Render dixit's flow page
