+++ src/app/dixit/Dixit.scss
@@ -2,140 +2,4 @@
   width: auto;
   min-height: 100vh;
   background-color: #BFCBA8;
-
-  .dixit-cards {
-    position: absolute;
-    width: 750px;
-    height: 200px;
-    left: 230px;
-    top: 505px;
-    background: #F0E3CA;
-    border-radius: 30px;
-
-    .dixit-card {
-      width: 100px;
-      height: 150px;
-      margin: 25px 6.5px 0;
-      border-radius: 30px;
-      border: 4px solid #56776C;
-    }
-  }
-
-  .dixit-guesses {
-    position: absolute;
-    width: auto;
-    height: 200px;
-    left: 65px;
-    top: 75px;
-    border-radius: 30px;
-
-    .dixit-guess {
-      width: 140px;
-      height: 210px;
-      margin: 0 15px;
-      border-radius: 30px;
-      border-width: 5px;
-      border-style: solid;
-    }
-
-    .dixit-guess-zone {
-      float: left;
-    }
-
-    .red {
-      @extend .dixit-guess;
-      border-color: #F6A9A9;
-    }
-
-    .orange {
-      @extend .dixit-guess;
-      border-color: #FFBF86;
-    }
-
-    .yellow {
-      @extend .dixit-guess;
-      border-color: #FFF47D;
-    }
-
-    .green {
-      @extend .dixit-guess;
-      border-color: #C2F784;
-    }
-
-    .blue {
-      @extend .dixit-guess;
-      border-color: #E4FBFF;
-    }
-
-    .violet {
-      @extend .dixit-guess;
-      border-color: #B8B5FF;
-    }
-  }
-
-  .dixit-score-board {
-    position: absolute;
-    width: 500px;
-    height: 500px;
-    left: 350px;
-    top: 100px;
-    border-radius: 30px;
-
-    font-weight: bold;
-    color: #56776C;
-
-    border-spacing: 0;
-    border-collapse: separate;
-
-    th {
-      height: 50px;
-      font-size: 22px;
-    }
-
-    td {
-      font-size: 26px;
-    }
-
-    tr:first-child th:first-child {
-      border-top-left-radius: 30px;
-    }
-
-    tr:first-child th:last-child {
-      border-top-right-radius: 30px;
-    }
-
-    tr:last-child td:first-child {
-      border-bottom-left-radius: 30px;
-    }
-
-    tr:last-child td:last-child {
-      border-bottom-right-radius: 30px;
-    }
-
-    .red {
-      background: #F6A9A9;
-    }
-
-    .orange {
-      background: #FFBF86;
-    }
-
-    .yellow {
-      background: #FFF47D;
-    }
-
-    .green {
-      background: #C2F784;
-    }
-
-    .blue {
-      background: #E4FBFF;
-    }
-
-    .violet {
-      background: #B8B5FF;
-    }
-  }
-
-
 }
