+++ src/app/dixit/actions/Guess.tsx
@@ -1,11 +1,11 @@
-import {PLAYER_GUESSING, RoundState} from "../domain/RoundState";
-import PlayCardDescription from "../carddescriptions/PlayCardDescription";
 import React, {useState} from "react";
-import Card from "../domain/Card";
+import {PLAYER_GUESSING, RoundState} from "../model/domain/RoundState";
+import PlayCard from "../model/domain/PlayCard";
+import Card from "../model/domain/Card";
+import Player from "../model/domain/Player";
 import EventNotice from "../EventNotice";
-import PlayCard from "../domain/PlayCard";
-import Player from "../domain/Player";
-import PlayCards from "../cards/PlayCards";
+import PlayCards from "../cards/handcards/PlayCards";
+import {GuessDescription} from "../cards/descriptions/CardDescription";
 
 const Guess = () => {
     const [isShowEvent, setIsShowEvent] = useState<boolean>(true);
@@ -18,14 +18,13 @@ const Guess = () => {
         new PlayCard(new Player('5', 'player5', 'blue', 29), new Card(5, "/9j/4AAQSkZJRgABAQEASABIAAD/2wBDAAYEBQYFBAYGBQYHBwYIChAKCgkJChQODwwQFxQYGBcUFhYaHSUfGhsjHBYWICwgIyYnKSopGR8tMC0oMCUoKSj/2wBDAQcHBwoIChMKChMoGhYaKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCj/wgARCADsAOwDASIAAhEBAxEB/8QAGwAAAgMBAQEAAAAAAAAAAAAAAwQBAgUABgf/xAAYAQEBAQEBAAAAAAAAAAAAAAAAAQIDBP/aAAwDAQACEAMQAAAB9iylozypJ6yKZfEprFyjJKQwbh7js0ewSLfM1PIzfr5WIzetKJIuGkisMXWaDVn1m4L1LLRtLQKrOgXKCcV5HLQktr1spLVstBnSzvR8xq+az03d/wAntWV0BkuB0vXWaCMGRZdpajNKtFp6FHoZzw0Ey93j17nA5BkltetlLallvNeC5z8rMdBNJrER0VVZlaQITBojajZelxrV9FscXsG6yYmHFkJvN467rXg/dZ6uWGTfCe6a63XWKEqRGY8SO9JOVbUUYihq7ibiXCcK84o0PovI3eTHc4Meb9L4HHfN0MVrPp9V6f4z9U6ef0M9M537q22pFC+bpCg5Ac2mK9HOgTDsltVwuuwspDpnNVFnNu0ZSs46vz73LOO/xvU0yZ6x7+k9PO7CfD1V4D1BUcouEVJ5qZ09ZStHOwqjsZdRbDKsqGO8ivK/k0HcleztUl3LZz1uvWiGIpW89GUJV6FeGgr1Eq52dd+rzxJTW6jh1r2BMXRnNw+XZNGlYCgKISpYQRpEo6wnab0kmlisHBefXH0hR1oW6kKjg+r8y7LxNdWW1NiTRIG2JW4Sb5NkpAZZxNVg1hbNJENRPJSa9q547cjeXhwwxuBmQd0XHVvxYt3J2xBtDlDETICxlt54gOudLQQfM2jmeApYayQRDJXdTdBWitrTWXdNp3zhstd7z285mKJ/NKXLSzv0lMl7ectbRSSVDrakWpI0cGig12lJE6XHdSQRCufsUlx527Vjg9JhqgUZGi+hwd6c6qPoSmFU2OlW6h3y0xJSDXKHTrUsOHX0JmqbiQrS9brijYLr36OtE2M4O5gTQXy6FI6UWmJCfjNvPc+ziTiXTlFh1tsEoSbjsOP5z0xdB7PVe3EahgRxc4CoUVoMqauNuO42/MUk9swAnFNA5+lmzo8qdXWKxXiRGmFYZDaw7nvXLaGkTNyq7ERj22JXDnbkwq+hueXY9FZfAafqiKgZ+ZnPX1AGIqyo6FpuWvPCtu2jDLszGQLchfMttd0w+wcuOq1mLKGx5tDYslJvCRNotievFZmZKdewnZjgI2gQKL8g6lgBRiDOppIn/8QAKhAAAgIBAwMCBwEBAQAAAAAAAQIAAxEEEBITITEgMgUUIjM0QUIjMET/2gAIAQEAAQUCqONMyK0IdY5DRthsPXrxVw3MMOxhh9yHuqzMqifh/sy1QY4wD44nIDTDQBoA0w0HKLnb4t1H1NS9NPqn1z64ec+uYaENGH1/3X2g2qifif3Glsbx+xsINrHCKNRTxR1efG6GJVNRXrD2C3XNf6P233D70g2qifiV/fK51VoxLY3j97DcqrShVpp7pR8Sv0yHT2V6nUaZOjUqnluYfJ9596wbVRPxa/v/APr1Efy3tPrG2u0q6pP1+vQYfd/dnmuDavwn4tf5B/M1UMb2n1j0fr0GHz/dkq8Darwn4tf3z+Zqo0b2mU3K4bU1rKrFcbjcz9bmEfR/X92SuDarwv4ydrjYOte4cGN7TNZV0W0V9dN+ns018Q5Gw3M/V/OixHV0nmXe/wDr+7JX4G1Xhfx6/vH8rU+TH9rT4i55v76LSpo+OE3aW8ahINzuR8vYKy0Wvpmfv+rJX4G1Xhfs1/db8nV+Y/taau/qXk/UpKVrfm34VTbSIPUwyNLqEdddqPlqp+/6slfgbV+1ftV/ds/K1nn9v7Wmp0/KyysifCqOvqj8H1K30Ka6YNzuZzGnpp1Ju0w8fv8AqyVeBtX7RYvCpgz2MPmNY4JJ75Jit3tt0Jr1emTn8Npxf1J1J1YLZ1JznOdSdbAF4cat+dmiUqVf6S/fn3ZxmoiCGV+3+tPj5a6xekWO2nTm1j90YGW8WCKqnquIL2nXadczrPOo86zxrXmr1d9baXXXvdqLYursl2taqofFbJTqzahtJi3sD8y8NjucnB8lgIzd87aVgEs7PTM/Urdn90zBtmEzM1gPUB4ltZzQ6jCsxYiaNSN891jbMAIT3OyT9K2HsrwRMNMNO+2ZmZ2uPKH0UDJrjdjtwlnkglWrjMBM7KDG7EGD/ZO+WVsYaYaCHfLYZMPavF9sTTITXjiGOdhhoqssYkwW4RwzRsLM7VNie+VNFu4Rf9TgidBnD8kOfQTkUuEgqWwtRplsNdDn/FIzu0AJn6gHKVs6QXVkBqwtthMO4jalotjIatVKbZVqOJbU1EcRZGqI38wKsFAwaSldis0SgtCoDYGMYrOwODVZOkGgPYrCdxLPedgcQWtDexleqdZ86cddiw2rH1J5N+JWEYWIEN6/X5lsOw7wArMrGZVD8m9N4/0PozMwGAxfAnNVU8YX7pZiLbmXDmQhD2+7et+MChhjvZ6B5dOYalxOhYZ8rZF0VpllD1bCL5Efw57KY2cqLIuYrYj2DNnu2ESdPaz0CEHkrd8wForsZdner7ghGYwg28B2xAxx5Z/Owle9noXzZ2i9t1mq96jJ+XaV0lXG1i53PZW2VsR/dsJXM7WeiofU5z6K2Gbx/tplAX0GET9We1sb2e7euGHxbtiYiDAI78Jwl3+ddLHr6sY1GmQNV0p0hOnOnLF4q1f+XhWXIZewXI4QjMKzG1caN7X9Hiv+szlGPJWQ1tr1+vRWhCGHouljfS3tdsRmgOBmfqYhGInYtMfSRzPCcZiNMTExMQywdTRTSKxKrMQx5ZLuXDE4zhOnBTOlOnHq7IMHjHXIWrA6c6UFU6IE6U6U6M6M6Any1c+WrgpxBXOnDXDVNdmq2quy5+hOhOjOlOnOE4ThLKsxGwKht2gxMCYEwJicZxnGcZxmJiYmJxj0ox6azgIQJ22x6HWKiMFUTiJwE4icROInETGw/wCjecTExMTExOIjqOX/xAAgEQACAgICAwEBAAAAAAAAAAAAAQIRECADEiEwMUEi/9oACAEDAQE/AUPaP0e6Gh6Iiv6JDeyJbX6ESeOOKk6ZLjVeN3oiWOFecTWqS2lji8Dfgbb1/B6sX3DxZZYsUPZDHp+1nkWyWPo4NZTOw5sbvZYrEiEbHGvTZ2E8SIM/C9VrEchsW62azHVlbMooXhYoooZ1KOp1Op1Kr0Uf/8QAIBEAAwACAgIDAQAAAAAAAAAAAAERAjAQIBIhIkBBUP/aAAgBAgEBPwHtRspdC6t/EQvsZOL0J5X3uZK92SFf4ePDMNKfRo8ELTOMsoLLcz9ITa9t4fRcUpSlKXpSlKXmEIeI1NP/xAAzEAABAwEFBwQBBAEFAAAAAAABAAIRIQMQEiAxIjBBUWFxgRMycpEjBBRAQlJigpKh0f/aAAgBAQAGPwJp6KuvNf5D/tdd4x1oAS1ww1iu72RJUuqbn/JD431Qv1C1C1C1C1C1Crc2xa44YDo5LBic6OLlSF/VcF/VcFwXBeEMjvkvCdKobhd5zElwb3Kn1bMf7lsOa7sZTLZvRis7L9yDZ4Jj0+VLjZ/t8I92Jz//ADMUMjvkvC+k6UULvOaHAOHUL0HANbUClCEyxsfdhwyG07o/p8Tw2hIIOqfbOnBaNwtdWhQHpuLuJxTP2i93uNOwzOQyO+RR7L6yjcsDo2Xh2i6bhyGQ/Ip3Yr6/juQyH5FO8r6vF56LWXckII3s3OQyHuU7yp6LEQfpSEL3lrtl5mEW21i2ZjEa1R9FoZahaR03BtWS+zPvZy6hBzDLTdAWEaNpc5DIfkU7ymoXC/siesqWzM6r0v1FlI0DxqiYggwRuMTR+J3uHLqpaRh5ovJFL3IZD8in+U3shcLgrToVqnOqeKBwFx+StBa2YswYjancEHim+iZZonWjhLRe5DJ5Kf5TOybcLgrQ8ZuFnaOIB5L8fovs+DnpjHPxuAgu57m0/bhotnHFVRbEG0NaXuQyeSnCRxTY5IGdFS8YnUWF+D6qibG1aW9U12Nog8Lvbk0Wl8hpT8BLEcb5WlxQys8ogkCqAadpam7oLtB9L2ieNFIAlar+q0C0C1XuK9y9xQDLQhNZaOaQeiOwwoD07Oia7CzGeCrZM+ysWABaBaC7achVWaqVQXv6oKOLVafd1NwTwUjVfkb+TmEMI2lLjJudlHZMupkCBTj/AFcER03MDU5hkoh2TcKl5VK5gJhwMqOIQMwtVqq5KmiDgnDlkoF1yUXtWFokrayGqI0MUKEqGfaBcK81BU2b5UOJB6hazkglqw2plnPkvyNFoDop9PxKrZDwtizZK6ZKLmFXVSv8RkojEKhW3TqtjChicOyLhR6kkyqVyaqCpsnu7SpKknZUaZtv7Uiq2RXmpKpkdfRaoAUhc1DR/wAlJ05XyFR0qDquqBajcL6qi4hczyUnKeu5F1FPFdL9lAI5K1CkFDtmjiFotFw+1q1bUZ9bqCl/+oo5aJvbNRUN+qGIyb25hCrW4FHM3tmOQG6AtQgTF83g9MhzN3QCf3WI6ncNG6blJQF5cmkp6ab9b/CKCN2qBzBSdxB0QPBMdzCLXaHKAjcO276ZI5ZdF1Zd0y0WgrRDcYTdAOXTJovatF7B9KgygN5ITO56ra1U/wAKSBO7k/z/AP/EACcQAQACAgEEAgIDAQEBAAAAAAEAESExQRBRYXGhsZHwIIHB8dHh/9oACAEBAAE/Iad2GO2KGhhP7mvfj/8AD8SwCwbHDDghcEOkhCEOl6x1+R39fzD1wZ81ua3+D7hiP2+x6l59yfES7QbOkHAYJC98w4eyNkErogkf/GT/AOEn/Ggm9hGBM2QixhdL9waFcU1qTPseWP6MT/vB3wndPlnu4oRfCVUSCgo6tfdNctPgH+zjYecy9ZOjp9kTCBD+BgJWEDP9zKQULYxfeFKJeyiyXkBp2587nNkuBpFLv5hsoZjIgQXF1g9IxjGP1g/CfxZ9qb5/R/cARsQg1F7n2wfMjpB1EIB4csIkIrRRl/OdMsLagUBxbiufiAvZR0LnJyc8zZB+ABsFNu8eYn6gqVcoCAYGfBcY9X0z4x/G2n6Mz5SLD19mO/om8Vj3PtJt7EOhDq6X3LmRpZIcDk/sjWIBpXHQsuPV9E/y6DQ6n+pzP0HmavX2Z9B9dO57n2k4+yEIQhNzl7lwjroej1L4z/CbfwD953nwP9TX6+zP8PqbzaPuIcnsiVYLRPTLwMDEL154uEOnL30JpF6GPQhe7aIvwn+E295r1fru8+L/AKio2w+4Ckva0zEm8feTj7JbNY7wuwl5As0dqZ3hUauZNT0rYw6cvfQJpKhCXsiIEmoxlKNjgjOwfz5j9J9k2jV1frO8+L/ufAPuYfpxN0GY+wnD2QgusLqCjZPZLkAYDcD8gGl2s7Qj2vtH/iQnP3K6LEvHS37T+XAhooixRXIGyu8eV25Y7+p90+3oGEX7nM1v3lnw33PrzBR3H2k+wl2sBQ8y3GnmOigGg7O0HVYVbL4j7SQq1quP6hOfuXFjDXQjFgqB1AcKpMVDoo6GrdQbO05+ofJPs63U/Y8s/R8sX4P3PoRT/pNv9QZ9iCA0E/MSpLhBHPldYlXV3ABD1uLCKRtd4anP3FlxQcQ6bBqq4t3LAAyGPE0Tn6IfJPs6tn6Hlm7px87l47oH94mLGl51MoS80t+5Sq38y1qr7C4wxnDdR5uoxGtawrqXXlcB2v1DyfmOWpS0ky4loXB4cxCNa4qOQK4SmNcJmbut6oIKKX1AvhlRc7itkcVzNsRBuM5m37nj2QjiJaXxL7aBdYj6+6Ljb+ZcLdduYud1zAe4f9TP47UKhmihBHP2Sxpf10t45nafjpD3ETLAbreYti6tJ9RyizujCU1+xLga4u677ljJ9QebOqu5wx6mPQ9Fy8oyvOuu82vf1Cj8ZBcr29F4hOoNKuU5wSwIz8hHV9Ja67ZYhTpD+QXCcsKOCIGxOIOURrQ+4GDg2xEhNrBmWJwIRxLgQXMskdf1S/36iyoQ+5FmXiJs7RfeT8VsoTyf8mNGrTFjXqej8SqZgxlljhYgqyVMVi9Llo8EBhdXi4bC7rmDmCXTA4skxg5PEwfIYlUsty5hZYGhl5ULa6J2fcEJ3mxMqLcR7+eXLG2YujrG4lTtYvZeJ5VRJUIeNgu4Vi75SwveXTceLT5m234Ynb8opeURe0sHkexG0WFZbtSS6M3qrvLiz+4ODfdSiPXgxDlTuag6cdnctQjiF3cL6XWRqM1E+I6eAssVjdjN/kjAYGLpDzmYM65VX+JwC9LgNYIEW9ZixeYEMaehmIU9JTLzqswr4OWK3EcS4sPKFQA6uZ5f3BWgfBEQWvIkW1azR/8AJi8zCHMXKD2rc0aniOGmx8zExhFNAuI7iC/arBLpbreJVp/1EgI5CXEZeYpcTkHvBxXHbOEB5Jge0o4K47szV1bQ0XnoY7tUxLlvuE6lxzMdgSqwvm1z4gNRWXCNa2+XtBlWzNagYldkCuuyWxbycRinhnA5lKA8dHMR/pGr/wCoc1/YgZCzUNeGPUlTe3Q9LhBBzJPjSlltE26ZHdxtrYY2NkAXDBs94pOcxYO0XoZjYjwyvJJRsOi9dELPrVc2TnA+cE7YVx+R/wDJ/mDAp30ZA79DoMUGeZS90WEz/cJo8LiFXuc7BMyHAW6Ivzx688Oo32o9zmO8Y9NiMp2SnZ70Yl+PqcJGIJh6jGTVefiEIb9nRQpm4Zxh/kzmrSJsV7wQXCFFlI01F+eL1PM4j902j1OM/IgBLIMpYmxv3K67dxyC1hBeCh26KlKmyVDU/FhKpFISf6nzY9W00I/Vm8etwlxYVMTHmewbiVPfPWA9QhDpSncqp7zlEDu1zBiNSybvMehHmcJ9iLPQTzDznqqbZuA7wPeAAy8RzFq1PeqfiAx8QHdnkZWKd2GrcAvyWi2Ju9UQxh0QBbiIp2cRTmWnM3nD1FSeWBVgFvQnyMu/WDg4TZ1NhA4ZgOvrigVmHzGcwTqre/BSO0CKYkhESJRLQcYXHOXPEdTxOHqJVehfuPjwI1jN5ZTiJeXlu0V2mucJkzb8So0RZaECoDklnFL8ynOnYMWmV0BcvBwbGOWEYS/SfCsvpYlTc6lGVG3E9IHaGsZ6z1nrAdkt2PxAhpl7BwzsTGKAPHVJHjbk393GjAcs4AQ8JXtPSHhPWU7RFal7HpBpDSDl/wBQCEWMJZISrxK9NZSUlZSUlJSAVhykKMAdAUqKJSIRqPqc00peiulMOx0DsdABqVE56USv4hiJAiQXSVlJSUiJSWS0q5//2gAMAwEAAgADAAAAEO7STpVIKH6x1os/Hn3hYdWuoivD5wcLhFfdTReWcxK5QUJmJ2s4yeUbb2pKWKpBY/DW7VDq3GUaPSU03/SEaCbUvZBIFUYnu5tyjZLI23yCrmuFIgmoBuYP1+WTKM2Nhk1HTYPQMNKBvsOkkLzLvPpApjpkcK/xdP0lKrmtkKrOoJimj6HYT+72vXi+X4QIDd/wAr2pgx8kfokgv3sU9p/Dcbg8+3g4aCv/xAAeEQEBAQADAQEBAQEAAAAAAAABABEQITFBIHFhgf/aAAgBAwEBPxDsdwmzghjohk9cZMX274Tm8jhmZl6IGf7bM/j7ez/l5f7PByYdON4b5x9uqQI/2fsgSHqpjjwEknW8HHL5F9vReX+3yBa24RewcN5es9n3ufIvsHZeX+ySE7FrG9DnLI3ox7Z8g2fYcksR7Qytz8AdQvlmOOQ5bxlmcnR42O2XP8OFfl2DJx3ymX9kOUEmdwD5C8k9Sca8rrhF2R+3ufJPFv4Y95EWpXh3Rkr4nTbHC93r8+ph6kfZI6Q6X3g4feTOwOx71YNlndvGx27tS3eBi7jbuxgbSZblWe2xtc47unHdqE9W5cyyyT8OzW//xAAeEQEBAQADAQEBAQEAAAAAAAABABEQITEgQTBRcf/aAAgBAgEBPxB9thltt4YGsF8jUtsMs+TJL4Tvdux1vUu4B5yTPkz5H5/I9n2/Jl6j8k0G3QY/gez7PnD5f5PDqPRZh871wez7fnISpAOJb8Yy9QdR7PV+cE8FtttvGN/23/L2P5dzeGB3PT6yS8hecB6kXdhnWAefRwkjjLgdwPtsMefetryf2w9cOnwcbHymwxjzg4Y+O+FbUK3sjbu2XDjjmx8FVhyTXfziu1CI1JtkPgatSev4HV//xAAmEAEAAgIBAwMFAQEAAAAAAAABABEhMUFRYXGBkbEQocHR8OHx/9oACAEBAAE/EEIiiDrgl9NoW+EPjUJaulR7X7Etjpkh6OfxNRTkzUNv8mSFrUGNZ+m9bKhFzT5gxxNIHiMJUd+wELFGV3ioWq2I6ScRmv0Wv6L1iXt1KSpqwzVypVrdtHk4+e0qVbdin6jr3bfEW246RWvr8RF/B0lVxsTzdfDLA3KMYasyeHiOcjkrXPWMKKdKEFg4u7n+JYbXtssfnTC+ZD/YTT+yywRlBjX0HVGaEShUaa4cmbMQBb8PqDIBWOnMcynwsH4gd07U5/qM5979T/UP6nO91/Uf9R/UR+1/UZNIJ6u4+pfkggQaAoIrnE/q9o93B/EFWsH+8Ntu2YPeFKkrkbIaPJ+Z6qTL/wAwyj6NppLcSyCIKWmjk9oezCwSjldGesGs0IwzqxYybvAVbHY6OKiqGaTKFFqzZknNxLIDXfxEgCSJawCrlhSneG45vM4MT2v2lyzPv/kgpz9bolfoKPPx8oAgPDtX7mtrS28QifxuVH+syvovh+g+psTMiLTN6Ik47KgEI6oUcll5EYKXRxhF6qAF9UMrZRAVEEEmRQsYXeE/Rqj8TAIAVFC7WXwQuikt2hHGE95UxANQFTqKqrrVat+k7mBEfx1ivwH2mKvX9RjX0kfHA6Pp87PWoGZf1UdL1+Z4ivmH+HWDJ/VfSMcrUKshsjUVM9TIyvrKnh9g3844sK04uV0ooEAHAGqg5+fzKo2+kmA+Je3b845t/lQU51ivFmprCrLhYf5cd9r5ItENmdWQqVO/zM/O+WLL+Mkwe36EMM1itmBVx0xAxT5iZt5llZTeIW4tRYZciup94/07Tber+JkSawi6P+uav93DuIX11Vo2D3+ZhV1fln8nclgdT+5Hcw45yH4ldgJB28D0j5BcVj0qAgphV3BmYMvA6yj1S6Cyr6xWfSoKYLn1Ay/BOk/rN+X4QyAqzWGEj9b58OD+7hwhQoHNKVjgFfQrpKFgeETnvPun5Zgn+LJv/fMSxCUZa0dG1Ol10jxiHsI2ufHtL2dUODOOjFC/UGgNI+02huC2YZgQEIcusySLJtD2MbNfgn42P3E4TkjlSLcDuwSeyDuMr3ftDl4/LB9l8RPr/r6zV6nS/wDXNH+LnjNLMmxkQ+UbXk/6y7zPywU/4yTD+u4VHEF7S+r18QkAKBeC8+e8OXQYY6dornFbE9BT3TOKdroBEayKL7wQ58/0GIH2RGvWc3Au9gRXF0HT7a1pSmax7buNCSIfU30IlsWq3VW59l+WVfhfEOYapgmG+kEwbX74b7ARdN3fsmObenoytDq/MfvPyzZ/GSDI6n9yMFZm2Br4CZ0M6E1DAUgcRns8x1JFEXQNNOejBw2wMSvQo7M3xBzF92FCXR49Y8PP5l4hJFEOkZdujcs0hyVXMambREiAtwZTLqUKj2dO0U9J8s+yfH0WmBSbFzB93lifRho7nwogHl/aZJ1X5l3k5+TDTd/kTA/3cFBjbPLcbAy4jSuxYUJ2E2dIoJrXItNXDoKdIcqY6Ebb/ccr7v0EdZg3zOJ1/MWIlGt8RTDlEWl26KehjEOp8hGWh4Md59lH+jvF/HpFcnhFgiAi74TyXaN5C0r3ItZAujJVb4YgaUAMMrtYKSau27uMyk0YRwtcbwxEOSC2BXTEduir4oAb5wxafLkQdkM+1xuGxVaOKrTpWIUojhxRo9UEUoV9FhUC1XTGrH2sesmJeEqMSaGi6JWOd0XwzpcJqrLoEb2Dw1OLDdD9AOOY8KK2tq8MthqDZ3jSiu3EDCiN5gAvdHeQXtAKYPdqMOwPffEBEHygihIAMXqBilQBlAhkvmV4rt5xtkerDs0YcL/sxKtstDRXARktpN0xehjrhO3eG0CwgC+pjt9oS4/unqRWamqJ/EolR1X+4AuUOzEuB3piODx+0Ub9phDZ5CAIH4agMbExZvmx6RTJGsGms05jNps2S+buE26sQ96kuAikEtBtGR0I9U3YfmX0i0WMV+4mOlyf3CPmZZqyeCXB2Ja6Qtroq9yxCLkH2SywQqguLhY88FfrHkZSHLZwXfzKGJsd+sWSA0Lv4+GOUBwA8Z/2JygBOW/9lCluktcKDcVveChGhudyWrMuUQFJxRqHfdgFGRluJUwF65YevjHYiqwTsHlDlXlmbPy7WahmFhwQ94sGUM8y2VMq1RK4bmJkum6BqB+DdsWgIvZolGHVg+UIyNimX2Vaidtn5jkOKPBKrBgcLN/zrMYrVXfm/tBgs6IrgXq280SuvEqM3HveIinQYrIHuUnlmCvAsCtyzrAJuLk6y/VmGiwSNQARdVBB958qJFV1pjLIvIzFXoE0GO7q+KiiuhtxH4jP1VURbvf4gEoaWWKb9QmKXqHeeD4lJ+l6BVdneC4NwZutSv0YN/vUsY9xiH7GA3J1VL4twNr9lwsEVLWmU875avqekJRypuGaZ5wOUjukNqml+aMyrtc7xXHWU78MBmBvYQqkQrGE6ukWj0hFTwcEcprykTcDscTyHiXo6gbBnm7xEl2XExLB7IJe8kMIFlK0RfNJQ8vg4nDygZfS732jqp4a/Q9/EHO9Yv0j3HRA/ELLY1dVG0dRRAI5SwiUJUn/ABAwpobHtDeWbBXegdBuMtdmEOVb0pxiOHBc+rvQPtHUPylfe2CR0AIEPYAKVlIO0CWS+xBrdvTkign92OkvHOhv98S1FRHMfH5i4Afg4olAo+8thLjAQVGnJDwi9lnFkVNJ2KxgW+8M+rk+8UHc21euYu3worA6ae7AtEWsXhdYS+syFhcC9adekdwdVfiFcQbBTLHMrIcO7IHbTQA2xAFgtBlWgI4AcY6ZmMYnQC9XrGRtLoaR6Q3NHCSloAOFzAQqaldeZRiNmAWwaG5ihVgO/clM4yyEUIjkrz26SsvygeDR1dsvBbLjdNztw4am0CIDkaZZFkp2s6SkkhY7t1b47Q1zm6UZlSTgBrpgGZLOpEB2IJBscjFAkDQQUc1fMXgIItXNnCQtZXRwSW+dDZx+5gLUBcDxFI5srqRRpdsHeMCEL44v/n0LhGAdMXwsdjkEIZc57PmPcBdFV56RytDRoPEqYIuvo3FbjsDQAvOIM9YYsEalWI4UNQhn2IjZfENG8H4l2eHL0gZU445ldi7mnVQ1chpyniGuSlO4WsX0YZGCsr+8JS32CVE4Bj6S5IcCJ5zvkdyDHXrs7S12KU9A/c2ZvHj6OvJLU2zwEeL6x3W4bKfWWMJ5J+Zl6836DFQHu0Ret4Yu6YDJ7TsqDkrB95hjiMVoaZTDqmPWKjgivJLgaubwiWWm1fHpGqU6XiLst3RmACn5I66vX0lq3yjti5PoJQLXEVWKcYT5X4jtBN/qDi6kQU3TZiXglSth7sQKoehn2uBT7S6EVwCDJPkdoW/duG4X0lM6maQFyDMQ6RqPKekYrg29GXcclQ5WMg74ue1lgkunIln3lQtIjGTD9p78+fpXHqYtwcP3ObsfEwU3jKlh95Sgw3HwEChQ7Zk3AKzVSzRoDolNHZD21gl3xUBNMT1f1GHluk3qCbEwJ1O5CjiF03CI4T1Cv4jyKUYLA6hEplp7zEu6LMvJHNEwycTTwI8oqYnT2ZWJZnRmV4G6Pedhh13CXUJQt3jMIoqB6XiO7NHZ/qGyyCGOC2JkNqgIEq+PDCxDYYmsh+aJPK4+d3EES7uLm0H7fQbGOgmDFW3EwTx+YA+sd2zM0Dxcs4e0K2twPMxfZmO5y/lS8waA6teY+7YOicdJVgwh9RGYFFzomJzfsQgF21eJjz7RBur0FSyXiO9XKALQ/MJLodXzLwuTUVLFNVMqsvE1waGrmqD6ShyiHQi4JqOzFnP+LCayAUeswgHSPcmCAhLS0ev2yGZeRBwG2BdRp6+YOcCj556MsCx89/8ACQH7bGjo9T4lPQj9MK4YsN1AVueIww2jEzDVFxD0B+7MHKSoEwxB2Nh2KMC3mJwUiZzKXJvzM3P+LHR04+6IZSExQRTUS4hBaHuw6UQ4jIU2iRCKwjpggFnad8H7J7QZxKGTirxcAFxQ/wAiBotiMgpzU0m96qr0jrXbIDb2YcNrnUhFtxBqcRUcVQc1UrNMDsV5l1lEratygBgVgE4jtEpXhW/ZgE5G15es7hKJp7TlHMtqi7xNsH2hfj7Q7PtOU3pNVpFRyabOyl5I3+7pO1AhRBMVFmiP0RrqOcaui8j+or7e+E/EKAyHNQCAKgJBEUgIqtHFQzDCEuANu/SCLkpoi0O9H5hXSEL4hqJXpOxmMYg+kHrCGlq5TiHJUQM1PFKukO3PHPHFjRDCpRYQgACdioCYCJcR6RL9BHokDxEOGVSjCXdASxXC01cpiCdiGEjBFg/SOxP+tDoPvEI2q33lgLpnrFtwppcSitE64JfYjAZOsWQukCnEKwYlLqNeowehK11FsRRm/eCwB2Z//9k=")),
         new PlayCard(new Player('6', 'player6', 'violet', 25), new Card(6, "/9j/4AAQSkZJRgABAQEAoACgAAD/2wBDAAYEBQYFBAYGBQYHBwYIChAKCgkJChQODwwQFxQYGBcUFhYaHSUfGhsjHBYWICwgIyYnKSopGR8tMC0oMCUoKSj/2wBDAQcHBwoIChMKChMoGhYaKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCj/wgARCAKsAdoDASIAAhEBAxEB/8QAHAAAAQUBAQEAAAAAAAAAAAAAAQACAwQFBgcI/8QAGgEAAwEBAQEAAAAAAAAAAAAAAAECAwQFBv/aAAwDAQACEAMQAAAB3wlfnJOTTUUCBCaQUsoJhQTRTQ09NLCghFBJlAUnIJooKaKCqSmq5cmqG5NU05NDT01RTk0McmIT0EMoBpya4EgmiQhkFUgimJzZWoklhokgBQYDw1VLkwVMijTJA0XL0w0nJqacmJN6YkPTVpLixA9MMtwarlxYmOTRI4sSbk1JuTRDcmoCWoHpqocWJD0xDemKW9MIOLHMcWmaMkcjI0lhYBagBDSEgblIG5SSBJK5SSAggaSQJJCSSYQk0Uk0kk0gkmkChApgRUsJIAUppJARIUtJIaSSaSQEJMJaU5C0zRkjkmmBCAMcxykDcpJAiFpKSTRQNJIICkWBJISSYki0EkhIFhSFJJJCSSAUhoEISSBJKWgkmUkBCSaSAFIJlAgkCm9zSqc5jstAC0lrXNqAU5pqeE2ojSUgnJQNoixvTrzR7CxntxUfdQp8Wt2ntz5ySvIJEEkmJJNJJAkgBSQIFAElLQIlgpIISGkkmUEBCSaSQJJS5C12Wicx6pMIcIOQ2uA4PQmZYHPvVDh7PitThSE0Y0jrZOOWXV2B40p9oeNQ+yHGhroedS25wQbyQRYCgJIhMgpoJJiIKAiBhJQ0kgBSTSQAoKWQhLKQAlrk3OactCQpopzotheODvhSPH3TSRzChTmsIJ2wjEz+rmpDQp+j50SS6uYoIEkk0kmkkgRBYkgBCSaSQJJNJIgAQhJJNJIEgkJEJpSz829JaemXzK6uhSwyjeDy05apzUE6S8b2VHJFhux8ZLlmriXM0CaeWkHFh0zlMU/Vy54ePf8AAakqlJIaSTSSQJJAkkhJBMpqBwSpIhJlIAgkgoIEQJZnhk830pbVGx5PrXpai2xEea33PEmYFplIWvz0CQhzxSQeB9DMxrY0BYlcxaXLiwpvLHJuCn0zjta1/u5eYqdiPR8/jI+x4a8p1Tub8iSVwUgwoICEopJ9nLUtmf4ft5as1ff8EpK4QSQgjLSuv5OvOfeZlq6nND4ntyywAq81rNMakcsf0XzgSTl8kcisJJOWGaH5z6RMeMt2RvbhtOo04kdC8DLXfLu9JmavteZKQe/nLHgGeQ+ufOSex3/j3sHRwuQXRwEIJm5Dc4PQJiXj+zK1oi5ZI3VMWdpUPZ8ViS7/AD0Ch27tS/zdb6lmLxPajqz1uPpUE0WOoa6KLluUbe+MUE8H1XygSOkOkjkiwSSpYdweB72GN4zpzQ6UZ6c8ugLXPP3m51gncYnpllT3vI2H1LGjcoJVXCed93ocnZx3Wa1jbhwVujq4MM7LgzJ9KTz/AEsY6o4O7LGkYqgtAUqMOs7t4ucWyPV8rHW0peJcvSYb55015HsZEW5HhpirZU1jjZTWJZ0n3GNX3D73h4J3FUYx1oU6Jvma1kHeb6iBDT3qTWYxMrUAmbLZG5ZuDJ3uc6ubT0Oftazey8+PLo1svSmw6LN2leiU9o0zemBjmFSNJSYSSCCGgihhIoDggTXAGlOQEQCY9JxuKBIBhMaRI0BOSMtloF8txC0klIRCAIFICcWRqRiDSuC1zBqjXOtpVt56c53nF9TGtXVzNKZAK0zDgUOAQOaghFEEkgSaAcGCG9NSZQQEtIEBgSJgQ90QHKIygoOBgIlta4Z2Y3iSQOXbhAJowYXPQLENgZDiiJSMlsyrnla6b1nkmdZ6D0Hl2+sO3h5vtObWPRSvFJKpSRAEoTA9ktCGeWkCxApDSUAUUgFpagBwlggDkTE09MQIES2pwTTSEBETRBCLRS6sgWOkSSAuYaT1GgfEecm8Hnast+rZVmbDTMN5WR9fylV5+vO8r9Gvzr6SMS1yaJSpCCxRl1NTB6GKjcU5RKocQ7WWkoTQVLiimjxthQzooFpAuBgkaNrgkNDmy0CJZBQ5mPO8Mka9ygUCIiCUNSB5H6L5HXU3RyNffo1XRTcHRFJnjV6OdLXaf2XE77j0gxh+RMYJGnmNjU9UofFd5xHbzYMUlQkmg9Nc0AQmAUhqcJbQ94QKwWq5nLII7bEV0RjYBCYSUsgpkiK2lSJ1xGnBNBCQgVh8Pw+lmdPazX0KU6XdHH1+esGOYdmNaxnzzvLpYt/PT2RkMufkOhnSTnh2sY09jPjWjpbWdcwaScTDMTUgODGte2WwlQwCJAHNGnxOB4CYGvhlo135VIA1uQJSIosTiLlz2nSSk+lEpGyN4rt/NJ6uZtaN7Tr7rifQMXPi802Mt+vcih0cNCtdzzZmhkWi/bLMFni4i6OUl1azDpNG9Oh8f1vD9VGtyeKXTByBpJJjHRQnOlKGIcYQyY12haVJgtAUCFoVYk74ynzWi2kmtCTNsRVtQWEMKSHFHRF7W6S8QRw7XF9jSnXgY6lV93q1jluirh8urtg17Lz6sdwytE+1RsCyL2TSxNDz8bL8+zM2AyTWBKHVPG9pxXazoEm6ZCJsdyYnNpBwLE2RMjT0iOOSCKmUTkSmNtzMIWpyZ1mbHSnPYLIBYiQZoJpd4JsN7olSsGq7SZlEE3kJHG+ce6+R111ut86sdAaYOmelVllnU1d/Ex3l1eXkZ6j1HnvoPLzSPa5c7ntdpIVK3U8f2nG9dns9J2mIr3VrNBmjHSzleGVUhovtZrdRMylpRJUm2JE6auFOmL6DPNynLYLpZRWgyXUsxWJqUPGTEsUukuIO0tRCA0qaGJuV4vwuDqMDq661u2+prX6+tDgxtfnVcMUx3w0fbfD/AHPlUUduvzYvkcrUcsTWcx1/O9CqUUqqHGN1JzmFoEISo32N8XuQ8vT9DXnurWZjqUcOnsNXzK/0c/fs4vs+fWrYvlFJXRU1TYbLjlYRMidXx0sz0H2pps6lpO7PlW6VxtWCldz5OQx14OpVu9m+9Eyhm9G3nZ41VadonUNqdHeveR+iYz09itY5uZRylpklbGH0a5nYpXlBG5mzJ+fHravIv0XZR8fn6ZX69EuujyExqewzPvPrNLgLHD3dhkOnmeX7zhOs9Di2jGOfeQUos70WU5kPcHBGxzcrEsUrUroztEgDKUjEJY889A8ojo5F8tPvnft47st3VjXuQmsuJNXKuw7m5Q0OXo6KzRwL87tjw1rTLpc2O9U5rdfnh2WS3pvNd3t1T55m+p4cXw2/n1p17m7g9BM8FH33NLTS1+Rk0ys4c9bWIrr8vTNvoXnvXc+3REFCBANAMNJFOFjs3DTSs8/PRsRUM7RXs2CnT1ZOap65dRxuppB5xT9H8822fsQdPL4kKXRVYZmtREFxu9jm9ZyaSch6TjZRy1qNRt1e55xraY9hUy4tMs3I6KB1l9rzejU7FbBq6Y62fnuZfgo4brfe3qsny9Hv54fmvaS7k6OErKmGeJJympI1K2IS3hjs6L43hmKjJhtJUxc2q01j7fVlDZ0pYMip0tTHbE6blIQ7jjehf3cGBmdPTT85m7LM005at6BrI8/7beV4MuybQM4jt/O83NBa0SedW9kJ1opuyVcNo9u/G+FzfT44vynptfaKhmMWZZyNd2k52mhSrWoYJekYqus2qgt5uYtW8JJAA5kuKF0PJq90Yh4UVaHXShl2dDrwzRsR4dFJ1iFORtbP1xs1rW1eXO+jcV12uIjmkrHOra2jlv51o387Dq6ZRXe3z1v89sU7vOa7ebfPs1pcbe2WcecdKtnUrs2LOtV+Xo1EjTmD5CPMi6p9fOFb5bfX1ZqXOydEbWSdRNVnWFpFeRsEVZCQCCnl8+upUhv46MdK2DmaGpldRV0ch+8X48mONdWnSbpL3xnbC7t5N/E1NriOuSklGv1ctSxI6dOLivs4+lu5h63TzR6FO10Yvssky1wnwS56Tz1X3N5VhLZg2djl6IHzP1yhwep4UOKsvu9mPXdPjWuTTXGY+jTNKWi2KpatGs9OatYbDyb0fI8u3Q6FK6iEJuVAEQ+ezdDm99azVX9LjniGpF19YWMNn1jhihguRdXOuux+umLG/j6ZMgk59vhKWVZo1e14TrJnclhnrO62Wsr8+6HjLjOsk5AB1XOVoFW70fA91Br3xJabk6VhrKnu1qWPoy20RUdMxfNx6fEc+ncnk5KXVP5+drVgxq47+vykOddmznoJe+cI4vdbkPRn5mrmPppQzWNTAb1MHTzc/br5uuWyzLsFXdvmpM31unxe1C6e3UN57PLdDhj8/wCqw7M30MGbZxN27hXuvDoqVhk3lRZb+LePl9mv0PB0L3UOafYY20s3zV5enORwioztblpXn0hruVzJFUsPcig4Or2MXBvyGfu4PRNBl3f1yxH9tVznkI+3zB8xbz3Xv1GpNa4LxYZ4ubtdYitXLm25unDgcn1QaHlfWdSyTgcPZdtOaehtpw79i7vxqDVrj86o9nzfJ1Ym3znT759RV0aW3Pq2syeX5pU6ihltk679xzjxaV8i/ovpXNh9Rmsb1N8FVmZ2vQM9S9zHRKr7oZS0DHL5jmOm4LKnuqX7V7Z5uLGtq3yoDq+bgl0l/R4liL6uXlpud3GSx8HpRzNFFuCPm+vn6pcbs9vJp0WNuMfUVZlqtUqZ9HRu5i7ntvx892dHM9n573mE43S2pcDjczr/ADLpfU9j5t6dIZJX54w1tBUZ1mauKXIu5PTz6DGQ657Dq6dNzr9NzU6Hm+gC5YxdtW+u9jeJwnp1Xz+vjeg2IcL5G9saCrncHuqOkecXuwsOs+3cHO45g6TFAzzpuVK1XSeh857rz7qxpdhxeh6HL2mIRy9jWSI1iMhHFeqTTVPuuB7mp5bt+I7DHHp3YVrLCXzn0jzt9Mvpflfo2sy0Z9GebM0LFdKWGu1laJVdcLscUeuelNm3NZfWnruauvQSq9t4d9XNMnN0nNk4twHFPJv2GQy7FzJrqavNXMtNMZDMtNp9a04w2vauiIWmNx+b+mebduWVp0tLv576R5e5EyIYSwTZ4Xum9LyfTjwO34buc8enCT4RwfdcXl15Xpflnb6vek5ikcnZV+PrVO3Tz5HUNW5gbZkaMxMmvzimu1r5eq8HZ2hxwd9Y53V0XQvZJdUYRjeR27AzlL0ade40dDOnZWmtVEOTDjZTHhiNyjfTrsqy50/zf0nzbsypa2Rq+jz6Akbyd5c0iDHhpoLGr2li6ka5vdcj6AuO9RtVs8OVpY9Povs24mltyunUVRLnDKnVk9KF66UuXJnrfrwOilBPGPY7DjewXFLwvdcPeOh0vGdntHQ3cnQqqBlHndcjmP1goJpV7CTyc/pIOXbn7V5mGjHkiwKtyidTacjK0o8v2WV1Zc3t4XTd/HI2Fc3bMq7mnh1mFTd0vJvPT0+a6F653octjDidA/LxvznnOi5r1Fd7HhZrW5RsrDsrSSuFA+YyRSSSSRJ7UMjlZU63U850k8NjjOw5PbDN7/zn0PXLR18LYqjHMebeEwuw0lUKCeBZwtGKJuGhAWNuIQ+Nr2K9+gSXQ43k5ukdG/0YcnL3FjXHjbG9mQ57XG35rdxnOh+b944dvN2VXHn83oNfSuqvN+c6/lPc5hMyTow1Hsk4vSSc5MPc6UE5iQD2VLSk53Oh5ro54ZOU6nl9+bD73gOv2z6DQzLUVqy1pDQVK0fn9Nx8UMVews6alpCCHPW/NzF3Ot1Z0qXNw6rTtqOtw5uW/jQ1HQ0ckaxchagcWOqaGL1UmmfHXew0nPG3Oxmg5BnbUNsuXvmhydXP850GF6uEjlJ0YXp45eH1GuKTL2GRxaSGNcqlNcGtjf53pX57Ob6fAvDlNbAPZh6xbxug497U1OxoP5jp+K5OmR2xcy0jnTtsmsMWdMa0c2kiBT5uJMy9CSCSK5ijlWkQSzPRXsXXyUL+5Y35su+c+p0oMKKa1aecLLNVsIpFEgw8Pcw+/OWxVs7ZaEkUvD6ZSU6EJA5zXKAHFwx6uVlp7OJu158WXu5uOvnL3ZfoY7np3m+xx9HeHGtzneTRz62RErUrWEZqz08qjnq2c7kMMgubhniz9BsdkOYH6Ntxn6Ng1i6raoudN1CnqMoOqY9VyCozfGyIzWZCDRSN54/OdjzHVMViqd41rLH8XokoRqU9pJNY1jYjmuKMjT03zGZvVBE9d510Ts7xJ+ovzfAdHq5xEliJ+Stpi0h6gmAtjxVVnP1LkXzy6J2kcq7p5KOecVx9sssZrOWSKzpEhZNtkOe0o41p53Q089eUy+xr9M+f3Ny91c+TpTO593uph53GpaczcjUws+rIu7S68X1LsWd033RNseonjo2s/SWalLpGPl08rz7Gy1FPQqau+UblQGuVfX599Cnr3k+afqHK+a1d6zU89qaa0z5Tn/Sc4rG1KU8O9Zp24nks7o4zr51/TMDF0LIQ2aMoVhomXUY4S4olFuQvFTRWpobKlzGoTat7K2yvN53o1UqTnnNKnXg6Ivz0aNG+nzEPa2HPFM7p6rk+iycuK7rKyTvz9LZzIAmzaUnB179nm5NY3m5NtzPYrm4mUapSKMBNVfHlRTKWWl59O6LKY9mPUwGJUA5zcskDiW07VcqvHNUrR0bjbiM8dRGyavcWVg6Gud3MuySxLoLEp2LZSZMZYIjIGnSwTEyOaahIvBliOEVjidnC6stm7Zk5toZJpJUUhTSTXhG+HEHsQ8voaGwalvK307swsCztOorWQ5RkNlbj1NbKQiEoBiek4o7CDMi1XO8R2wxvKN4hnt0EzmMP0Z/Vz890dk4kZlWbibOgiUqajLyDC9Ca4oSa7GodDqXKdDVcTFqKSaSQAcgbFPiOs/RfrlhOCyCKAIgEigaQ8P/EADIQAAICAgEDAwMEAQUBAAMBAAECAAMEERIFECETFCAiMTIjMDNBFQYkNEBQQhY1Q4D/2gAIAQEAAQUC/wD9Sn/3D/0eDf8Aln46/Yqra014qBVqUTXZ61cPg+XxLF/81RuMNfKrjzR0C856ggsgacpuMRLqqnB+/wD5Vf2f9sMROTTZhJP/AJf9yuN30JxnGcT/AOifgniH7HuOymaDB14n/wAAIYK4mOXKYVYhxKTLcHUI0f2D8N+PgO7+V/748LyMTyFdkhyWCvnPDmWmOxdv2D2H7Q/H/vKu4TNfSn2WEbB+/wAzCYIe4/ZrRmhx7IyETif3goAZAR+2ili2ln9//C/iPskb7/Ad2+0/r9pK2c1Y6r21CIZ1rMrwb8bITIT9jU4mInHtrxaP2BUs9Nd+mIW1BD9wfCnzv6UMb5Hs3c/Ansv239Qla8mRAg+P+rLVPUulW2pd8lq8Tc33/p/PyTHseV4dgLYzjuZ/TTY7r4Fn5fNu5hPbfffZTMFZv4uwROoZBysyjl6o2V+Fc38P7m/Bjff4Jk2JEynM908awsWhm59yft/Yifaz8v2CJqEQianEziZozj2HmVjhUrQdiYJ/qrMKUYnT2tbF6U1eV518K/sd9t/K5dfKr8ddiexgE1NTjEEt+/z4Az0VnorPRSe2We1We1E9qs9ss9ss9qs9Di1jaVX0EbcB8fc8hL8M3WUcuZTeR7asz2iT2iT2aT2aRMdVnpCeis9BZ7cT0BPbie3nt4aRr2gntFntFntEntUiUIs9NJ6KT0Fnt1nt1ntxPQE9AT0BPRGjjAn2iT2iT2iQ4gntp7ae3+XmeZ5mjPtNx/tkDxK21PUBFt3pV33svSqvrxyz1Zf/APfUHwP7mpqcZqamvluchOYnMTkJubm4e2vh/YPw2ITPt2ZeaN4IaCyZdx4Z2z/p7pTi7BsoD5Fv4L5X/ua3OM4ia+BPbx+6DqZibXl45TI2ZdX6nR/9M2f7Pc+h8rc323+zqamu2vlubm5ucpv9gmbh/b3N99cgfotLqCPMoHPH6biDHv5Sv/kampqa+O5ubm5ucpv57mp4/Z18T25TkZyM5GcjOZgeA/LJza8eHNua1blJq+g4lgVrh6b+qJUpL/8AQ121NQj9zz2Px4zU4ThPTnpwVwLqamu+TaKqci8mFni2WCUX8IM36f8ALNatAQV/skyuzn+wn0rv/t7m5ubm4Z1bMF59OvQpqntq4ccCLWspFYbAyMap1Icd9ic1mx3uPGuhh8tTU18CP29f9LrOR6VJbkXPlB41GE4ThHJWYt1lLYeR7invofDJMwv4vjv5E/PzNQ/s+Zp4Ofx33M6pkGzKrP1INlfs0343AZb/ACp9+iWfq/O+oWL06y33P7G5vt5mj8vM8wn9rc5A/F9lBvW51LI9LGuPmo/q1xDLfC8vHKCzUs+6mdLfhmN4bkIHE3NiMTLbUqVLBbXipxsit6k125CeP3yP2ivhQJqHxB8evZG72O4o+seDV98j+EQnzY0simUNxdu2hHqWKoEPkHHT1snJoxxTW1touKZHotTk/eemCVHGHtr57M2ZucpyhczbTz+xrwv2HyJl9y0VZLmy3+8HDa0WJ6bg6j/VSgjr4t8ry3SGlZlR5VA9gfPa5wp9ogv0ArBMmDwP+kyzTTTTRnGE/DUEDb+WtzrNvK2xfqxKGuspoWinrVGsiUNsL+Rln2r/AIgfND/VjD/ZhROIiicdysraKVs4IjBMfI904rrqrqsWwfJtwGze5yE2JsTY77mxOQJ2IWm5ubi6P7tu7HalTOmfo57fl1Wr1MW3xKW0R5cyyV/kfvjt+rhneD2HY1gv2xV4WWYge1USte+4zhY1y6VlA5iepOc9Scpvt5n1dj5hDT1tQWBpubgaCyI41+yZ1KsV3kwt9OJd6+HZr08r6TT96yC9p4yy1d1aNrflj/y9P/4U123OcFik7ETxkdy0LQsd/Azfbc5/V8iNz0QX9ECcZxXss2floxiwhdtbmbR7im8slvrBB/p/IDTMtGPj5VhchitdH6dDNGPJsYfqPWedFbCzB+rp3AzhasHOctQHufGV2b7Q/stFWaMG5szZnJoS00TAO57D4bM38fHbrmLyRvE6dknHyP8AUGQLbshvIYtLTGOytH0J9FqW6atkadFaL9vgCCJYdZkI3NQiemZwMKmanEzUHKaaaM1+1yE5iFx2WbTtubm4HeczOc3uCcY6jj1XDONePoY3Ob3Mq8n+kXcrmSNOx3KrCp6XeFyT9Dj4Yp0qncyR/v8AkIG3N9+QnNYSDAoA4ia7aE4iMo1uuBQZwnpz05wE4T+QikT0ROAhrmtTQhG4fsO+/l1DEXKpyUZXrG4U5StIRKauVeuMy/5B+P8AeGSH4z0xofRApnmBAO2X4z/6Ii7WA/Ajfb7TMYin37VtTnBglqPLUFiX4uRWar2Q153k31BK7q7Ja54jGZouOFhpJnFhPrn1RCfiPgZvc0ezTrpF2b4CU1bXUETwuTZwljc2ddArMT+RTuufk03OYheWYz25PfU3N/Fl5LfjoY7qJXkPtma5WBlu5p4l30hXD4zO1f7G5zGhBMk8cdTtfuQO5mVetNFtpcVKbLCPHmVnZRwkyLfUsjfj968f6bOn28sH8uxnCBBNAfG2wVJZm16XMHKm5bF2IXWZd2VXkVZd9T5Vy5WLThWsr00ULg5Ia6vOQzw4aozIq5Tmce/Gs9Wr9h1DAVJoUrscllzM1SPaiYxZU25mmM9Nw31z/UNpTDsaYT8TYzCldVzY9M2c1fwYGlXmoTolo4LYp77ltvpVtnPFz3Za8lGHqpDcgjWqQ9DpYagGW4JLMxdNmto5ewMsmLlZrT22TlGvo8yqvQYNKbXqNPUZkZNcyvE6Zv2gEPKPZcB69ug1pgDQDXY9vMHyPb/UjfXYNlTxiv8ATYf0cfj7Z32XO27Yv1DWjjOUbJ/Tupy3EN1rQvkaN7Qvjh2yahPegQ55hysgwe5si4eS8Xpl0ekoUqeYVFDwYdIldKqOpYdmNbRYJjZvKOq2Jf0tIw4tw+oMtMU+vfjBBV+yOwHwJ7bhM6s5szD4bflDHYsq/Tj/ANj79sVtZDLo4qNY/VrTin/JWMRkWWyvFuui9MuidKaHpI4egtVvAA12+k1Do4mpnYS2xwyNyImDm85yjWoJkUYxZKWgtNQbJrnrWPKce51yKXrs9NFPTsa5G/YPYffsTqZWX6FaZvqKudwjZ0948zqLDYW5LZ969k4tQLN+TD6X8Q9m3K8ysr0iwzOxxeV4Md6lNz1nEzUtM1MjHqyK7sb2VfMMMfKehqbhahtAjZKS21bYcJrT7A1mz06Z7hYc2yBL7ZbTxfkaz04Y7geAyh1rpSv5b+DEAG4wW2qVzBDkLLMqFrLTrFpj59Sw5L3FMRrovTDM/pivhH74VRssylSnG/J3hWcT2/rpXT2FXMG7KUg5uDXlxsS+hmXQDCY+YKoc9Z71zGyL2no8h7LR4VVy3JxUL9REXqGRE93dLcC8Ak6xaRv/AB9airjvJxa70r6XQBWi1iHxPqM4vPSfnxacDCpnnsPPYtP6ZoX4xzyd3qx5wy8uJ0+tZbhCKQWG6movWwL/AA9S6ct5wKPRp6k/lXELABrIX3MTp+RkzE6VRjPl+q2F0nBOPMj+Bsi12b7FKZuoT11E9wderZNXGIuZwOFeVOO6wNwtw1otK49awACCX9NotenDoqhlyys8l7O4QKWsI+TGbgM3OZlr8a7cho15iObGpSqqVuBOU5zISuwWAghyDhZKWtYNNrxfjVWxsPH3Zg0c6MfCZaqaais/+sc+VlwdkswW9VOltE6Ykv6UHlnTyjVYmYWoxjwFSCBV7WVh1XpOIDXUlSs4WLcj/FysptAnMRshNqpaL4+b+O+4DHbYu1XODWsSKgtqwWtPcuCMwyzKJDXMY1hC02tXax5giMIy6lON69mQzVviPFmpUdODAZ1Kp91CzhpoU3BQI1bLKr1Pc6A7N9s/IPqYjayO1uRqKmTZD0+wsmCywYZi4YBFJE+oQNv4E6nq17uyKxPXE5sw+uF9Qv8AR5sesKg5rG4GGqsw0LHq1CrStGL+n6qHBs5YyMmPBGTcrr9HHyalsGPjBJhPyV1i/kviVNMs/RygeBoO1lSvCr1RctOWTYpoU77ZeWtA6lk1u2FeQ2JmeqnB7TVUta/Nlgfz2y7CoOI1gHT6AUqRO5mR/Bjwwqk5IkbIAhyGjXkl3auYgN5XHu5LjpoDSkSqvbBJkr+g1DEpiBYy+gzLyCxfxrl/8BgM3FM32ybQiY+GbrGxkYelxAbUz8dmtt0Lcdt2dPpKRbPAacpym5v42pyAsathYCCVtv0NH7/DJH6Vb8S9phbc2YxaEgRaDYzgrMfMKRs/c9+4swX9emr6pSmgfzuG6ywA9VJkWr7Wp/pMB+oa1d/ACDB3EL6nm6wNoBtwbaeiBOq9T9E5N9LrVZjm7DqstCVnl9QIPfZm+w7GXuqIGyMh8Wn0Uh8nu/3tI4cozrNznuJRa5qxkrnCXooWwLFft0LZNo4PUdprb3/x5GSXPJpU5lZ7V7g++YdYeM2rBk1z3Fc9ZJflitLc26YmS5YDcqoghMtwVsavCrQGhBM5z6eHjiil6wwephGu9OVZtdk3N93sCz3FcuSvItUpo/JzMl22VjeDxBFdq6DiBvFl6oHd7iPEdvrxcUq3TlCC3wadT++tWlMR3+qlTa3t7JhE+nFlf36idYKW21H31mvfPPeHZyOUaxItgB6MnNGOz9oPJmxLGMpp4HvmY/qIm6WOaRFzNyu8tBfLWFsKVLGFcpdKrTnz3zGe+nvjBnCe8EeW1+pUKX3ZTGUrAxgsiXbjNyJJiA2SmmusamJLn+nGblA864CRwd7cbD9BBWZl7pNJFkT71/jngtR6YMbHqMyasWqPVsmpt42GXZsKmlKF9LHq8xov2MFq+p/W/jn4q2hulrtulmDpZl+MEhqM9KDBM9i09mZ7Rpz9FvXaUnK487o0ZiIbHJ9B4uP4fDSW1FC9nGI8V7ILYLhsO0wL2rvyAAmH4hM6xv2tVorgymgyrJ7l7AlbYzY53ZSdvbt6P8g092zh6WaHGeJhmYq00KFF1t/gV+F/tftedJYvjEyOJgMHd15KfEMy8rcLTcwcUmFdx1XYRN51i1rvZrVrHor9KqGcFMCiIm4F2baSyXYWSrcWR1otyLE6ci15npYpVi55DSO3qYwdhjK22BEWv1aThuI9diTg86VWFyMtfrpHFqm1KvtdyrJvIKZjbTqBMXL8evynTayEyGOy+l57VPxv8ky0TFu9RFg+3fql7Y735b2r/YV9UcUb30syGd68pFVsxdMGseuhicBKaALEM5LDNjtWYnbU4icRGnXKuV+PgVitcfGWchwxlHtE8QgMqoEZlL5eXRYjJ7n1sTynNMijRgbS+qBVksPXTHD2eyZ3Tp2Mq2YeMrLgVoKfCWn9RvsGlT+LD9VglnmYreneD5U91nW/Acks6kBMSwr7SyHHsnBxOLz9SD1YA0AEHoz9HtoGcdRW0PX4xuo1rPe7hynhvuMe4LMm+qxsZj6S1ss9KpCMl6qf80iJ/wDkNHH1snOopsIesCzATp9jGmhaq8kXYdjdWtlV+XkSuoVVBZxnGZB4pj0IA0Pgs+7Gm5TLfv8AeV+Xs8WxT3+z9ar50+3DZGPgs1nFgONuVEQIOMy/rLUEWUUc7PaJFoSIgWaEPcS87w8jPs419QUsld7MV8l6dG6sBsqNkGNZuBpbbX6dV1atvl0lTqdJO8JBNTqCcq7d88DgGA3UBr4eBF82+pK/ufu8xj5tMSVfzZHhgfAb9T/5qbkLPyyvsEDRUAlw1XhqPTahWhxllnTaXb/G1qUoFS6mviSBGuUXIeVGSeJ5/X0fNtoHUs05EB1NzXw3X6H+05Y+v8TOi/8AHHbIHKvMfG9bp9uMljFRHya0h6njArn0tK7leFhFbU2OI1y39TylpZ5KxP8AkZQlLc6rPvWd10+LLZd9uxAYGl63W1tLZuc5yEt8jzNwTfa0PFrNi/4ur1Er4TL82/8A1hebG1vvqahlfD0VOJyoKf4QTpBPt3ozHKLkrPq11QY6N06/HW7Jx/Vu/wAfjEpVTWOKGNTWQFZIfFWvNzrWFLMA4cIZy32+1tg5TCbwfxw22oGms+1nw1CNTJtCqcgxLw0DmM2z5MqDD4GoGDHbQS0TKH1oP1MTXrfDUPbHfg3LHssDKOiidD/j7GdQCcultWclv5ZubEewLL85BHzWYG+2en7g2stDVtfzXKdZXkLzP3P3+4e4Y+Q33w/FjfdpdOc9QT1BDboNYzzHx+JIluOOe9/HUHYzfjI/NP5MQfrfA969Cz/aNbaBX0oTon8XZmAnUa0tfBroqyG6hX6j5zRsvIMa3IaemDOMFF5jY51ZZbZclYQeJY4E6Xl8ExnazKMWdX/n6bkHI6eraf7w/bNOlL7noXNFwrY/T7GIosrXk6RMgGeohjjz232P37mZP51eTi/zMPI7nvvUyMeovfr2AWdGH6EzLGqx8/i7rSNqEAfIpqBy6yck4/pjIreu+7hMTqVtF9mf6kXmwX0ai2UsZ3YtPO+lXE5H9pOtH9b/AE+36G4jfTMpeTIvH5PWGllJEPrLOLkgaiialrMtiZGoLVMVwe2Yd5FP3xP+Q33+JIlh+nJortSutG6fUOT4qqmPMhedOZY3KvNZFq/WWvHpSH04zII71mG2tYbQ0DMIPM13Pbpe/c7iffrPi3ox4PZ+Vf8AEn2u/PzB8jDGgHdgDPQSHHj4zLP9wGcHlT5mDo3P+Xbc33HmAD26MEwMR7vXxOYx4/idV+nLY/TVbpMPqKV122bsIaBIEmprsBNfDpn88onWP+RhtwZ/MpP6dZ+l/wCX+/iD+oRuEa+Jhu1DcY2z2zUHGs6GLQ4xtzzCyrF28NNuqunWWTH6alcuysrFz8XJN1/R8g2XgdmYCWZKTrHnLMqH0oNH4ampqDue3Sl3dKPz6sP9xX4UHlTR+FH4WflYux33OUtfjkrZtC2/jYxYwTxP71LMepxiNdiR8d9jDWL01Hi9I4zHCCG5FhyvGSgyBarUX9Dvqot9YCPkHTMWBnVR+sw8URvEXyvYdgOx+PSv5G+9H8nUv+TV/Fhtyw6fxrbQb60Rty1ZuHzAIeKzPsErvqWv3KTnuA9vEf8AKKJrutbGJhO0XDrE/Sqj5QEsvdobWrvWxT3tx6rpXRj45W1SF/UgxCYKK0nXN+4b8aR9Nn2T+P5Gf136b/IZV/J1X/k0fw9Kt5Y9MWUnYT8ci3IqauwuFlmRxBvsut9pe89m5lvTiY1d9Tot89S0T1LJYOJAgXUFbGLTK0rWPlVpGy7DGuYzc/uZOOL1PTrRPSzllYzTK8fJaL0vnP8AD2oPZ5iDd9UTIu11hi9rfakfRb9qv4/mfh00+Ynlur79fGP6HTbuGRX+ajzTF8NaNizNqrZ87mKqLL5UqVpDOQjefhZWCVGpsietxJydwtuFxN/HUCEyvFZomIBBWBB9M5yzICy69m7dS82WTH/CwfSPt+z/AFOl7LmVfydYBJxj+ljv+rgXC5B4gYbJ1HP0L0bGaY2NTjrx7b7Gf3vvafqBh2YYwBhTc9FItYEBIOoKyZVjGJigTioG9B7BHyAJ7lo1pMLQmbnUz9dsx/wt/Afb9oKWODSKbSIvhuqrvHoQ8fTIarICnDvW/HPiI0Ya+O+1jagMXtqP+Q7N8ANwUcpSvFkTxHOpbkrDexjNOSzcJhhOu3UvMt/Gr8bP4x+A+I767Yvmxf8A9jvUChp1Xi1PsVli+icSmiyul7sZq80FVyA03B8rbUSC6owOk9RJsS0fX31ApMSncqqCzUywErrb6HbiMqzcUee+5uHv1D7WH6Kvtd/EPx/YLgQWczVXei1V2Uth5W7jpEvutzcmuue09wKsD2mQMNBBQgiDj8dwmW5B2cSy4npk9ncB7O+eyuj+ZqagqYxKgIq9wdnLBawMArWCNCYbFWC6uGa7HfeysWLkVenE8R25QfJrVWAsxTHsaJjLFUL21P7tD5NdWKBK8ZFgAEuRSuLcbA0UgQTxBYhIIjuFGRkWMlNisomu3/1CvlQIqjsPuPvNbmgpavKuPt7pdXYpdbY4yZY16SirJsWsZKxLYOLDjBD2ysXlH3VMenmz1EJDbWsVi89O0wYamLRUk0O3maggQxaZWu2WvU12uIAuvFOY+YNjPHGvLteWWWsFtK212tPSvuNVIrXNx7BE6iyGrqHOJkcoH+rsIPsuoPLaHZftk2aqr2FEv+7ahAmfW1qjSryjcWHplYl237GE33mvHCwVGuJZZDiCwnDo0K7knrZCQZ6Apl02QFYOy1sZXiwVASzjUmAD6cY6ltnjKa2019DQlOl4yxMSlBlMa3T1oKSWCAdyNy7Hrsj9NoMqxBWqJxaXXnkcjU94IM/iR1BYM5YMvcF5M3zmoSALX8t9451aYIB2dVcHaj1vOwZR/HBFWATXiAQqDPZY7wdMrg6dkgWXWYxozwZVerwEEdT0MbGXhQTqXvo/qWGigU16nISxgFQfNvv2/wDqNQhKVVr2OjAqEe3qntqoMdBPxjWajN9THzubjgMorZCh5T7Tc+8GzM2xFVLbbBX4XcBimbm9wblKAhaqotVc4iDxDxIyOmUXsenZNTY+W9YzbhbMe7dV+RwnrWu3C9oMezfoNBSYq6gfU5zlOc5TlOU8wkKGykEpt9TsezAmARhF3A03HePPAhmprzqCEAxW1CoM1CJbWLb1WCrcFc9OLWJwWcROM4wDtrtqBdQM0dEtAcc6cnklacIdwBjOJmhPHbhuaj3Voxya57muDIrgtUzJyGWJVdbFwzK6+HZvvC05NAxmyZrs8LQnxuAz++M1omW1CweqaCmUJ6iyn9VwoVQPOgO2oo1N9vvB47Dv9obVEvz1QYdFuTZj49dFeouoCD332ssVJlZ3pLj4mbdB0+6JhkQYwE9ERUC/En9gxhsMJw3PTMAYTRmj2adQQutGNc7Y/S9CutUHETQnETjNTXn47m4z6FmQJ6V1spw61g8d9fGxwotc2Pi4y/u6mpr5ampqan96gmvJAg/6FhNlyUVrB+1m/wAmJWu/2f7n/8QALREAAgIBAwMEAgIBBQEAAAAAAAECEQMQEiETMUEEICJRFDJAYTAjQlJxgcH/2gAIAQMBAT8B91lllljlRuLHI3ikbjcWWWWWbjcWbi/YyzcbjcXrY9Nxell6biy/Yhv3P+UitH/LiLR6OND1hilPsL0v9j9L9MeGa8fwlrVno8HO5nqoboMaKOwvV14PzP6Py/6Py/6Mk1Px/BQiGFyMfpIrlnY8jxRl3RL0kH4Jegh4Zn9O8X8GPp5y8H400SwSWi0ihaVengaPUw3Q/gegwbpbmSRkjcWPJP70jphiqscPIo8CK4Ghk8iaol6aL7E/SUviSg4un7KZjwym6JelxbdtGSGx1qhejyGL0UpPkww28IcGVwyXfSIzD+oiKFjHDikSjwZ5c1r4PVR4vX0fp4z+UiGGK7I2V2KPXwqV64Ywb/shGXkhEiqEZVwZP2ekSjHLYdYXqj8o/JQ/UrwZJcid6RwJcyPWenlN/Dsfh5D8PIeli8S5FlR1onViergsy4PxMn0fiZD0/p5wmpSFkQs0TrxOujJlUkZPSzlNtH4mQ/FyfR+Nk+tWWXpkjbIqiuLHLlEv4S/wNFC8DxmT/BRWtFFFf48OJS5kTww8HRkhSqPyRJ7v5FEIbnSMcYqNGyP0bIk8KmqM3p3j58eyPcn39y/xP2Yo0rMb4LLWmaO7GyhrRJrkyKvfZZYn/i20qIquB+DqVKiM+WJ2ii3pe7mYp7u64HKy/fwKiv8AB6WNyt+CT4comGdS58nmjJ+xB9iM6Y+7Ho5WT+m6H/WtUWWWWWWU/ftMM9kuRfFko7Z0OXI4bqbIqpKiMJbiS5Y4laZfbZYiijaJ7ezNzYnrRtK0xT3KhpNpsm7lR+zJR2pNEc9OmZ1U+PI9cng50ooooooooUbOBav2QdMUbe5ijzZ27HUtUU+xl8F6UT8e7Z9GOalLb5HiRLdD9ux8e6N7NzL0VDVig7JR8Cj/AGQgOduhtrgb4oQnbMq+I9Emxxa0Tj5JZMZ1oI/Jv/adTIuyIYpTnufB0/op9pGP4zcVooMqtXrRhXJLh2J/G2J8JaQdSGTzNcHUmy5s6c2fjy+zpV3I44eToQ+hwrsRnXEic4Lsz8j/AIl5cpHBGHL7+1KyWOXgWN+S4xHl54Qp5H2RvUuCU+VFEpfIsb4FJyjbZCVotE4o3RXkXqIolnT7IeZ+EKeRm3Ix4JL9uB+nx38j4RVRWn/fuk1Bcjc59hY0u/IpEqMj2ytEMlStl7mS44JzFK1RU3+rOjJ92fjNn439kcUV3NsPoW2Pg5OxdlWhRbOF7EitEuq7fYrSba4j3H6fdzJk8W1EWRZJbu5JURZjlTNxvYpp9zbfY81o5fLubTel2OrJHUn9m5+TvpHGqti2rSa+LMTW1FjZPdF7hZ0+DJ2FEjwbmZkJEBrW9qpG9+Tf9kqs6i9qElLkf9a2SUodiGRS4fBX0ck8MZcnS2rRKz/UJ7330iOUvJvOrXgWTd3L1oToeWC7o6kTfEWSB1IPybo/Z1ofZ1Ym1lfZPGvB8odjqyFm+0SzJqtFLbyLPHyjrwlwJEXTTFlxvyOUGKMDJS7aqX2Xov7FigSwRZ+HFn4H0z8KX2L09LlnSrzpI8lio7y7DZKVsR08b8HRi+x2ErJYI+DoEvTP7GqWjGLXZN9nwdLJ4Z/rI6+TsdeYs8/o6z+tH2JNUmkSyS8G+ZHHkfJD0z7s/EjPu6JelWJbrsyY0mpocjoQlHcTjGNbRsUi2+RjL4GIQzG1t5N5LLTo3xFResr8GSO2CIx3cMx4tqo2rSHcy/p4/wDDJ2KMT+LXBmptf/BuPgtMmqR3HEfbSItF20ct3c6YsVGwfGnJm7GFfL2J0zK3KK7f+GZfHT0t2z1Kbas2r7Lih5L8DZZtVXRKFEV5Hw9MULiNRXki4krfIm/PsSM3Yxd17Z/rdUZ8i/QS5LklRkhXLFB+DoXyRw0dLgWOKGjOqoX6kuSIn8UtU6I5Ps3XokJGeFoxVass3HJPJsR1VLCOKuyEWzNwyEr4kbTYjaVoz1HcYuzIkWUUNUJPVa5IY33Lh45O/ihtMpCwNN/QsaR1EjOuSIu3u9T3JdjH5WsIuim+ETx/HvyLYvJtXg2ifBz4JRb7sWNayxqXLFGMfJKSI7l2Iv8A5GZ8iIdvY9M/7D7MXDsryIxPgeTwtEta4GxMsbsll+j5TOlXcarXL30xvjV6ZZ7I2W58sx493BP00o8sh6So2LDJeCiiiKHpfGjkkOVkf2FitkVSoY1RVkOxJ7rX0NGJfHS6HNDyDlu7ihuZGG18snnVUiMnXxHb0oUbLrgvVjGQTS3EZyJLd5Oj8uWJVwyapWj08/BKjY64MeSUY/I6siU7GmJFfZ1FHsdRyQk2fpwdSIpKrHl+x5WxTbLXkYn9FyfcdeT4H/RsK0ZQoRx2yM9/YdlSYsckco4FNoe2XczY3CLZDHu7EYbUPHfk2IaZTKZFCV686UJFFDQ0/BvT4HFPSiih61ZmUttCdKi9afcTifAtI3vW9bN5vNxZkwqcrEklwWX7aLob9kVbJvx7v//EAC4RAAIBAwMDBAIBBAMBAAAAAAABAgMREhAhMQQTQRQgMlEiYUIjMEBSFTNxgf/aAAgBAgEBPwEt7VuNWMWYMsRg2dtlhQO0xwsYmGliwo3HGwomIomCMUcaXFuYmCMEYmOthIaMBRLDRiNGBYUfZIivche9f4TL6LVaLVf4SJD0WmQmXEyU1E7wqwqkX/g3L763sdZX/ijpamM0JiYtx0L+RdP+z0/7Ox+yMbe2/wDcfIydZRKnVyey1VWS4ZHq6kfJHr5/RQ6mNX/Bl1MI+SNZT3Qp6PkSJMej2PJ5EdNPGYv7F/ZfS511e0cdKdVxkvoSWk9KsvApDluMvuJ72FvsUqDW7FVkiHUXe5GSlx7HIq14wVyPU1M8mynPJX1bsPrYFXrYpbFed92RmuS+4uNGIq86VJbkq22wqm92QqbnSw/HJi0XJTdpFy51nUSh+MSdaXlmd+S6sdDO8NasmluVZU/4k5kncbKDuyHxWkhMnHI7I+jbH0Y+jfgXSS8lGO1lpcn1HKgdP1KXzPVQPVQOpfdf4joyPTyOxI6SUqL3PU0/s9VA6itGcGojpNj6eYumkemlYpUHEp9RCMEmepgepp/Z6in96xX7Mf2OK+y5CVhzuJ3ZilcS3erLf2baXZcuSewre2yLfRF+BNiY6ibuim+XrfTY2LobLly/0ZMyZkZau4zYutNyzLCTb2FBj2ZUg8tiMcdbaX1sbG2ty/8AZTZmxu5i0d2S8nddt2RldmSbt7JcEd43LFhWLI2JW1sW9yWl9KKvIqFvyO0QjaQ9pJl7bCkXHbgg9L+3EwO2OD9y152KaSHNSdiSsyw48FSJP5Gz5EJON1FDp4JWe4lbgUUtGvYmzJjb9yelibxRn4Hs7klkrkeNJxuia3FbWnteyubrnW9xRMWYMxMTE/Esh6pHA5E1kjHJE15IxeNhD4JNWJuzFL9CYrlLRaIxuYD24LmSHNEkpeBQSGtUxSRc3KUrOzMfoSsjghPJtMqUciXG4haQ2NhJF0OQmZFxl9HJLR6w9k9iErxQ5eBshHe+lVcltLojpcvrnfZlSnjHLwKqxKNT48jUuGKnExRbTci7DrJIhNNXHU/Rlk7GOKEriV3o9ipuLS6Rct9GFR8IjRqHYqC6X7Z26flkq0YwxW5c43hyVd4KT0c0ZX1jpcbI7yFuOG4lu3pNXiSd0R6eL3O1Tjyf00d6HgfUI7zlwOrNHdm/JnfkavwRpVH4F0/+x/SplSs57Lj2t2IVEuRzT4FTlPk9OvLLUqe9yP2iT9lR2ljFElhKzGRb+jGT8HYm/B6V/Yunj/JmNGBnT8Hqo/xF1NRrYec3dlj/AMRd/RuN6WIUXPgtGkOrKXGw0/Jhkij8NyxmkKonwhNstbcm4J3md+C4R6lD6qxLqJvgVWflk5OZdCLWFJRkOpFcsu5v9F9GTluZFjLtRsuTIvEjGMt3weoUNkiNVSZJFiM7IjuNFaOURUzsolRa3iXtzp+iNP8Ap8DlZ2tuKnJ7yY6EZcnp6a8GEfA/x50lVd7Inu9KT/JFWDcjtipojg1iToMhFpmQ91YsigxsnwRejEs5XZ2YvgVK3Am8eDtvlm3skk0XlHZEV5Y9IRuxWlyTi48bmV+RRuRqNbHeUnYiXsYwSFgn+OjMI+DD9navyyVK3Bawlq4KR2Kj4Z2agqMx0Kh2KkR05+EenqfR6ep9EZGTRGr9loy5O2idFvhkKEou55LZFSlOWyYunqQeQycck0jsVkU4TjyVak4rYoZPeWli2liadth16i2uU6tTyT6rBH/IfoXXJ7WHV82J1rvjSPIuBTfknKTIyUYXZnfhCjZD2IVamTu9iVZx5OS+O7IdQ3ux9UkLrIvwJ++OC5W46lNcjVCb5PS0mQ6enHgqUVLyPpo/ei5KUG20KjBcnbh9DceCU77HccfFzPJ2Jv8AFoV5Lco1HwVd7kYIdNNGKjZI8i9id9K0pKdok1Jq7IU3ON0Rpyi7lSrLiw2y7ErkVBfJnSPKbaJ8XHK5cuS4LbkuC9kUfkVLpbEYyfLHGS4ZTd5HkjIT0YtmcEvmywqbp7xPUPglWvszLW50VsnYqfF+x7oWw9JNZIk9rITl9GEpEaFhQ1zL3diO+nUVsajRHuS8E4VCm4JWZOMVxr22M6D5MqfF+3/6W8jkRp3/ACRF+DJHcHK42MR5ZD5C2Y2Sgu42IuNJlSkuUWtpFvwfkjpqtpWZV+JiWHUgvJSn3ZWRKLUmOrnsVJJI6femiURMyMhy1XLKZLZrSpDyhSMzNE5LVEmXKdapxyZ1H5sdxryOk0bryS6uMo/slKRizpX/AE1oy5f2LllIrcXFLStJKWxdcshNuXGxabGmtMx7kVZ8Dk2LSNeUNkOvKXEUJzv9Epf7E1fg6X/rWkufYtEin8kSV1YWztpWinIVPe/J/wCkpDd9Ujgd2XsRpt8mMIGX1pYsdP8ABaS9sd2P8WOpi0yFaMuCr1CcrIlXi9rmW5kZE3tpYRkXZYfB3bIvdiXkTuXsMotJaTeuJsOT8E6llud1W2W5GnNvK5U3f5PcWiL32EdxndZcRFXErkpK9hW5IVsfA+o/BNIlLLcW7KsSjlfZnf8AA2i8fAlLwX+xyRe/B2pS5I0VElNQI2n+UjtDTvwdtvdI7KLKHyRJR/jopwStYzp/6mUfBd6J7iEkMvYjJ1PxZOnhsxF0h1ImSMpeGNZfISt8WU6qk7E6yiTld3FWt4O8xTSR3ImcSpJS492RcRcUhNPkdGXgi3DczuOY5F/ZexQ/2kSs3t7EruyHTS8j0v772L6XKfVuEcUidSU/l70rl7cDu/be3u//xAA+EAABAgQDBgQDBwMDBAMAAAABAAIDESExEBIiIDJBUWFxEzCBkSNCUgQzQGJyobEUUJJDgsEkYNHhNICi/9oACAEBAAY/Av8A75bp9v7xJq1NmVuBWwk4BaHe64H+6a7LQWq4Vwt4K+OohT4/3ihKuVcqv/ePX+y6qq0l8N3uq/3a+FFMhaZLh7KZ/wC0aNK3VaXnyuqU8yQUm4DE/hqBaqnYoEGlhJeM3RTaRPiAbeVZTPlyM1RarKTBL8WApDaDDXKxCHBc1wJ3Tx25u82jVdqmJHyj+EL9pz3GTW1UWMfmKb4Zk6dE0u3pV/B3n3W6FYKZ8iqP4QBUts/0sHffV3RqsSAzN/6TZzAyTn9JQzX6fhD+H4q5W8Vcq5VyroVph2VMT0T4kR0zEufy8lJjA2F9R4oD8q4hXcruW85XcqT2Lq6ut5VKut5b5W8VcreKuceOF1dX2LlfMruV3LeK3lvLe87vhRF3FfaYzd4Cia+V2zX2TxAHGMDx3egHJQz0I/G1/Dy4qWxHUB3NoUIug6oW7EnSSPMVCH90zt2IjKVEk6G75H4S4gTP91LTxThi9vSajPbZ5th3H91vmePlCLi1kieakTI9cBNU3TUKRNUHSkBz/uj3n5QifmdUlUVdQ5LQ4w+hqFPL/iZpsPwXNh/Wb+yGSRHMcfwcv7F4bPuh+6mWBTyBWVyroZ85b+VUixA03a8KbHBw5jZuFdXwKBH9pDAd7EbMgptcQeizfML7FhsAL1/CW8u4XDyXmcwDIeSU+H9Q8jkRxWRz4Rgl1Mpr5l1cfg6bJy3lRCd8HS3jQYV2h1ROEI9ZY3GF0OXNanVU2oUkcyqtBnLG/wDY/DBowV7oppXRDYBxa7kdicqqgUl8XdNkGDe4AKjT3TmR902WaFuG4wqqeXwwthQYXwt57or7D9ynxHbxM1JHKOFTyRh/RTD02GFDCGebBtNafm4rNEE1SgVZGXL8BfY3itJK3ldVJO1XyPCadLP5wAAnM0QhMsP3WeUs4/dd1l6I9MXDkcAF9n/RhcjCSew6YguE+DGGdvByAe6cllM2NDiJKhl1K0EHboZKoElWY8m/4NznXJrhCLt22BPFtcX4FPwaoH6dnO4aueJAp8RZnOOX6VKG3ZqtPk7yqcaH3Un0KodiqHl5gKPr64w4nGxTp2lgE/vLF8uSKb3UHthTCy4+ykDVXT/1bdduyAkfIJ4dFxVVx9/Juua3SuKlZwsiCJSuEJVCfA+rUE97uCmU53GSBdd2pTKKf2wBUKRsSF94Vpfm7qrVVp2Ikvq8+avjZWXBXn+BovHZvN3uuDYg+VQIMMzbLxXf8YCHWqAFgpKfFd6LUFRRoX+4bUxhE7+XZW8y6uqV7Y2OxRpX3ZVWEbMpTRHyOq1T4Iun8svTCZvwU8AnKeEN/CxRbskdSinjbuqV252W+33VNqTLC5VvNtsFp3vlKyuBBbeacVPGeJbysgm1QXJVNFfCmDjjWuzZXwm26kRElzC1CfZaSiHKbXZh1UnTY7opOkeyzZxJHI6yPhCbl8WI49qLS53uvvHKjp91wVZFVHnFraZBqKoVOwxkFNTKBCogmEcRh02fEs3yyEfEaO6+G6fopzNOKzanEXCs/wBlVkT/ABKG/Lq1ZIrD3UMwQ5/IhfFh5D5XBUwiEcAh2UhsviH5RNOdxcZoBADgtWSSkGFHPTir0xaUCofHLpVRjfbLnTkEC3UOalx5TVD3V1vKRiuyndIUwc3QosdnY48l250U48cfoZUouYMkIBa9KmCCOmFbot9U13MeTVboU26eyvNPaG3Cl4VaLUDmK5Lecph59VwKDR8xUk4uvwTuLjyXzSNZlOc2I0GVKpgzTyjGU6L1wiQjxGYK+xmIn2U2hpamOy5Z3CvLut4K6NCVmgkBpux1lm+Hm55lPO0HutLplcVqaT6rT9nzdySpQ4cOEtf2ljuzlresr5taLciuimx2VSi+4U5zXivo99h0Q7riqH3VIYPqqwTNWkqk+dDZ6qmAqiTyVWgy6KmM08KRTZXRbnfXVRWLh7KbYJBVQ1vdV+0MHaSnnJJvLipCESOpVIQHqqMauXoqF5XzKoRDdRCMxTupRN/kTRbjVpEk6NWLCJmfqCD4R9lKLLusrpOaVmgO8LpwRaW1/LVazl6Kcs8XlwHda3TPE8k1sNwIH4SI6dLDGvopeiaOZR2APqRKDW1JUEOc3MGr/wCR7GS++ef9xUwCVUALU8BH4hzcF4ccFj+E7OW6qgFvULMyxxzsOSL9QRbEGV/8r/lZIlxxVFqIWaCSx55WK6dUfiyHRTmXLSMoWlt+Kyx5w2fz6pjIPfK0TJWeIcg+jz8+Rzhxks7XUWp2ZaGH1Vh7ovhZS36Nh73Tywh7lNB+VN6qmMxwQzTzdl8KC4Zv9R3JPIa0v4TF1LwgyIOBCkKIFp9CsrxkfyOJZFYHBOq6LC63Cm1029FpqFmaVVwVyey1QfdaWuHohmiBn6nKX9TnPIKYYqU7BTzH1WV2cnqhmaA0mVFzeqKTgCOq0NDe3l6WErVDVQW91OYXLuvhinMqUWJM/S0I5IMhzc6So0AflagZCv5lV7fZQosH74Cv5kQaFZG/7jyQhMaBMp9JoLquuFwv6mM0DNKTVDgt33/wobmi1CtTZP8AqF1lia2cH/8AlVIQqCpOiFzVTMfRaYas0IiTZflagTbqpeNCb/uUpuiei+HAH+9aMjf0tUzFiLMAXL5j2agIhNeSm2h6rKRlKk4L4gLz1Kk0ADGgV1mzLfW8VxPkyYNXNa/ixeSOY+HD5BVmUDDa2Y+UiiIlq5HgiYfeSkTqRl8jlngSbEH7oZt81K7BSmuuIk3Iz6nIOPxHji6yieAM0WkgjG+0EP8AtL6E8GjkE+XCqI8T95LXFb/mpmK1fM70VIZ9SqBvuqOPovnWQOc1vdVJcjf2QESRZzzL4TWu7qjRiXltTeq0QxPnhm4oHGq/L5pyX/lVvxkssI+vJUvzRxzSk4WOArIhFpIm4fujgDEZmK+6ap5ZDiAUJwRPqtEJjfTZkziiMzFqePQKpcVpflHJS/py/rNaILIY6oeNkJ6BWVsCCp+EFlYJBVK0kHts1KIV1IGZ6Kb/AGVPN0zmVIepWXK8D9KpEV59ldVV1UrVZNc2cx0QcOOxqnkG8izwy5aqDEIYZoLQXO5lNzuBfxkrqrirn3U2Gqk6hxrsD8xom5eJljlZqKrEDAp/1cWaPxXGZWqIVpdLsqPK57NVvBXVFQK+E0SVW63lwJW5XoqFw9VOaKbQZRed0WuAyrLDiDLzdVMa4zpjRPJ4qTt7geSdOI57jx5KRuL4hFNPXastFQssQ5SjldxCOA4k8FJwtbLwUww0+pVo4cFrdJvRSaPImLquOi5WqKRPkpkF56rS2W3YKn7KyopONUZGXGZRMKY6r7wSRmJk8VDA4DDpg4L75y3ymxmno7ZO31RjRhT5eoUpSVCVKqERkjpLS0qRbEa4JrWBxJsiX38vkVliCXXDKK5b7YwvJccLKrpHkgZES4oBwmAuYWkSKlQqfGyAKng5Tct9vunagh2wpg5UO1WyGFBNTci3+ldKe8VOA13WihmD4mZvABF26P5VHBw6KR8qcSUl/wBMzJC+py67cueFsJQwXnotRDegU5V5nCZU22XLvg+8hJOQwdO0kBSS4I5t1DEKKeiqdPGl1vfst8LfCm3Uif2AQDstcJuVMC4PcJ8FUAnstxvoF4UMnxX2kg27uJVVTUt0riNmq3gg6JFBYPlmpNl28iVp2K1BdFzClLKqVWpVK1GQ4DANazxIh4LPGOr6RYJ8qVRdwKvg5rDqdT0W6T2UmgrdcoefeFMSovZUA9QpFkH/AAX3cD/BTMCB7SVYLfQr7qXYqbQ4LxnT6TUhhPGTBMrO6rzx2Dlujqa0jmFX7Q1UjD1X38I+qrGhf5BaozfRwVIrAO4VIzFmMUHoqOYrNPqt1bi3Ct04S4rWR7LT7Ky5qxCo5VEyqiSk0gcytIrzwd0qmhXwYV4cETifwqnM/iV0TYwE22dRBwMwV0wA/MtTQt0ey+7EytLnBVjH2TWl5LjwQ3i7umtRKljKJY2O3Mqj/wBlR4VSFIaitz91USXBfKrD3VB+6lMreWlrXBfcj3wupMBJWoy7KpJVEQ016rXDPoptcVXKf2VWEKU3L7w+61Tcw0cFSreBGMOI2chQo5C5vNfe+6pFmvDeGmdE4A6P4XYYRstSKhVYEQxoBVSVYq4zdVp3jdx4ofSNrmvDeacNqSlxRmssK3NFTNAhGjAjkxVCytAmf2UgAsoGoooMZvFNZyxqFQSVVZSbRTHxR7FSfDiAu6UWWGMrOL0GtE+rkAYIznrREu0yplauK05j+VCN9nB/pn0cHFS9jhGguNHKWZodyIVchU3Q2H1TJtl0ngUSndlWa3ivvZ91WFCKrBH+SkGuHqgXT5oAYhSxrcbTcstSymIAOipE/dXCDn6iLclVqM3kN5AKQmjIImaqWj1U5jxOa3grjC+3bBpzEGTj7IExnEOrZbpcepWQNY1h4KEGcJqVsCRdRSwNcJ81me3/AJWsjKnmHdnFB09WJAOo8lm8N8h+RPdPPPhlsvDZCaB8ziEB4Q78UGhjpn6XKbXO903tsS2JbMlDMpr7sqbgQEDLC59FvlfeKkULeVf4VZ+yrnX+p7Y0MsJuVWxJc8qpBdLqqNb6rfA7BfEiuM+qbleKTtXgmgMMgBU0V2EdQiXymURBc0dE0faYUWG7mBMKUJkV7uUlHix2GFAaJgB11OcmlU4m6IdFp0CyBoA5KcFgiM/VVafssTN1KG5BHHiUyVSRMk8dnMNRPHCXIJ2xTAhT2mXo7gs7Q/Os8epG63gMJl2WF/KDWmS3v2Xht9Sm5N3jVZW+p5L5lQu90eKsNggoP6Iww6cHkmw8uV/5jRGs+jGrXOfVSJaVpCorq2EMRCKtnVTD2L7TlObTwwf0diOhRH9WWkGzgj4n2snsVC/TszkF0USqriO2JXrshN7qqonFCd1UYEuFStP8qQHkCt6IdkZ8E6iLGmQdZNByzHEbbfGAkDKoXyr7Vk+nCKOuLkRFy5uoWnLM8gh0CqVLxBNUe1UV088k4TQ2inJhHJTQT24MxkVNjprU3Yoqq+xpVYrh2WYxIp9VvFO7pyyqlNuKIgm1U/dfaclpSwj5d6YQI+0CFI2DZzWpzHIzA9EPGZqTfDH/AOVvkDopxGuf+oqTITGjsqsb7KjQOyInpX6sMznBo6prmtJbzCm0oYzGDmcl2UuWLe+11VIUXvlouM+S+ZfOqBy1bFKKYiDstTfVO7oo+QZiklMZJKL4cpFwGEbuNiIH5aHimBpF+CdsXCIbN/ZboHJb1FuZ+68KbmSO61ZoYLJ8XG6FJqVz/CpjCnTMZLuiMW4XV1daKrM+rlRCJlqPKkUe6cj28hhPZT0yTWtbR0TCJ6bETxBpumuMhIp2UOcOgWiGfVUyj0UjFI7BaiXdytI9lp+zxPWiPix4ME8p5kIU9wSm35grCfPDqjoaYhuSap5PETONTRrVCiHesU1DAd8KSZ3Wv7RLoAh/1TsvKS0hpHRWKrRUPlnuiUeyO3McKprsra1UFrTTMaYRO+Bey4UJzXnUNQnxwsFre1qk1ze5MgtMdzn/AJBRfcND+bXpsjcyPRRoTI0oJJKrFfE6XVso6rM5wz2WlpKvTphNqiMea4xZ/SjD6TwGDVSQ2qtCm1oPqpZHLkq48lrE1dXwiFv1YdJI7ZUGK4ndtNOMMj4Tv5Qb7oZLHBzeYUnbyObUFN0SfQGSzSbPmuBVFUhaZuX3TfVSbpHQLU4lW2Taoxi/pUGtxg3BvnVVlQqfjkei32Fvqnk8SThPkjt90WRGyMN0tS+0FtqCTQojmfZoj2mXRARGydeWMVOQWQwmRG9bpxhaW8lvHyxi/hpUI8pIFDAdttw9fIoueLaeqdVGMB36BXphqcFoYSp07KbyV9231qomUBr7CQ4dFH8Zwc58nKJChNmyhJxqVeaLuYTsOXmntgU/sm9kw9EcGlTG0w82lByv5Zm0T5rLLxGWXw8rW9V8R74nSypAl1U857XRA3grrSF8W/A8k4WQY6Tc9BVSWkLUSqLhZHz34xEzsmI4dQh1UxsakwiXFNFaDktK+XYMrbVAq6VWq4BUV0/POTqiSEiMfiNBRLWgdVcSWg+61PkOi4lTlpkihg3zHYv7JvZOZ9BRxkqNDmKZF1VSbdFsPW7pwWoAf7lvtAQyxK9lLw5jm1ChW4txVxsqqbv3WgTVJAbFFKZErSXw4ysD6rdY3uVN0Vo7NU3xXzWaDHaTye2X8LVAzdWPXzs/UtQBCbS6M0MG+Y8YUVvlTU7kaHYOBEwSFQ0PAKcScOHy+YrIxoaOnkTwsrKhVT74U2ZBUmunVTOxRSmqBM/8YDAeZGlwIwCaVfin1sUNWoY5kT0RdEDnONalSgwwPOrhUKik4YU2bqlVRX2IeAwb5kQcXAOwFEEQR8yc4G/BAtdUckDObuONN3yK7B26KqyvxnjXG+ww8sAim9vL7JmU/wCnVVQksjjIFGUd0kMzD35rM1ubuVODlLfpNFrYQeS0tKrt6irhUct4LeCO1XEO4oKqutRJ8kdMAimdvJqQpQmOcnSpmN+SzMeS6fFeHGZlnZ3AomwCJYMsNtF16LLIEJzc2l9l0Vl025Q0TEcWjktL1LxApUVlx2OexJNbwCuq19VMCqM1vKh2KDEtcnBDspcENqpCk1jj1WpzW9FNwVKY3AkgxxpxPNSAEuiqFQLqix3A4WwqVKauqp3gNzOCpQ8iuCt5mbks4ySNVqkT3UshVGH3VIVkCWub3WZsg081rZMcwVqVNk5X1PNZIg9QqVHNTZbDU4LRDeRzlJfK3918RznKjRsX2CBwVMbplQA5XajIz7Ldc30W6UQ5rw7lJUY9TcJDqpLxPs9+IUomeG7qFSStRW2abJAu6ilgMWsbclAcsKqhWUmuFMJQqAXK1uLz1R8Fo7Igw/3Wsuy/SFRleYXwokx+dfE+zk/pK+I17O4Wl49Vvt9FTCi1HAvIsnPPzmex4cA6z83JZo8R8Ry3ApQ4bW+ia2GyZN5BbhQc6+zKIwOVA5vYqQcZLnhJq1Gi+7dLsh8M/wCK1UW8FRXCbPhhUqgxa702ahUUnzCoUKy8iomtUJq0ZmKbHsiDrRfGhvb+4QmtJBw6zomN6YXWn3QG87mqqtFpqfNmQtwKyspFoK3AqNXHCmzIrS7MORUpSOOnBwdfhzTWw2EdUBtUatS5qgVsJOAkszB4bvyqbD4g9lljA/8AKhZbIOKuFRjivml3VT+6vP8A3KvlVKpXZ44cVZVwpjbYsqqTzhywGZs5BdFZWVlZWCpt1V1zWoJwPyGRQZCGYhTdIvVZqlFvbGquEnOqqVVVxXGakwKcSfqqnZ4qytt1nJUGzycpPmtVFNF/CwVMaDC2N/INlGcWkQ3G6ywm+qurKmNBhVT/AIRe8thA1qqxx/itUQn0W8cKDzrKvsvmViv/AEuOFirSQGSiDW5mtXxXuPRSA8+hLkJ0Cm8Zj1VB5FcBEiCbuE/7dlcaKYHv5YCJl5n/xAAqEAEAAgIBBAICAQQDAQAAAAABABEhMUEQUWFxgZEgobEwwdHhUPDxQP/aAAgBAQABPyFmZnv+fzPn/gFl+f6OfMzM+fwJmP8AVvpcvpcuX0uXLl9Lly5cuXLly5cuX0uXLg/g9DUd9Lly/wARj1v8bly5f9K/+DOoYj+Fy5cuXLly/wAbl/jfS+l9L/G+t9Ll/hfS5cuXLly5cuX0Opr8GPS/+Ef6p/QWLH/lTqy4/wDOH/ljqdTv+sC8RpHGwibE/wCJJt0egfifwUDdbYa02VhdA/ECaH1PiK1hnbB2E0FHhiIolP8A8T/8F9TU11d9LqUhH8xjTK4Xh6Tiz+4D/llvCexCzhIYo992Or+wwUj/AO6+h+L13NdFOeI/hXV66PemLb+6f+5NgX3BTT/wzvpUCXHM3Yc3HEA7SyJdRRPBcT/jTUrMCcYNE5lXGOu/SriajhlpdymOShKH/wC1/BdXWZgyfMEBgeYb/fmu+Biq7DtCICk/Blx6cTaG4S8xsYsvOmc56aXCXWsQ5u39V/8AiuYri2DQtkMBlSkCztLOIPJMNZPEY134h1ej15w1FmG5eehqD0uiodFbGpWb/wDgfzv80vXG4l1etQ2YljoWImIapL/Aj0wuK6jsmj0NxR3LnENdHqNqVupkI4tlLsk931E/qepehblmArEcNP8AToZcZVV3qMUssYeTM203fhc0j+FekXCdCG4w3BhggzbNxobf19SjoBMl+460HEoWAAgCXD63aZd/6G5bswTlLV2e0IgjFf5gqAZm4WlNLS3Kw7TiLW4mqdzFS6a7PMSKpTe8zd/E11a/gdMPcJ2ItQQWjSMbzCIYIddviJbMKQB8W3MrYz1Tw94eSvH4mcEGqfBCgwHRd5iwMRkU9/kNeLy1Lan8zM6+IFb3Fqzl6di5a40ZmMcw9yoDP3PxIkqG/wANHQpGFfMHonuZM3Uvs9EycQ/A8QLLHc3p64mXqcnZiLKRr5l/hzamrB6DBkczbNIa/HxwdhNpGwiCEE7EW3pcI5JThmVtmfSU7nTP4GuqblPqCY90xyW7TFuIcTxMF2ie4GkI7tBXyVHZDcqmkOXpNfBEN2xDu6i2TPK4f3LWBhnhfS+pvCUFVuXU0iw3Lg1OJhJ+W18dF1LJzLsxGMxdxyuCuY8n5Nw6PEhUjbOwRarG/cX5EeOB/qOhMZdzDnFFpA2lEQ3M1/dBEUXYM6vAXLvRPEG4NHLux+CCjZk/uf7x0bZ0ALNk2L5xXdxfvHvM8iFWeg07ICsPxMNn8Rs/xn/lTH/b6HVuNnS336Vnf7jE91nmTyukX5nnYFswavknl+0P9pGcEi75fUP+5K/+Zg6V0Fh4S+yNuJdxEdiEWvW5gy8v2l7GCtRgg2Qm9YsIziuvxNiA2f8AgHMSu66QK3+C7R/CpUqVK6V0XjFZSA6YmIwvghCjqtkp2ieCZHRXUdCAlzEsnmICVcxylwn7E2W7n+GUlHBHMoqKqaUNeyUjKMnrEQ0g8QuPcwqf7dHYSrLqMtlv4V/TuXHoErrxMzcapXtHsQJo6vQvvL7ow63LZbM9K8Snoismnr5ha8SWfqIIvArLvHsi2McQsRDqn/CBlYmXKQR/HMCU/gEqd19K657TDp9Z6z0mcMl3L7S2o3LZcXoCXxNJcYfmbnr+AGjCCzy8dpQbW+CAWKTZD5fATE1hDVblAzAUQ25+TorKdIB1+Jh01leli0Vnv8bd5kWsfKb/AAuDLZcVKlRIbmkx2io0R7s88e9AI/MNJiMqVUWW47LnmQuFYJmU5p6EoFZOxBRA6ZfpiZz1HMTKu13XD+hf9FXVo/AlEqVKIxTiW9pfZBV1LTMI6mJtHGWYuCrMPONcrV02iSpobslkVd7E3v5XMJXupd9+d/8AiPTL5kF20pzD1DvdH4zcqV1oIAdCfxLl9alS6dLC2rcsE6meldanzPnpX4MvqJ3jqVK/Eah0aysfGZeO68QHlXL3/wCJlOe4opDDivplt/JCFWis+YIJeDTYH2TzyMvrYbYkWp0I7CWdksdS0c6nY1xGV7v8zUqJ4nZj/S03KlRh0z46cYlqzxK60yuuMhsvqIpxcsp2lBPRaQNRqd5U7J311bcPBp6J/MfUIT/yJR2PqUdjowcSLHyreunxNwxLhFkslkuXEnvp8Q9dAtldh9TKBntK6uZXf8G2noFG7yvGZfh6GOJbriWsbf0wmpBh3pgQOSkAmGWRXZ4xMPZK+BeHkhrpf45E8CAXwaI7f0Lo31LZXY6DScnSpmZ4JcWJ3ZuV0zHo66DuU7w8BMOn8CTg8j3iFS6Z9x8WAywGjmZNc5lAh1Uuo8wqeospUMauEPJH0FS3N/zKINzBeYz/ADQHchdVLh0BSx4TQjgbWP22JNFBdx5igZYhWRGDU2jjmZapcFO+lHafH4MqVWiHTErqO6lRlPLGZ5nHRLuA51AOBXv0doWW5/B8OZVuwvNRQuXUroWGWEwNosSIgS3qcqldKlHYsYs40koxKkaoSIUR2X0scg/Es4eypq4hvRYxWRdL/cYYr6cC1BbsahklldpKtIjlbG2FwNATLdfUrzHyY+J/PV6dj9wGDxnslezO4Yt2JXYPiZmYz6jr8GibVqWpQaivozccT2wAy4mnhwfQSrRWr7zLiteJSBfowr+1v2mI+Y7yGcOJshmcme0k+ZzLDE7jo/UBa3LrctrL7dBjoVHC4lVG7lJiGkiw8awCjj8kjK/FvgqU3KYprMCJ5iVkPmdweZrw+oqbnqccF3YAQya6uGogClwFRcxK6OZ8S3CVQt75/wBZQjRLTwyPPKPdKsoU2dkLMK0JnjGI9yTsfqEX3mXTKZSbl/oQWyUH6Eu7p3egRNw7NeTxKceCYhGrES55OtS+kd3bi9pN1H30uX7l9DTApaWu4muEPJPNM252El9KxLmWGniA8XBnjHtjXLUWTBBOvErGZfaXLhEmIYcBHjxvyuaQp7yjc39LKIGxKW7eCycSoTeo/lXB5qGwwNPCXGinwTLbpi5dtFnmOipvS3KfHXhiGMGo60zzsP3LkfPgephwPEzM9QVqiKyW+psl+WNeGI1PWWuXeIp3HPKR7VDyfiD5udyAmQ6HwZIdjld3K95Xoia/SIydDmFAXLuBAlxzx1YzCsxVCLfxLudPnJcRuyyD0sVlKCbqBVFzcPy7MVgxOITjXG95H7hFXl8QvklzlKch9w2DSWQQfb++cTXQuGIcxVIl7XMdKg5JhL8MUc4nZGXB89ajC4EUS/CBVj94Y0mvcq7PlADVxankYESA+IHgivFRpxAePhU75Qtog27w/MX83SSkZ2yPPdZf8kpvo88vE+QrxMJaRjYUwh27tx2IfUu+BLbFXPKKNxRZrjMsNPqCW0djENT6ZfQfuW8/qFb5gYnNv/UNdM5e9xF7mGPnoMv8HREtTmPD9p3Kudg6Z4ZTqnzNRX6SjEKnx14y/EIwhtBj7ldHLLmLiCoO07O8twm2dWnc5jJFqcJCBfsmXqLNdpY2aKJTZ3DanSOlvECPU4huYx0RclfnmWSwuB5hRDUBUIyjt9SoyMVeIAxTLTEUYlxO3j9pyQUtvGuOm7r+oV46VEgL0qVE6bilk8OaFlB6lRLOgWd5SjJDyv1HvLNKZf8Acz6SmjB3nhEJUOMwk0XUceI0B8Ikm4C3XAne3D4RUF2sy5cqUriYvEy4vUHOcTKqj8DGUgVzcXk+4Qm5Yuz+SKr4px5JZqWkoMs4jcFzBJZHlemVxtDBWoWahHoncJRxA2IPMzVmga0M8JlARqjFX7yuH0QZQIA6PqUOKuXZKuXwqFWoG0ADHQS/ctGnZPh9S3tMu4kIGhnss4wOYVKuOv7RuYZ27B5mPDfmJkUzvbJjbxM+fkpmGUFX3iAyuwyx0vkQOwaXENyHd+prkANTAtlMNIDKTfax7EsMTcJcBW5eJXZRvlfxHNndXaFGvsH1C3ydvqXhYmzkgZWMQh46EbUTfBho9phLRhUALsRYdgXRBXO8KkVw4c/ux39xeoezBOieMShK/HRrmHgi/BATsD9TJkxGLPeF1gnd8QWHirU3KDQf5guXGFUQX2iIxcDYh/Uz3DweYWoC/qOI0drX99EB0BDVva5hQQLzLxXTMeXM+0A9PfT5gIaYVQB6TL1zuiPqBXCZK1g6nB+VMDw4e0E0HzV/cCypw8S0G7QJ4gqweFM/hjo7ly40lgzl3Y7chPcMXYlSGyy6MyPySiJGVOZRpkhwb7+ZgKEqDwLxM8LJ3dQFCpxWSUAWqnmIzyYlwWqEsnCWxKh9y5civ7Sl6T3AowSzrEpeVA9/lhoB0rEroNfKouDDU0cH3HW1Ys2Rc5DTtK+EMVQfE5Elgj4YEou7ubIe8P8AJCRhOXBEVnvaS36ojsP8xAGB0sw+6CuOLNwahxdsFNwaQhzk+ZUqJ8y4xetM1gSL+4OgqR/i4tBVP1K8dgX8VNk5ak9JFn8WI3DnZXE5wJJGtzxFK+UE0qKhLMGq4zCFDtHmZKrJC7HBYKsy6HsTIL7J8QRBTmWBozHclbp4g9kgiYZ4VARoPeUnx9cfuJdOSXZ+VKN9+GXd9QWBuG2OEPzaeO0J6pQsg7Ru/GbMPYJaAPpjVoRVBL/oVMQLvQfzBTAbLP0S1MhK+7tivnnpV9RHaOzYylBMxLgHwG2X10b805ll5hVZ4gdz6cSrU+BE9udpE5NoZux41FfzoQ7z4I5xidzXSQet9D6hPZDWONOJi5ykYyLc7l7pRp8RorzS9pxQO3aKVwc9pdGU7RtvG5bmDEXdFN+OZhxgBTFPaNUBd8ooLHKqTgFzh/lijAeQioRACq93HSOhuC/uaRB4j+IBNifJi3j73KD0UgxDbO6sVDQ+QlXVXqaAPGIBCxlX/lJW1GR2IKDD24+4iALTkYFbzvyoqAn0MXSkcHLAbpjDa/6OJZVbeKAxpw7h1qZvePwZpOdfc+P8BGqg+IjnoM6x+ghFa1LZfuC0jRC0q3iHdl2dq5lNTolG03cuqzdRmdL9ziTxL/sQCUpTHWyNsIxoYSdnyXDqf5/3G1kPMp/SFxP9tI3Tf+sYjeUAn5cEBqMvjotaXrk9ysv/AF6S+NpWhxH19HnMcrEA/fiRQZpAro8oFqLCr+VYpS54DcrBXtyjVtGwkvsM50kyCBWGX3D5ZTKi4nOetTCGu0p/tDUfcPmTQD4bQ7xsTZioKqpeQLfiG7Z3wl1/tEwKS6OfNTsFmETJAan1BQz2mb8aPriTJUqarDG4/SBojbVfAShER0mDOnlE5QfEBzo3AjsvOpFX7IeDL8DiZEd45ol7A5OlO0QEeGZcM4cj/uTFltOkCNF2MN+0PE0/9sDaeiiDkOFRTXeMJj1dwIwPYNzKZXnMzT1ITL6HuuPgWXWA+pqXhS2UsrsViCsFHYjRz2C5ZZbdKlSy656McEbcT1qEqWJ4jynq2Mtlx2HMXh7wx9xRVjuMKmineBVhTV/aKcP1Y/RLxW9s+iXDuCn9sqlKYWsAiX0mCBna63zC1NDkCGvf2X+YGYL9ZlqBZQlBarmc65alA82qU557ssyQDjli7CRw13hUC2g7cmYDi+KDWQaSvtBW5w5SA1Tvi5RCT+I72h5SUL/VSvkrywQo9+4YRLOiF+t3wmNnMtRRyjkxh/5j/ETZaexcX3otRhhtq7gg0C1TNe+0eyHBEKQ5OUQHwWIfmQSeSKTw1oK6Ii0xS7geiUZ91VQ5F9TLcgafkihzBipM95dxcvhA9w8seoNae06m+PCWAO4tUoqxS2ovaf4iaAPAw1heV8HqVkA2PMtVTIfDmKWQvWPCKU/M5lDC/NC2Mr45lGq+HdmxQvVBqNiHqpRsjSFH1L5845HqMZdC72RQ0qEBezYU+JiN+EXcH7uyhl3sKwPYt2x/mU35AkHoP5RbUHgQ1bfMUOxLN4U7C/3jVsB+HuGErW4v1OabkC18wTC9TXAfE2idBryFsywfFb0ChTDn8HLjFz1/KUDUvqxgcb6KtT2hCqd9/EUcTQsaRDkjCH3kYIt73WYQ1Ed4BoxRNAPh3/sgK97Ei0FU4xq3wjwOjUVK7sxX6oXN3IUzPReTLa14nI7DF5doWPebEcVXgr2mOZus2wmlA2feSkWDjzMDzypGPpEfuDwXHSwLUVAzESmbsPLcIDPtC7APLMwJ5XBhHDccykgS6HF1E+E2meMp+lYFK0/G4wRepdS0AliXjtM+42Z1M44gM8wMLjeL3ieNcWuGGXySr/bDLZ7EEg6VQ5qIRYuSzsFwG7j0JLHmpdBoqj+JiLSzgr3FsLJgzdQtairlQdQnBmZgu2ZUe4QKhkKLgMNf6LU8x9orcpw8y1SlpGX9TcdrTrgrxWYMW/0IueKUHM4jQWwsF8PEpnwIthrcXWKhouWZiAsfM0peFTcr2XE4a+IYub6gLVESoyrLNyymlPmHQA5EoTwBBojOZmQNlg+hFxdlVzhx8EvbiAuG13l4U4cpKoXIs/SuJUHIpKlSDxCx4lwRymb6HUOkyCN7CmGON7Kf0jM3F/lKGDkTEoqwwcupe2d+WREsm6F95l143NQ3gGVoW/yQVlz0sgvUCBmUrou0DuuYulYDFtlQzh2MbmPYldoY/C+l/B/KVUKJLvXQIbdE8vM7w0a8u5oSE6COttwi+YL4JlXLUJccgjmyRFUsNeJdTthoqCZkLrQ/5gbPw4EtiVKT4pmR9xM5GZxRLNTBFDt4Qi1qu5eFWjxKDPEFbiuu5fcwSrqmLMK+rpBiqnLU2c1C4XF8Ev8AzKbBjCmMFXebIUp1sDEdh2rFStefBxCi06QV00qXLlxmCt4Ul6HZwf8AEwRGGKAb8GAUI6QS7jro8y6rhmqzOQB4nfryscVZZwLMXC8+0DB7ZY/URC1PEwQnaYYDvi6lIs2iquDtB5RKoYc1Mx26hqkqXFQWge5VuBOUXADMGp2TNxqXJ7Iaa7ymftHsD8wZlTXMcAXeJSn/AMpVsYK6JrUlBTPYiChgLaT4i8DuUqrmK4ubN+2JRddxZcI4UwwhGLzDiDUJXrMGPMuKMdDzFg3Yc/BMc778zMqkx2jroQipbAbjGlHzMtGNwv7kcPlgit9v7j1AvzMp3lhKPMubewmdM+AqDrPqsRF8tT3uJT/U2+8TJe0RnVYrNWE7wzfxlGh/yloDsMNTJMz1vEzZVPMAle8Zf6Q7VvtB+L3Dn+2YKJ0DDaAO4YGPw0xkUbjh+mVuhghGOWbaG5H1BncTTM0QemHomA71yx1QKMJkxQ+GXFVCa7sZhG3QqB26PMeL2cTOIENPuVlOjA6Lol58wlXM9DBfM6uKcA5a4Zl3nuhiPIMEWadtTAX8J4xl3gZUFwv8wX5fUO6aQ4lMiiv+9bH8lhBbbQxHRCAReZTJfH9pUj4yZ3fqJ4avFygJDuPbiPExqp3nGolMu65lb7HbB7HzGOpLbl6X90N1X80VwrAEl9xmwNS7h95QIFe6GCPcIvPTnggKnOXHg/C5OjkgTsVHF6jJdOaILK48iZSt6EoyIYIlkGDwAOL+ZWzPFkdFNHAiF2Cf2ggfn8zJHJ6D7Td64PZixpdtplcfKm1RUFvCCZUxeTGtDuPEBhB33EAZeXic4G7ZfmC4X5gaz4QFQd7j2bFNVMyV8waGM2wNuPfQ7rDoVM/4fEW2IDmNAaEsnZg1GIhIZi/YsHKBt6gaBPmgJbeIhOm8BWJfos4oTMpwpSnf1HqtIHlBrgoj9hj0OZzRUXMo3/4EEpWoF/HEg4Wo7DIvZJ6gnT7zX9/DRBvf4nc3u8YD/Op/0MSGNRyV4i1u/VGYMAKG27tqFWboNbIa6VxUv/UszKKIY3hLhR1wiKi+coCfD/6lPL3wF566szH0EnuUDTBZYXqPMQLNkGJvllK+6ivSSs4mRetoNxnOOGfAgEC3VUfMowcwzt/BPQBKQDCYYP2GPewVxRbD3zF3ZdswOQLbNgeopz9lX/eI6nYvzFX2JiXoi3bxNcvTljvi4jAjuPE8y8uZHUkXM5GmGYCpUVZpqKS3UaDaGMu/BHAGrJ/bAzOplA7UMw43MeRcdopAVcA6/wBEMxulXDXQ6OwBRehNnhMLp3jhLPMUgbG4K2VmbD1KHe2/x5g8a+VZasO+AO8RTLYKr5gVczzKhPJQv98TvhE4TkO8AsO/cgj55gXVoriM2IadL6j2s82ku2TtlFile3KZJwwHGLn6pOJ3QZinIOpoDD2BXjSFfS2f3g4Oe4iKq843ha2jiJNnKrTln6Ue1iDYxDWyYnsmYWcTk6MuyMdVG1nfiBzlKn47ygz+YgpyAuD3h/oy6G6qVA5DvZDlu/mNkh6mFJ5hWkWGDLKLn/twHeFtUuHiZEm3RKZncB+J431KXiFYwQHkWQFFjKW2d5pQoVLHB7pgpSt7iGIw4iUdM2Qm242UYGDrcd0rdj9JnWOGOPEJmFSjR4HvElAeGBupYFmzEcNjUeBByIwOzVfWJCNnQnjzLm9d2XuHivWSelDdpcjF1XdR4yhzLhwqXNgbg4bmUacMyDojxfRZYCRyGY+wl8M4Fzb/ABCZQdYSNN3R7gbqqNXOxHN6hU/YmW0+qIUzIuFaeI3YfNQx3BNx3AgNqFs5II4KqXTS96vuDXySCX8Q8oboxJF5UDUmchS5XunaLpqr8sS6vgjKfMyPredHqZx5tr1CsYcg9MRXCwrcRRzRHLnxLaqKxiiPh0oFLp7xzN+1af2hkT0IQUeGnpOAtOCFafhol13cEqIw7aZZ5l3mnvPIlKECpZhwRvgEZRUCLepkGeJUhZYcpMytTTBUuEjsrmZCGuio24uI4ecG1lAgCttfUP3ZFvzRtVWVbxNQnEGsR9kO1ADg/UyScrcIlmtVhgvEKe4KJUrQ95/5U0hKfEfeYsIChkC+oXBgz5/xLe0xpHzNSRn/ACoJVEST0VuYcfBKboRWVXObaOa3r1KdzQrjEBgZ7k0Lg/aJwmCZ4Og2gd1Ci4hqRErPFF+IFTBSpU3B6UPA+45Bd4mGe1QTZ5NSycrW4rp5jpJhDcLnvOXlixxWMYnojbqFzumXghSQ9YDgecqVzIcPEKwX3IrVy0Lue81RPtD3y+8t3er4i7IZNzabgako/KVO9az4gQArYwVUti2+VRs/LjS1LMyO0O25UNTlztiuZWh8ajbV1h6I716DNejkOCIyMf8ARlVDCpCl4P4muaOYRtWqjvKTY1bUHLfvDM2B8NVwAFFiGTUQiPJEEzLBTiWKj7W42hphcqEFZt5lo8TwC7g0x2W7dMOjjIM+EYwnL8QdivDERRyn3MaQG7EnBi30KWkINZvWcYVC+5W+EsNOdmJXrJz7ic3x3lA2sfiahAY1Lm+lJkeIIhcml8Q2tOKuJa6FWHvpweujD3Vn0RmDF75JrSEgrKFGahQOcUZR2wwY9InparLT8Qz2cJdp/Ca4uqVUaC1LzHQHK2AWykIu1UoYK+KUeJrEMhpIYqPy0oN8yrv+PUsqzunlSgK6vcoyMGvwVYYu5ri24zL9Va7UXtOemmZW18hipcLVa8TLMepGnmiKM98/vyuJdzafRzVZwxMypUqHd0JKSCncWcwbwQKSw2PMde4l9KJDXDQyp7iRlRDcpGR2Q9spyjjsh4BEc/0BqZuT3hGCfgZz9Yl6/wBv4SVhvcYiaRTYFwjKnfc55ZRLSgsepszgR9gS4MZpMPlHzmfKPiXWCuBfh8y1LLUC0vndTEi67lRYNTD+CiYEvGouzFLRqYezM3sTBqJmcaYnL0CFRIdpczucKVV5E1Myh13KKVrGsEtc1XjqBzAcFUBeZTkC3l1F5qYdkafvKiVL7UFRfiIstiGYjqdlzbjya/bK97RP1HsyxLV7h8VMT7qm4PDL4IQ6pkDcZJ/0ZOJXDwRzqcpm4iTI/SmhafIRUCmZdIFxisWCrVZz494bavsZaSjfPBAA7qnO+SYLN6lShbiFjTBIfEArTM944IomJzNIsbd3ueYLPqR9l1PaG8G4c1O8FbY/0nehDX7iOCE9pcwBq4l9KxGjENHlSaiEtV+ZpUwh2rLAmQiMBu0BEe/Ar4ZUs1pLR3xHzEChd+9y+D3ZcZrO1ziA/Sj/AGm5E1KdseEQEAf7xBLUTHEud/MwCUNSWZlVAzBL46GNkp143Up4+ggfhnvcTuz6nwyNogUx5hlj0YxC1yv5lPP3HuXsj40O5DLEE0ZWeoBTFpsPLmFV5OifG+hFh5ahjl8Epa6o3LQZAjDEfoch4hQDG/Amfo109+nDbQe4NAFMm69S5aB8kqSgnCH94tsow5q+4BQD9QhnbbfxS6y+YIafRQNvVrAGgVqb6aSsw9g5DoNuu07VecW1wIKZwxXZh5iszFD5nlqKXLly5cRGcQWyjqLREi15fcwaYEhHaiTHvlHBjwtL9x0McsFhZRo7xlUJcqR7YKz1uHA9Dz4lEwA0wkO5VvHPaVDQBSxEbmcDddHRg9puDRntFq7A1Hn/AFYuNYFWjxGbp7Ru37mCYMO6VWo7uVJuMTMsJeKeitSC/tZwVLH3CXlfjUJMzBHOL5i1RlkE/DMnsDxH5SpuOZXV1tohBBfmNHfpFAziyXIrTVC5oihZacCADkt3S13n6hR/ZWv5llXXRtlGudhogw2VzkgiQs9EKkSsrk4lKKdHgvJ+oRUE4g90Tuy7BOZ4mfVUtQ1q4ndeJVal9KYCvMpCB9RlVBKuF4Man60DVPN3hBJ52j3Vhi7tMp49k7oQc56L5l4onhBf3loNWX0S8blrrqgVxeoy2SwId24m003B6844Y415g3muzF4mus0jnLchp9Eo6u3Il2p130ZtzgVhDZa4I9ik7sWcORwxXGx/EVUBRlubrKdpeYPNwdY9SnJ+H+8SqgK4GNoMyh8k1MdDK6ZQOgE7SE5juVN+uCYVmz1MjfX6itnDSF4ROPL6HKdwmU4EGv0Q7ZmwzzPuULUS5Lt9RM2oIbXD5IbjL3LTv6ly4q926G5pfbEXOZmI0z4MyoXTzP7nMKqky4y8MR7SwmwdnJClSmHO/mPgihamB5IwVNqAX7EyUJerEZLrNKJXlO7cD0gCfoQLPiCDkO2YSpVwOCVXTaJXRrpcTPGZsMdHtqCqjfRzYq/0OY5X3LsqG1yGe9LmA6F8JfKP8IeyG7Z/wloKo5EGI/JM4Vs5xZmlo7kqIIOGO1f4hXK10uKs1qz5iDUV7fU0z7gavQCp9KBuM5WKZXdP+mCU9WuDMrY+YLvD8Mw/ORIO3wdXK5xX1C7W7l9wfUdt/DBIpwta/wBRFv3SOVyj0TcZgvCGrOTqb6E2yrj6BcZYSsVGwLge4ldtnaWJQb13mGOSWNOGZnEqWSozvcFBLJ5KwJUeEy39QGntz/4kNDfHL33g9oo0ZjvZZ1ucCQw1W8R0TLpcZJT6JtfYjM2JBOxUPlmYV6nEydkwQt6nBV3RT0L6lHiU0CYPEYRCzDMHeIqLTX1EppMP/TOXqfoTIceoXL2mpr3Ao6hGUuZUvSe4cUwIXNzBJgJpqU58mO0zFr7Ea8txzB9swl9e4FBBblHKNI7ZWC8Hq4QARAS+lMqup6Qbl+GXU7RYthj3BRmaEQHtrtLuSCNqO0F4XDxHXJ6Ivgxe2GZJq8EB2IDUd8ouJQiG1zPmJ8SlYjo287mj6i2dptn6UIfvquSX09RlMqHPEALnBUsLllWlge/aCbgJzi4Va1zCZLJJaDsiphO8FU4i7m2yaly5cYvpVXZeCUrds5lZyO85Te4jPQ1DJmMqEobuNcQGq8Yg0xADRKxbUCVF6zXeKOX+o6lvMOzUw/wnpKG4OIdrKam/0zEr7dMotOvf8FTTPQWt6Z9zWFRJrJnEzZFMuGfdg4uKjmWKdbGMxg+Te/hAf3Fn1NRBVFP1BTyIZjVM7WS9d3btHjP43EMQzPCLEGCpkz98P9/EP5IdHcLRzUXsEAO8BxObPXeM+mDYo9KR2CWXaZe6wGo5tc33mKzriKM38SrwsbTKWrJqG3CL13nC7xEdTxh1qfzDXmEuFWYOWP7x1R8QUA7DBYdZeJmMDMg8h28R3gBcPQmnd8wyqPKkXY3GuYdjCGnHeVF2+25g6eDc0MHYS+ly2Ybi8RNO/IRME1yhYRHwS4FH8zI/dcK9PuZu7z0adptMSvUt5mbVQCXUoBtnnz35hO+HeMu4oyTIbuVtme6y8sfGY5QXzKYTmCDJ8zBuXaJxiA0dp3idnEQMZiqnTaWaZg4fYqYz46mprcxYvbKYys6EEaz5MSlxmgfCNOMw+BM86rl2XcMfKDYZxBAovIwigDwSwIU0wYEwHiXHOD3Mhh5isjQ0B5iAF+YswGX6PDUB/wC5g3Dk2Ey5XqUY+Knpx0z846biko7zGFBqZoWJwbYgpPS3NTGJ8SKjs8S4FnO5paPBjgn4bh3hTS4yRNty/wARcjNpTMy1Ku9xD5VuQIUa1mJwrkDqCMw4nE1i8XmO/Sh+4ANsWK12GiMCevEAzr1Cg7G5lzPggeCcQtmQujxDHBzlXhcK+Y6lZRuYLm4fEsnzbIxLvyuZP1HKVW07VFDEswYgQuL2JXwqhbXlrcxgOwmZUgGEh3j5cTA+64b5Do7nOVMmor0s+4XJDtdN03QgoQ4OLivcwxPufcS7f6hpGRl0QC9CoI0ygBb3jl4vGIPh2mXcsqsGY83Rolbl2UgloPh9RVUedIC1HtBdMOWhlOGbCUnzH6wWjf7Imwu7tiyj4tzP1zwjy8flKecz31wR8AF94Ija+IbiMSMM3a/CBKEqpMGuHDEG3PKrU5G95ggF4htmYral+m94hNlSpykroAlyK5s3PrQhZj4eInpbmYdiAbUirdtWoXtL3Lj0ivs5cv8AnIXD8jA339QVAPiHgJXCRRL6WeY3yeZpzMQYh+s8SzxM/b5j5+JtF/MIK4NDzLXNTiqgZbCorfC7rmPiDPQBxU4sBKzRnoaS9iOyXuYlrPmsbPqBu3F1M+buENXd8f4maA8MuEfkAR3SDA2m7XolYGr5aIePJ3ME5wPcRaZ85YDdqNV1vawZmfEqZ6JvjoSv16PMhmVtb2kUONTslzONeJlXduH9lIhZ+0K0JU1zLQrnmZmQPE3cW945NxUWGDNjcNp4DzKcJc5iuAr4jwL9SlM+T2IYGfjUJmMEwZ4PmFzGviG1zLNlTao9rlR+0KY/meNAaBEtDy1AlJyKqIhhkFVLWdNUlCNW3e0i7g8SkqT5jZZbdQ2qUhgWJb8IvdS5kEHvfmCAY9pK9pTtK+Yni49p9xba/cvFR5nOYFWFTMeWEQeHYxBDVMu5U0Z15Zk/sgeYVYQqBVTyS3hNRsJQtHlsiHwgVWviJhr7vaAClnmArkQB8Tt2GyXgVQ0VqYF/LUG+EC7jiCpT3JyCeUMIYYlZ8ytc/crxj3M9n4iQJhaeZVDviKwy8+BQFGZay3fNeCeAPcGrD9y6rXxANty9JRKK0/wlJQKCaRXAuZ43EWhTvU1nonLauMsCoJF3ig+VSvE2Q3KnS/iPafPQeyCA5uUEpjKepV3fEusDczdEUfrm6IX4lNfqUzVMMnMI29hY2/lpZYC+GGa7lbhAZGFF9y/IhsEH/ZAJVsAQ5ViJ3/cPIvYxHSEW5fYeigvic4vqVFQO+5gyOePdQEQO5tYkX7UQDJTyzapRNaI51ZPc7qZejDoMmCqzQNqeoPh9jaBfIVnNHwQz3aT+4Ez46c9AVzLlxrt0omOgEyQaP4QR2fDUSUcPBKP5szCaIyb+BEuHPiVbs9EFdrvNFrPmW5h5zUEtrbAnCNSwqsTFVRTiHdMdWTKzco8wA4nxL9ze7lu0v2iI6O7CYYmeo/3KRJ64HxBQKDx0ovUQ7JrXX4i9dVGD/E9YRw/o8z4jK37lICUlSpUqVKHEoRN8yisqxxfaJpgDURNoUuYtkqBKlSpUqVKlSpUqV0AI8SbUPygzOPyqVKmfGqXBZNXKlSpX5f2dP//aAAwDAQACAAMAAAAQez9QOO428z12x859L+HDGhKFCf8AksO/9ZMd/RqMTbcYClLN7oL4u0N/DuaJmIfwMCAIW+IMcVM/fi08adB5/wD75wGCk37lmT5y/wBHAdrYEyDiXV/e1dl+ZckNstjTkX8UDzD75ABCViR+zjg2N85x5of4E0bmbxSSn4qYlyT9r0/jmkdWxKuWpC1shnBr5N/p1pkd90x/rjNmLwe7M/oC0GbLqC2741V9bned/P7Nn2HwsxYXkYuutcaCw06yUsc+dtzP5HtpKB5Wog94QrBnX1xIpvGlI8APL4HTT9i3DQrwncN1+qqt4EYxq4aMUi0up8l2XHM6/WafCAM7XSaIK+sLoPXsJmN51Bxp30eBMmOPk3G16FsprRf5DF/2yiVRgayh4bri58LmAeX9Q5tHX64CC43BcVpEwP8AuYVWNU3kliLm/nrKc85uK22JvV2IPrwTi0FDU7y6RT8x737IujdPenP2Vx0q0MGwHri8h19C5noRpSqRhMgxSXMH9EkI5iBzRhYFavbSe3KO4aFiiIN189STWt5AE2O+E5NeY4kR2PgQeonaapQ65Nz3OTfLc3SuHmAN8nfx1te7aP1dBI2axCjSK0ONmUvl8w2qAFIOjzkeqPJslcm51RNtt7lUlU2UegNZ2QAdwrDaePQUdqAYFeetmXzMyC4wa3GMcXlkyT3tGDVk7w0lGNt6m+5D8M4kqvW2Poj6lBXaolF+JBgCG7SnEKYUygzyGx5V8jxvZsxzLeQr4kHvoDdEMOnQ6yx1tUlj4SgShp3SkeV79LVq849PcxCz1i3WFQlVrQO/gWbd+cfsh8x6yJQM62z9y9tN8Zwe27GZpyOR/b9TwgufXo6iIVeh42VnJyjX1fdkzGt/82fbC6Obv1OHsuiFHUIdqKtCiRI5EjrSDj4aLw147xU5Cwlq1jdoJP3WjEIeS5ezpNxNy2K1VKkU+TShmHrAMy2pzo7kU1m/6Jl4HFCx1QgSISi7bfxAILaK5aoodteaNw1ExpqYm7BOVYmvuDgC+aYSwnQCNuplBnERPdPgjaxA25UJnnCDnuwLD2gtFYRZDUicpKwC2injtYrN5aqkklxqYo8/IN3/APnd9mDUR52NMvPbuhi/ndLZyOYPs/idNdYhdu5zETW+RY7/ALtPjVkkdy4c4t/TRLTRfASb4uA6558kA9NVAZ/nff1024BFVuJtqGF6r9QubIgTWlW4TW+u93UwNh9R4x8CkKxDhHfYh/lTMZnuK8uLm6TbKqAutdRajr2rfDTy2dgEng4htzfo2FJ1j9//xAAnEQEAAgIBAwQCAwEBAAAAAAABABEhMRBBUWFxgZGhILHB0fDh8f/aAAgBAwEBPxCXL4vhpC8CxOoBgwFlYJKwWUgZlEnFBAs2qWdRrLdJaDfFRVFxZL1LS8uXwVsGphG8uEZweRi2DLizwDAsgcLUWYrF4X8F5eA5JcvghH8eUvEWOXfDy8HDK5JcPwOSPDSOuAK4ibQSsRJnNIrv6TH/AATexX43L4Ob/AlQYhFhqjCP0g4GSZYxljGFViu35gahxolSyjHlJUrg/EZcuDNY2olqpQpcAFE1AVg+0PhGyL5yd/ySHLK/EzM9YiikglmZqVqLBWYlSoXEMCxvgGtxO/D+BxX5VFYMECsQKG45bhVy8WOsmwSy0GeDBcYFsZEwzaKjWK5UnPJBofJMTN1V9fmIr5FxWUqdGz7hGqowTtFrVDT42izKwsyotlTbKIBaTX6I8OzCrTP3KlQVCztBsBK6ItZ6m4CHquKMicAEqxn23mwmDLDqAFJBdYJvEo4Ry6wBiV2hZl8QRQNp4T5nhPmUmezlBxDDeZf7EO39wkMesFO2jRiZZSCFkUveeI+YdP7E/wAkjwvE9EF2lTaE0YUA3cRyeY2h9v5jmVDEvi5f5DXGJRAIKcRudPwtg94SXKdIjB1cYAdkCIeJnioEzMwGHBUrvAQE8OHrCoVXDcplMJZLuEug+4x2GJtWeP8AXEOy/iKreNS4ZiI0yuL7TMLg8VK5Nw5ThCViD1pRJYR7b4nYI0SgwPK/83xjrMQWJgyXLjctmYnrytQeCHFpF44SFxncvjojqJdyo+P1mWSyKSnUCCY7wEVEpp3AWamJiWx4fTB7cEOF6kVWXwC6gjnJG4ltDOTEPa4dEKFdoBkit3AIRo0d4Cf7BMRdmLikp4HghEJUCPVAO/LncSMGA+oQUi0vEYVX/bgNKXK9T6jUxNHl/csEHYtdvX+oi1V5/wB2maUh2Nyi3p54NwWTEPSUlIB6xBC2LiZdQuApmOMRize4KrIRujhgPBLE97PRhje0Gwq//ZSNL/hhL11gLLWX9yrSMeqo10m48Ny2MKSvSIS2eqXYPvHkdxiu3gjZiv8AJLdZjrPIyb3Bp9N/x9xKDv8AUL9SF0FOqAy0dMv4f44MHEFi8RRiKsE6JrmI4W/AJ6JUIjB44hNR7CF0yvuq+oCobFw5KLNulfHT4jQfr/EG71FWW1FYO09YnaJfGOs74QvEJOzLNi+4lUHsi/WWbmXSOYKZhaMRQiDRxLdwa0spD0lbBt7v0iprtDxdpdfs3/EvBTUy8AuLW4ctfc1hz6Me4faVFHyo/USXi8CwEWDb3Y9SqLCNkouxip1iC8fJHrPBqa5iQIZRLMA3dKVCmrmj9Rab7ymX0+YQG3EIwGdAfUuUD+obr8sKtPiZaT+ojA+5hxFmgQegiVwxxlF9OPohK3G+X1iO7gEneWU/17RzR9QyW4JsusBtcqwrXNLXp27ykYITYuo7p+YdVr7TLosToJkxfYlGm/dqLZFveUhZ94jqXpBuOcqUd4B3gRIS2R2UR+83mENBEmNaQAzynqRX6JY1G7CVaq795ftfd/qCLs+4kLB8Tq5+pd/6lTA+JTPEtWIrJivRmmIg0ZZXBMUylUWxejtQRqZ6zHGX+uLkV+D4hYpS2xFAgWpdxnT0lUewj2wLBMJNhectyQhI6L+oBLXEcE+7BGGMljAGRwizUqUTEZo7SlduaK/JHcs+YFstwVgtcr0y1F6yiCklG5cIk9sqbgz7IhviKUGpmU8pGyAWxGXWhDgARC9e0Q2IsixuNWjlC+n+ZiMql2ob9iIcdeJUsZNxcncf74gBj7Q9FMyY5JghZtiwOXe0FP8Ake9OonUUgexOjB34JgiXEia1g7CS/LmEQxa2+J0uXfCAyh7RsZl+JgfENia2BqEjW6IGo1DUWiINQQMpdaRWQhWYdV+/7hZ/kf7hZivuw6EgMRfX9QJpiLUVJlus+YUoDEMKUHxEVeCA6musqKESbSzRMOle1z1nxOpEwB6S25kQR4zB4clLSlpvpljSfX9RAsnwxr2fccUj7Idt8wm5it0nQgnSv4Iy033gVTcAOV3YFyrP3OgxiZqJVO7XwfuoaBR7PTxF6RBtjZUwCKmUymVM05CmkAtBAuS7rGcjc8IgwLleBfvEA7D+IS70KYKViWFVZOuveK7myfY3C4+T9yq7IxcwrepmKVT4dILAWGEiQHplUCMMwpLqorsijkmJl1AoDMC5cTYGHezJXArTLUuYp5YlcVGBILPUbe+fiZ76fuVKqjp1jFS6dONywtHzHqbgCiEMkVurmyioPeVjLZGYpLP6NNvr5gamYzEVxUSHGPN+C0M0oizJt6ZK8xJRy1MCUbaTrEVq3EODXpBrYibMLUY5gh6l73TCkaBeCvSRlRMiNoXGjhXUTSxQRUzGEfCl1Hx+cQr3L4zE1UFaPZl2y1j9hLIQgFZEuISawlILms8TEisJtBwMeFKMZT0hwcRBhcKvF8RLAYC9g9kVywSwMMPBy+pnmW636SxPeYVUyDKmLqJUSM1+k3MzUgVUMkRdlyFaOx18wkG0X2l4MCtIFo9IPNfMToRKj9Ge+ojl/P8AyLKBfVZYzqGmly11HSTIxxExHGor4YwbkmpGwCKypQ7zKCmp6QHcCjHBSEDEWXIQU3KcCC+J3GYOnA2XLFvH8QxUuMYTCMs5SBZYq8p0pPn9RLDdmKi1sGHEpEXBA9uCwIzQR8CFpbEZg1onQRcmFlRNL8/uZJhhLCVCMepDu7gwBWUPE1MpJ1KaFQRBllRRsji6jXQqWizpLMYbiosnexxmCaZ1yPj+oMYifGZQsv1gskC75q/3MeS77b+YklFmu5/ceU9LluWq8b+4+78/4IgsbivSYbVK0Fg9Y7MmnXvKMVMQCAFZTEVidU/mCGB7St4i1cdQ+v8AsWp/3zGxqZWChbFS3VRMGxYaGAFGoa4i/U3EpMRQ5ILRNHCdIH9xAdEwClwlbd/3mJOGKdTxTxRDLF0I2YYagIqZqmInagYmdmA0hkGntDAcxHpBsqREARly3RBU5tMe8IAVHwl8CFtRmUfqd2YKsLjVRFvcGYS5cKQUz6TPpN5bKypGpaWJcXiiAY9CItvFy5gmUlPy/8QAJhEBAAICAQMDBQEBAAAAAAAAAQARITFBEFFhcZGhgbHB0fDhIP/aAAgBAgEBPxDooi9RaoqqIFwQuKNzKkQl47BCPF1qLC4lNQbGUhYtLjUzLIBvpBodFQMAgkySkoxJVRvmEGJZKuYBBQHcAaiMpdysCAGolw7QWVKYkunMu+lXqGiAgSoHRlQd44g31Xq1ExK6sIy5c1uYypt1VkZpLm/+GodeYyovR6Mr/gpgw3NooGYB1MkMq6GwjXRMuSaBlyrZqXLlSoxIyulR6L0PNF7wJmuAl/rEtOGYYm4iL1inlRUFcsNitwj0HoG5cdf8McxiYhuGarAt3BEKlq29MGh9YxiBbDAcYe0GXBj0GKS+hLjHpcWi2H1T3g1kJaYbl7XMEcmT0VFjkQIUqUiiw/5vjouDLqMXGkMi5YZhYsErEKJtBY10mgZRToQdAFVCLDUzeEwAhl9Fy4A3EbZnrV649oJiLF6AOBuLt34jZ1zA7pgSOx0eIMS8pLAHEsBBvixSzPOn7dCQ3T6zEXhieJaUKU949yMvS0qwmK7S4sZukBs3HdSvVxKjpc+OdNIFTZIJNxixgpiA7oZzgSVTTCzbFsH19qlja3Up5faYdvtDsh3Mtbuf3cQcZi/Rj3fiL1l8dOc4YFuZQVF7YgzYdp532iW/gz+wwl1LIbwAxlHgysUyjcxLqKMPExXfP4hjEvtAsYqVE6YjXELZUyhiELxctozuMrPRXvG2CbYg2j359IgNbzFGphNm1mJYSkUh5dDAe0D0F4dEYHfM5a5IrqZoswh4Io6icCd4jSFyAFlT0Hr/AFRVVjfpCFS5lhhHEA5l9DujD2ESCz0RHaMzFxUY7moNzJOH0F2zMQGvSKqF7waCmql8wQBT/alsy6ls2MxoIBB0DBtiZ6FoszDzHEegDuB3jBGIssntCmoopFsVrmYd4lK4gOpWpfbS61LRhaWalDcG4rLYhzC+YXNw8pWwzxMxOgZpgAiZjiBaBMRUOJUszJuTBcVqUVCACiADZ2vFzeD4Pcg2MBzXBGbDMbW52IkbjfEF0e9lvaK9o56DTc4GAMYoA2tTg2onxpjUdi4mGXMLPuH2I7ahRmr8Oq/cCircY+n5rzEXZHd1/kDArfEuOriMCIYtfWO6FEWbIWjXMG9XHTAGoF5hLMwwxAHMTBBn3iIdxX1lStxooQU1lrniIBV4PtM2URHmdyYCQgGUSkYW5YVUpa8yriCMEVzGkhjqVLqUkpc+zE8SuHQBtf1fXUKBRiRQixpgUOxh/H/CsoXzCcxHbMmMkWZSsXMAxSYTmCzPfROgvvErNxt1MFm46ky/3zOFOOVzsgL9ESONwAlUW3od0Eal5zBrUDiRgOfshXLZCqVQR1sFmsyrUacwI5Yj5JLM/mGGNsr04sQVbCykzGPYaPWABiYrgt5Svz+OlB3LFQC0wTpcWUgekPs+T9wbdH1/UAz7YvziB4RfKH+xpvY8EbFOYZmKLHM8zAXBtZ9mdglsYW8QWNCCkqfjMVKlgdQfUSuIiVEJIy2Sf6DLO8feAf4gtCwJZPljN8ekdeLbX3inC67QzPuxE269M/MAbN+7K8oBid49DjF4TySdsQItoL5PeMSG4gNQgQGzP32lxqgbq4QotDRftUv2g+v6hbkEHtfSdp7sLLJ7XBrotfQj5J+YhL7/AKjQriBWIdxReyWQYuJVAKnLCcqeJWyuWwQtAZJ3sdwglwzDAPuZRi+16lOvbP3Aysn0jqqv6zjx8webPaVlKfX8Sh8xQGErK7+8PwQEXHZDsmJWPgzHFws4iOQbiXZMmpmHCWqVQVmXlHMQCvEEKaT5jBZzMUNB4zLtstbYv8E4jTMVZHOG2IYZVrzF2UK+0Nf7BqEKX5MqKwiVULD17ibFisHN3iV4guWA4uPukC2GJ2DBMFkX8SsTtMMQv0Z2Yk0jItEQVK7cxHcga7mDyYJ0sjIZYus88fuIFGWPL0YCVggCxA9sDWFUigT6zBiMmZRKPWbhvzLrErWdQ2w9vw/2B5H0l/LDAiiWMSXPWoMavrMvHvOx+Ygwh8yhYXF7Pic/4T+FSrJBw3UvJ4IYHgxMbjgtpUtKxeEwyqBMGuzGWVzE2pIt0+8EtKygFv0jGxfWMKTTEqMIssx23wjs1iWP98SgZ+B+orS30JdkPrMtT5/cGOxcy2hSspy9B2gJVX1Y0ZE84hAGfWaIELBfG4VSEe3mYFxPiQTh9Y1QLNQviWVfOZhIMvtLvo5xLGCkMshL45estWX7kVst/SYq9eAlvb2/2GI7BHauJgL3D/ZCgOJQoQUXR4nEVh+0SGgXcVbzXr+IkXy479/MEZjgGoD2otrxHZG7hFzKzN5i3sS9eoO9qHEXUoZDzHbYQ0IpdHoXD0SKn2Tfg4wL5fTcMN3p36Tc+GYDTFbG89tyzQ2oZglken8zJg+NRSsjhpMEN6KmZp3gUmaSiUxpSjtxE1GalNCohhDUwM5mZHaP2OiwzmAQZTW9cxolf2JcHkzCwVyceGBwTWziXrVg+koMRM3eCFlDiXGAZOyOIBG4+y9bJqbOazNGnzMg7l9MNmZc3Fj8EPsQbOg0QFlaBO/p8xTnomKKaLO04RqJGUuFsVc7D5iXGxQdXj7TIMovTIkz/kRKQ2nUtW1GyoTkooZfgfvCQAXsVNx3hQqIC3UHtH3gIdBm3HsTCl0X8RhTBKReYrHghuTDKNEXHoLvcGosn9qLA+sFiOrjI75bKG4JZMkrPQmEpjIMMPMd4ej+uUqu/wB5hAhY8wa3Y1BQjTtzLVUkGL6W7iy9DA2wb6f09CZUeWCgOJcS84nA0AmAccvtFtoBNhBJpjcSsz9gmCcekpNw2zHbcNpH0/2NtZ6A/UNMLfb7RVjUFH/bnMWTLuPTAqEqX+4jr1H7QWUVz3NZYEtr0L/JBMn1TO1EMkZWwhDvMThFyMLCC2sRA3BumIWwlrKgHn7x3DT/AMXUDVCMXcPsDEaVvpX3lUqR5gFIS9FT0VQUFZgIMw7pS7VxpBFXMWuUDRiBc3hEG5zcsehmWq3Ew8y1x+ZkBbO8vLiXBT7wttkAGXEMWqMSNjOFb9YeOKZpmLpCVMLggtQjjX5goAX64gKwHpL0GYlWf2IDagN3qJfy4lkL9sxc2vzr4h/0r8ZihoJ6wmbl9NxfCWMsOzuBre9AW/qVRc39D9znHzLOBAuYmqw7mEmyn0f3CWBePF8v8gmQn1/yIOGWc9FWIC2kh46SBk+83Tb/AFxBuVsvTLlrBgcPqwVdPt/e0A3737hBbmG5iPFWissNS+LmeSeaCApVyo762hlGuYI3BSYrWELFktmViV6BGChb1FkQFuADcaX5f8MQWsbpN/VgTBDzHLHSpUqVKgtJdg1LyiiO2rlSoEzLlsuSmA94tj/wEtgSpUroE//EACgQAQACAgICAgICAwEBAQAAAAEAESExQVFhcYGREKGxwSDR4fDxMP/aAAgBAQABPxBNuXcvsy4W9y3tme2Z7Yr2xWtsti3tBe2W9st7Zb2y3uW9st7Zb2y3tlvbLe2W9st7Zb3Le2W9v3Le2W9st7Zbe5b2y3tlvbLe2W9st7fuW9st7Zb2/ct7Zb2/cQ5fuWeX3Le37lvb9y3t+5b2/cF7Zb2/ct7Za8v3KWn7StLfuV2+0zW37me37me37lu2BTbNn3AzMEKrUupcY/5YMXMcoQMYO34XGLQj0lwyj+Bi0tD8pUbQUt7hCzTFu5Qy4sGLFHM5jqFeEIV/Ai5dxg0zCYmCy5f4X3Li+Yufw24r8NDiLL/D4h+Ll/hy/Ny5dP4GP5v83/hdEuMqVBl4h+RqXLe4MuXBiuMJgTZ9/h/DC1lu5buK7lu2W7gu5b3Le5b3Le5buW7lvct7lvcF7lu5b3BdxW9y0V7lu4L3Le4q8y3uW7lvct7lr3Le5YNxS7lvct7lvct3Le5buW7lu5buWlu2W7gu4LuW5g2fg6IbiaR2/l8RRfwv/G/wwz/mw6h/mv4I/wCHEfwfgYf/AMbly46SEXELYXUdvuMJUwOZkQh/hf8Agb/Dv8v5fzc3D8O/wwj+D8LKlfjcWcQj/wDhcHMGcR1BahtiqMPL+b/yGcQY71CMuVH81E/DCEWe4zcdf46f8Lm/wa/wPyePxzDcIag+SC9kNvv8GPL+D8XCcf48/g/BhZf5Pw1H8kfyTcr/AAf8FhGDiX+CV+F/F/g3D8jWpy9zaO5shucSpX45l/klQ3L11EKEvQTwwjbMQpjVJX5PwvFf4X/ifl1/gsX8D+GbIEI/5G4Q1FhqcvcXD8DLKjtAS3USuJzKgeIeZefxjogpoIfhtPnwcQeY9H+ZkzvgEb/+54o6Z933Ao9rYT5h9+Wo/mM2BpHZ/wDhf4zL/wDxVOP8RnP41/geJSXiO5oxNLjlgEFwBcezjMXGHMdFxMXKv8cxJQ9/4X58QqwRoLmyZvZEqR+m5pkeqwXYnzHOkfmOiv0kr5cAAWUA1MfzoQDYKX3+SP4P8rhH8v54/B/i6jL/AAv8Fhg/A4itZVwA3HaJbMu2mEou6dQ0pV1fmK+KlOyJjUAmE9TET2/cD1cMNketF3YQCl85UzXm1dpS4XBa6gDYXun/ABJX44hCP44g/lj/AJMZcGL+N/g3CcwjDUGVQJar4nPuNlfdSqW8RWAus3LagWq9StNPhmwj4lR+5lXMMxqsq8RR0kZzX4Jcv83n8n+V/wCD+HEbfxzLjCE5lQO2LLPxdfg3CEIzR6lxSptlk6Szztl5wXyiFOYdvnmyWHUOZZ2vwRoUB+omoP8AyN2y5Qo45h4JTTq4oL09/gb/AA/4cy/zcWpeP8eYRbnEY/jzOY7ioxDUNRaJqFSwuC0y8rUDsVrxCqlc3SH5eWSEMJztv2REiqR7h+CKGEGX8MLKpsZdcwwHITGE8k11FF8HcsbDTBBvIVW9yl9pi89VAIRTomLBheGbgVn/ADP8hLl4/FxhLjKjmLmc/jmO4iWFol5LYw/C6jCoXmPChwJe53YinEUTkO9h/lRsj3OgYUYipjlxAggxZfxxmbUYjaaZ1GqZcPuou0id8ywGVbdywCK0zqCKGczcZSUdTFDUaSLmX+OJf4M/4kWiGoswfzX5YWX+LxGKxOnGS6I6sAKwiBvm0U3FRsrURWsjCCijz/UtoKMGCXLmX4BiICdEU4BEe2Is6/cHziES9QLqdEL4KuEO8NXOouIKO6uWGTAczJe9lOfMKt74xKHpiojVYUPKIfij3Mb/AAfnmMGbm0NloItGuz+pisLHmBYKZcPwy/wtQYxWAbTb4IhPZpKUv4Y5o2WtseATDxE4WNx12o6L3/yGm7Y/m1VNM1qBibK7/FijncQuS8biwb3BULc/XLhSNWjEuB14heAuVJ+zwEJ0ca4ehACgiLHUVhD5SjYB0oYBzdoRTl5vggQcBBm1Yacag34j7ZcNfhYMR0FgrR8REgGfJHEVCzoyiwjyrMauRphCLH8qyVoJestWsFxYtvNvEb4Etqn3AVjCsZYRZQ7Xcso4zbDW+MwWwn+0EllxCCQGGCHE9sfwb/Gr8BxP5v8ADtc1EtWqG4yxNTJWUMOS2VGDRo1k3EFTllh9sLqF4NFfzFA/DZ8D9zA1jUOXB5yrBdYxmNiwiUDPQCWwq3kjXzONTcuiL7hUBa8QqO0v/tKYMHZbBDY1XUci0sad7hUTCR2GAoggiU1Lh+G4Vqvkv0hEybpv9QdDi0VMbQicn9QjmAqJpppiNrS/MopdkoR0u+qiqh8mJoLZMRXZbzcoEeUYfg3NIrkgq1GCnD+DiHmJgVSZlmT+IoAMEXWIaN28KgiinsgraHMYFgL4jEeEz78spTobZoRaMSsQctcROQdMAXcZsgl4OB9BK3hibWbCvBL32OipevPUtqqr8biWQU/EuCzZxjxEbOeIKNYqAXeIbZ1qBYNQCHMA+suCvP4r8c31LoG8ZUEOCtg/7jjD9Dc+YTB7iq3PEySi2UZYnQOPUYq4OSFHSg4ICwq3MTFhBzHbPePwIblQ0JdQGpdR5jPBcMUDEVEj1gpAgbfUuFD1KLJ9SptL4ii1BDGrAZwyoUlxIMAL1NKoSb6ghxFfgSjXMyCcxNl78w28X/BD8wm0hN3lw717hEg7mjA5G+CcikDCZBqwZrxc0ZcRs6WDcdURHBaOpk9HLMLRrMRTuXhQ+Y0VjDpGWOZVi3C6lAKnL/hX4s6CYPcMh6ihgWo97KXZG291MjWdQbMh3FW3KbgqtRC1boZTQL7hDHaWx/BV5m3ExNRWXBxFaTR1AgA6zkhYWBVXHcMxcVdE5L3AXReM6lnPwCZRRfiXsD9MzpfMqW34FxDOUWaYRheCuolzYQ5coVbIlzy1mKpZlz4jZO4viAsp0ZoB6HA+IhnSlJVqqtUbeDESQDE5wqNSxuUWZnAZ9mf1C6zdaxLk4+Gv9Qt6XV1+4CiLzozfLncprb+5YsF+IvdfVBoS3XiUaueSXF1eKnELuFJJ1WYdYoLrhMwvkKzasfSYbX9VK2T9R3sK3+oveHPlxCwD6kP5B7nCGKJp9SnQHVEO2/EONMZoheUCSsFvcVOOHfqPLINVS+Kf6iLAHhsZbz3IRixUaE2/uKBeb/7zP/T/ALhY1HJDyl5iOFiR/eP/ANM6QinTEz2DuYuIExwRlljqJKaNoGpn0kNDsMN9kwywA97fMSZbJpxisfMG7Ab3kNQpfyqTCzQBS6qruFQ5va5Ka/TErfLBRylSpYLEosSy27le4HxKmW5XuInhPlEmnMPpCjRUB4L8MQdZngm2OZR3N+Y4u7i6iHI+p0H4iC0Q7jxMLMv6g7FMc9qc1+ZTAbhuzBNM5lPMds0jPBYdw27IBMMvlU8hGys0a0ZIlFWjYrhhPaFP9S7YBalp9SkursjGOlet5JkeWgAzm1iiwzyNH1MICpoRY/UNV7Q0FC74bIsAHsvT/wCV8xLEEuimpjqBcx7oZ1+Ab5/Cpxm4QGPzKXcqV+Laolu2WNMVzMXq49hlHUQ6laBqKW1uZ2m8MSou1QHL7mQySviJmYZJYeZUL0hRpNmOJcDnTBgSzmKG2L5MOop0pUaQi0YhvN2dwsUGMJoFNcQS5vHtDFK06IoE5jQbhxGvBoV1W6+7qDlx3AVXgclnoc1GrAzkEqmpkwMUMRCFGo1URdSukVLcT1SzLS/ZKXccMEDOyOGz7iy4A5QYqp5kRFOsoQngE4pGm1xxGeBgPP3FLupU7JQ1dxiMC3bUH1+fcDuOGX+CILJdR6YRdVQQB5lUVf3MiDiunslqKnlgjSfO6l0ZK0yPSwihI0Zs1K/2wjSkfhSVYXp8ipMEsVMPlPDpJWsMbMv6l2rYgLWaSb0XCzYfcU5ULuj1AvmZMDPAxAYInA/UW5fqad/bLeU+Y6gy467VBAAGUDSXiHZD1GjF2Sr5jgliIag23EE4/cEW19xoSl8EC7ahpTFLoINNvqZOq9Q2x9RVUymOb+8S5de49d37ikDZAzUUNhBcs/gIuPIGAufvg+Y8nnDHbBRTiX+U4yL6XESBd1FgeJYoCfhzUU8ouDBey/ca5r0S/oQhdDQsVdcHuUrf7i5rcETCV4immcYgdxDxF7mTVP1Lj7Uh8yscET5lURLdELGZXhhfm4vionSN3ojfcMbzKdSzqUOpRj2EoczsL9TJj7sS8PDEF0+YuFYmcFLOyDTmXtaJHpRALbT1MoLXeomyp6lpiKXO+4oWPcv2XMeiKGSNCBRb69vB91LWtNoBXDGZxdpoPBLMHORsSFlpyoK9mbfGJn7W6OL4w03GmrBsnZOZXJJMluVd37g+iayN+453kl0Yr1UB2xDyh4RMRE1qAilIW3wR7yGAbvgwTw1DTzHubIpn1jQ+QEHCglgu1YLf8Yr7+pl3G3MT8JeW4Ca/SU1tK8sp1EK4/GZhBDiNVvMUrYRigFOoLpKwpd1UMGKgK7ydEfwmkUbogeRJ7pgwMekIsHIAtFA5WJEIK2Jcvhx9wVkaw3KMV5EEYYqt0oSwB1whluCbtpzAYm7JXjLcQ01E28PAjzzOjqjHzWoYaHEoXVEU/RcoRHcLKLPcQYT5ie/sgDIfUOntYS9ksqzZv9MtWjxXMobt6Sg4r3AK7hl5CAS2UqpS1NytVEHEFMiOZpXMtz/hWYzHcSErgjvo+Iqo/KeAQUv9E4/ExUyHyTFWLzTmDSWT+425YeKgN1+J8T5iJVBArq5faYY0pxT4X+IPF1FO41RxgBHslgRTm0iLWY+lNypTLR4jDbeUqEA6rUGXJpljAkyhXQ8JEvDzBWkPiO1B94gO0TORD/lof8+VzQfEJnI/RqUUYpVTsbf7gjSC2XG70uA5N+oqVWYUdQXZCqeT8D4x6xgmCK7Ayq4jYWJiXaGU1uooU54uYP6UVTAfDEKVt3BO2YOYsTpjyxF5W+JTegI6gYgBxvGJsyHkjqPMopIMVbdrlAz9EVeKInI/aHAEVW0D5lN3/MdtEC2MwQUcA1uLet2iW+Zcnv7mQ+GYFugwxiqAVl7N+OZYadQEqqtuiNQm+0ZpFN7y/wALNStPJMc1DoS+aueWvwlZdTMUihwcj4lDYtwLXdrrVSq8S+j5cQt5PifzL7g5/wBxxtibKV5YvslnTfonT+2KJY+41K/BcUZV8SvMUmFPiANtyjlIco/UQZK+IS8k6IOxg8kbF7i8F/UVlNQvLULN38zT+AGzE9j7lAJFmwlgtkMdfcV5p9R1ktiX3gDBTFwqqTc00zUwZ+cj5PnNEb8cyoLLkVj6SxCZy8Er8GixCC0sMWUJWY4LO/TM8WKL8YY9Ycu1jUt1aC+ibWPuDB/maSDJjECdAHZF+XraLcN3kiXdHF2wcj7Lf9IWyotn0upsgAHdfuAnQALz6dRkgDlahb1WO8nEABdncQbQHa4lytDYMwgr1co5IuLT6lCXHOvqc1zE/Asl6nQTzCUdH1K6SjEISIB7hMH3LVLXgfcq5PU5Vt7g0CmI5BAigCmKgwel2l/MQYhmmJTkJYGU6ZeNTxTcfZD2XFazS+petHCHh9DIZT9GPuFUVMkNRqAyy+OX6hbHiFiPMADtxLQUNpZon3XEoMlO7maFFmeLphU2DUddNwb5JjwixxCAK5suC5Wx7YZT3sRODGB/pBcm7oggQES9wyowI4s4dpeWwY2Xf+uYw2D0AXcK1IjkXWeRgJ+x3Ryf6gXBL4vBM07i2z6mkjoMQcV+4EFCImV+aFMZe5mqsf4QKN/MWptLaukjYFPlOWXdzDkr0MaEsRV3qEq+3eKhtUHNsV4HicDKtbXEVuLHuPg+zMrXzn8GoX4iMWTbAUBYyupUKti9vxF1K8GVPF9QwcnQS4ACN9e9ZbR5YzrgPacAdGo0AuA5hY5VUo3ee60S/HBmC7CxQg6PxArnaq7CWORu7lVgByXLPbVjZUXhMt1w1VfcKjQ2IxaKBI9napigGxwSmQqu5kggGE5hq5HLcXBkuKzsOUfJ4jVnXkpXhjaoNowRomaWV/PUMjIwQyf6h5lMr8ZIdW+okb3KE1UQHW4je0PEuuL+YHoyysUSuOCsQwV+kuopSOW+2IlKjMFM6RdAQ5VX7mVAVvvKY3CrAPxAhftHJ+pkz48y1gj3CziC1xMjfe2WNUCq1HAUGsGIAcisoi0y67gUVMB2ghEZdbiHxqawpl8gx7YVld1+404ZBq8W+At+IHilbb5Fg3gclf8ASoycQM2tzFqkIc5wwKhVr8OJS1NjQY5tAM/3H5fa6EuPU8KsJr17Wo18xS8moClltrdfE0APLXq5Sa3ATLKAJTxlqJXlZl17ljMFmlLN3vP3MNiQYvq3mKqlU5WpfHKXFa5zNq5X+IaVxT0wPCPiY4/cQPLXUp0viOGk+JUQmlclyoB9JS/1LVVHYl19Rw219MsazfUyCh7/ANwQyL6dyjvEELzKmnB3KB1eIgF9aFx5v2Lc2TXuNwQHuOGnzFAC9SxpRG4GBYTMMKUQ0qupWgDjXULltX4gr4qUTUIzcRYyy1lgur5eIaIgKassgVjbdNxG6VkMO4+2CdimoZ6lHrg3+ohe7dBHwQCM1Ur1g3+wMzGSM934ho3drg7gB4CwOclwWMniLWFuB8y5bQlXfi4cFPmXwJzSMIoCJQoBsR0wyHFAfBjZCVIBoMVCWNm87jWSuCsirMKPjaNp/AhI9lpze1go0XWiCOD7hjKylUscjDbFtqtGWPMKVbtDbKhF8DHtBrtjfg/EeBZqXsD5hzi/Ex0eqxFK9dg3FRKV2VRiVleTBSlXk3LFNzoWfe5W2MKfoZdFnQ0zAhhOTM+47arXDTK2kHhWpm42clQGtS97kVh8VqD0+jM3A/BKm2zxKHUppaUhuosUU83uMF1QLO3q4KqyuY8zR1FQzmfiv+H5mSPADZWoY7tavDXVTuW7pI50CebprP8AMsgFQDKhucMUZDI+TiZFseHBbK2DQ5zf7jW8rFxt63Cz0yUvBnzuAuNbVqVOJ6lFWvxlYML7Ax9RW7YA2QNjDUQltbXcMGFBNbYCtkYwosDWNsViz8w2D+IBWInkuApsIBU3fVSlYR6CaRfqXoHGkXAaw+4AmSYlOoCoJT8hcz4OkrMYj++PUD4htMWmjoP7i7XPeWIDLKhV/bDVsMS4BCupWop8VEOX5qUcCc5gHYvq4cTfkf8AUUaO4VX3/cRYC8UBBFhc0OjXDOZVWh5iQxTkbCNlQbGBAPZp+Io1nQq4B8tQKjeb9FzrjepiiWg6HFfzGNhKhd6IwtXhnV7ICRbV8ogCizBM7gaZToELTZ7zL00YLcrz4/mX9oSgIVHxJ3v2Q4Qu/wDbEBTvVD9QRah6wiohH4QHwgaqlKPhMqRAzzAoDXaygCGdTXCJlHJqpg5RZ5oYRb6mL0RS4PFbq8wLW/Zcc2HoWQAKvsRA/kYV8n3LGBPLMbhcq/gmuzmsJXBgIOkC3APcBWQv8LAh5PuCliqqBfMsaPUa0ogWoWyiZWeCVFqNDoPcZTkc6i40rxGtIAD/AMk/iXS1lwxr0OnWkfJcrgCplmL8214iYRu2AWItEG041Uf+9SgG3K6iJY7VWPUZJUWxzxLaqKNWmcYjn1iUwxqhrhGEfFSiAyXdQYYzzHNo+YGsC2dBjqFgFhlBf1gk1Crj2QCgXHcsFp1qCWTyyzBniLMF93AswqjT+Ygo3TdxodAgS2X1s1lmKcB6G6hWrheYG6ZdsTq4tOog5mgxKRzuWq5m8xdAEyFZ9xFodhyQw2r7iFAWwXFZuc0mokLzw1P/AKUHFBXuN0UJdrUrzf03Fpe1qH8wqogzhGYHqrb8yiWWGXKyeGYcscURCFk8i5lSe2Dh9yiAEK6Xt61Go1vak5gE5y5RgOqjWAVFCl2Q2ahQMBfN97xLsiFvF/8AyWLKOV68SgwzpDIirUTWcwkFpFGoVUKKrNkrhQhfHT9WPxEURYZ06KgN5H5RVpLlnZmdM53mAnsoF0FKhESyM6DMQFFWXWacwKQGtuJVsqYTqAcCoo3Etoe4Awx5QZ5HM8U69B7gttSrc3G2oMAbomLwV6qLDkvdy9gVEQSZVQRcUdQDGIfsblbgGSwN4uMFmSEDijthkR3Hl+z3KTlG6mYVmsItBHQi5XYcMwEh1KS3ddzG/Q5gxVrFRU6rxqFDaepk2vymALUJxFrxFKK/pRgtEtXQzBSnqNQx40kY+IqMzwm9t94qKOCgGLQOKvuIWyNVx1CqhxDRXRFAKb8L8wYLGrOJhp3jxEIqyLra+YhwFsympgyGDhgCGyVwuU1lSNrgyseC6NYUi9pHYVn3LoaXgqyV6T8DETVBbuu1thOFE8EFXdEq2Gwi/T4xASQv5Bg7g5lpsiuCQU2LZrxLuaGs6IPAAq1UK2D8INRo4Dg2g2mx2HtHoxQV5uVahtY9HyX9SjPSRPYSuIt6R7JbfnMJ0mr8lTkIIK+yHAZhvR5p/qADGhd226K3cLC1phDuniPTbAPm8xcY5APqDGl0ZQUp1o0uE3ogLMxWz5GpXXBu7U8eYgj33PZ9SsxgyA8xCgCfURM1vFQe4ngmDcyz6JgF54UMyiBsAruAKhR5hu6xd0TzSDAUPc5f6mFgRp2tSxZuqaVLblQ1UVkXdX1AVjALb8y/0N4mooXnCoTXw9YuVDHBarNXE+EemTIiAqh7iF3uhwoPqvcsWiJXTdNacy+ZmhtGlMgchOu4CgvBUxocykcVDszp1BUUBusOYhhi26a1qNVhXuD0Y8E4bULJorOn4gEOKvQ7vqIU1oOA6vkh4PgFGuM9RCqgI0dCHuLcPn+yrcAevqTyP9xFEdhx4Kytz4OqvN04mTPVJTaVyeKlUDgsabA1CmhlVonubVX6mnDiXbMwOoSAVdQfpNNBUywR2YEALEb5Iq/3USwIkA6suLRAGU9Q3x9ag3YwBLipHqIEikrdGozVZGu1X+oG1GX48S6wsaYoOZVofAInk9xXEgqod+peuNFSgYfqF1UoOPcchxuiWUFKU58MtcVBVOd0wQQ7wg2q8sqqb28MD1D7sv1AIpDjcJ0rd9QRrrKzx6gBVuoait9T1RUyoqko4vEoNWwJipdwLq6ODmC03W/+qDQLW2qc61CTGFXPvGhQZrfMcWDA5I9wFcNuYHMY5gNLKtYLp3fuIEacih6D4GEDKFaHtnMp6OAFHjo9saqgcomFebTSiH3jzXMzwxYQMUI00DCsGil10xMRQOxsTyRdBUsj5fuCdH4ULbT+JS6qPNTJW6giWpUK7Igs9GpRCzNLtiVb7V99yzoEc6J8EJWrRS76QS+3s3jM/UFlqGYV4L0EWYrzJcULlKpA/iXw3f1UFeN2NMAKqnCmUgBKVl56lghLWBvPqIkhERRi08bY0HaIp5GNfMSwFUMd4u7q54oa7lxvfuGOyuSsRuHH2PEYZYKlKO8Q63CRXHNxgWrznmILE+H5Pqn4l6CNnCoLVgfMwBrq4irGHeJudkG55raEYFGkKzfw+GVuTLRnr/zORNWAuZyM0bzXmFlf5xmEMxQVzLXRKa8uK+bzcsiKH3q2rrxKE8VSpe8cw4pXgHnnuNBUN5ipUKm7xx1m4gZ8Cp8YS6q+1NPlrAdTo+DvqKtTL3iIPRL9oRh9MbCOIt/01ND2B8DqBlODiPjmKX//AOp/MIBHjI0PQ485iyvUFWf0zcELc0GB9wgvX8n2RE0lgH4jOidGle4pYnOSvuVi6nIK/iCOReX3LJYKJa2JfkgFtB0OA/3MOAfEvtcQHaQCmorfRHyJdR/TFiqVLw2P2+D93FkVaOYFfCFP3F9LewfB6h39Z1/8zLb3IGl3uZm7KKHtVRgHGKjAFwM3olMEt0Xhc7jKytYyHv3MQ0l7irIicKYH1MUWy3YLXXcTpgLqgHPcRSKD17A/1HWpmzR5sD4hlHjaZjdVbD02WbQ8uM5YZYmtg7q3MAeZWzUCvvmCwTZMcG5HWCJWtXbYhfEsYMeYFCa0I96gYZqiF/GyIojKKPCGYoKXq9+7u4HMDQV/jKgOl6nGjHmVwDgdf+OmIsGAqfDj4ld/SACVU0tT6Tq5jgNbtxJsGgvrx/EwlQQCa7L4UdwBYBkULgOa9Sh2SwbcrzA0WjM+LmeZTeDxF0D4RVXVeJ8yoi5QB+oZBSns2SgzhwERu7s8ww2fuDRmoATp08L1LxqksbLGS3AQjFR7s/8AtcYejmqIkFzFpy0WMDWJUgFNVpUdwDgwo6gD5ZVjcsxitUdG4KlbS3fMFeJZIF5iqayaAcvn/cRtS5SZPmswEEuyf/nPiXT3wK7Mo8VLM1oCk8AQyN8qL6YYsHnLp+37ity4Ia9y2cRq5CG5mQVfyQEncmNXxvpgyM5OKfHiCTaNfR3uAg6xDI9MwmREulIuiDE+g5I8jZPFtdP6iqOIIb8vUSMKKH7MEADyuNqQ4RGfiWugG/3HfklToGdR/wAiTcZDqoeiuTDf3EFBwE+D/wDIh1UZ9L0EBdFu2njQvHmKILnoGHPKXcRnpmS7l8PqLFBwYnsIqslvmAK2viBdMzYRAcFE8aJ7MqvERUXtjmCjNDnlcNCoTwY4C3MdL1KdxU5DmpSXkoEpd5mF0oZR6bgzRt0GV0Y2zWfUtD7eLVvy91AwJRRAclcRd0CqygM2LoOFdcRALalhVQPl/UKJVrq/MJQLbQMlwBrVQV/KIHAHw/UKDZa2L7iUJXhpOb5h06QQ2c1UNBUMK3lWWsrWJruy7Rx7riIRMop3T7+ITlTAKX4qG0DZGBeh/wBxUJZyAdjLEMCOdRU0hhPRHJ5Hh8yxGD4gLxec9RZxiAWr+zqWYVhri+K1H+gmVldMuSl3RcIQTsz6gAJYCU+CZrRtcXi1heKXm9ngZVgGFYHeyNSk7cD8MCYlx/YoVTDYo08BgiMBhTMfh8TFkAq66u6AM3FVWLVZHdaPuGXjgCggvQoQPxE0xUQ2rU0C8ELpShk5JeMw3jiNuCwgPKohRR6osy0UtRpbV/cH+JSv6naJgKHvv4ghra3folbjM2vqORBOi/RD3Zys+osxy6TfWgjpbqBVdpf3BXg9gPJnDuXlQIQrjBHgpm0ftZwNv0oTh2QVtLaX1HNMgwu117Z9EQWQ0ZBWfN0TICUSEO0/BgVksVM3fcGV4Az/AO+5eUCnLg8DGyvtC/AH9xQSik2buOTWOW+oU5pSrPNPQH7lwh9DSMo/YQlmBRL/AEhoazKXQ5ruYkA1lR6v+I/g5QbjAEMKVf2Q7Ekuyg+4v2HkTFIUo8rdOkis5HoFu3HN8xCBx4SjVwNjCyxo779Rgx8frWGA8Naj0KDHzd/tMxZxLsFvqFXjYAeL3L2QXSjgG3ucN4yfrcOvGgDjwjK0c0iKa6YKa0oOE+YBiNq46xcEHBVIfgy6jbJO11FZ8uz+4oAtsFVfUp/rCIIZcb/fuGnkFdyjBp5y0S0b48VDRM5SV0QlTgeeJUDpLRyNQ5SLXa7/ANQ08Oax543F7Z8lT5ePW4Xa1Y4eQy/MLsJwoD9Rq4VO0423auYebS/jqaPemEuLANXsPHiAjhiV7EaQDTJvA/mKlJVpyWeM8wGG1E9FgWROWk4WlWQZAsWl5+blRJcB2+KmAEMMY8kewrwUqGosiooZ2mz8Q8QkIp2NccXC5VRMo/Vu/Fw8bwNu8+avbzEbTWi0td/qH2u0yR/qKql4k+BZemrtt5wQdbmRg82v6jYi80X6JakmF5I8iZt/zA6rrdNcxGhMoDOzGWIwLLiPSty9mOTB8x/qKwrVZfABuAxv+GHT9y0oc0r+IofERTCmCetUANNCFw0xDas03tvmWW3rNygm8aGU6mQUvCeYYYsesvg2y/2txj9IsIL+pa7o+o5cyiDEN3U1lfYis5V8zB8mW6fUc5pgE3YtHhLkoSkgvLmj1Lhp9DS1mr3D4ty3HynMIwWOcupUKT53BuvxN/uDuDVbAZU6f4HEMKlspT/zw69SxohpKfP/ACF93OJ/2H8TJFCajDGmcQa9aHYvM2nxzaNyEgJg0ZxHewtCsO7uPO/AU+YdhsdS4ocUyw2QFYGFLHMoXiuaRTUal104HmWVZwH/AHAhLebUfUcTgURF5dwliogCeafrcsHJhYvoDLrnA4j9sIoQqqMVNKj3UPUJQjcPJ1FQzLtP5M4qeConPs3RPYrjCpqolMmJiAsf1GPL8QRW4wsaJrQ3gz/2Jg2dDcKSMqyw+YGVbsMB7gCIAgy5dy4g5hERf3E606g1sUXGieAm8ZYaXuBWIrGg9cQzF2TNHVdy3GVaF+SGIIvNAb8cwgWuYAg6oBVJZBLtytfcv8QoMJ7m9F5BaCBG2mqPmpR/klqa2L6hzAxTpzK1oZqNAbC6N3Dwg5BlFpqqcDQfL/EBwARdii1Fpo0KPSndTGqSrMSp2Dv1DGULz37gNBsl/LTZFNIrZ7ydeITcHPnKjOIxkp5WBoa8UpRFeJa7sewQtWi2qfMtwD3IUOKihqFZ/UuJaZYhSoGydQRWoILBj95i+uhMXOIpUwjqW60Be4pQnBv5RXrn/PYizMXHKdIEt9Q3dXrqIKIUiqr4jxXYgW+YXQHRQbnvap8QqHmkdj1DDczxG+2uogAGVZmQVdXBxUdofzLAdw2v1NCs5WBTT7Jm4bq4oG0VefU7mgOvBFI0bWMwRAfNTbraASxIbarRb7IQFk02/mGEulOY8iFZLfg5jWdIwjwOjb9TJnRBuCwM3H2GobrGoUVbrxACxVYhyvxB7kpVQmQoUas1+4Mlx2ndafU1rihANANRbY9S3Zx8olz6FlJaMuZteAeeKj+5xq5QDBfgcQRUtwa7eIYI8OGJ0NPxGxtMUwwWzfvt/wAR5a2lwwCSAZ3jLTwKg4xRLo7gvZT+XFqvAR5DKZ8FPOLInTLAxtxz5g8hFu2ux6m/fwVhxncA+RLa+3mNXpCAoI/g3BGmbMzMUmZt8sVUKByHki0Ww0jM1xcTuPoap0PH/YhksUx8QCivBDDEuOKYYVRRRrwwarteuoSynGS5VQTlqiKrYZo39sAUs0CyuM6FOPHmJlLJULd/DKOeuQ7L6P8AczesVZ7aX/8AUZ5ZrJqXqrZN3xXQQSRHtqWH1lDglW2uYwUlQ5NCGvUTBliwrxqZjJwNygiEQXbas9WMVqzY8kwati3cy3ItqUDZC6aNxIR0npJa0NxmFDBnqPeWXaSAljiakD4QORiNuUX5zCFXjLqq1jUC2lrb+5QPZVG11aQqJw3jPErSFMp50zS7Nig7Y+MZ3JTq+SIJejjMoWzEByhGJvMChDIuONsCtEBrmWTEGsAOuk3Xhl+/tPgWXqL24QHJgzqBgx1AAtHeo/J4/FRN0gOvUwXixlXTIpuPWgOsmPShB0PqBMy6KlYs00cf6hhCYG/htgASGijxbKXFgPdtwidGtIRGnmPwlnuT/QTF+ItFS+wyZIYF7Lr1q/cRBKa9IQEwpUuByd/UACvKDMXYPGQQsFpcrVMQ9GA8IV1CW7IQSgH1BVDK5rEKB0NFOI8rzGAw3D7h7QvSkXljXQOOZYN8Xjp8wCLAAOoPuMapA1egigLtRcUsA3SOb4PR8xizRn4lHRjUEQQgRJVaA7lhoUotucPHmUXdqsDxyeZhht3z8y4rr3EoLxzKeFagm26hmafxEWzEbNbISCP6iWgpUGT5mp8RJZc8VV5gLwqzl2+YKw1Cg251ALRRJXCVF6LlbqIoiGgNZ2xdEPHnMQCueDKDB4M15hvd9DAezB+4v2XOSH0jk7ZS35f4hnyL6lUVlLwjk2SqFV78GJfYU7WnxuOrQXlKCMgkSxSwfEuLoHCk+oSBUIqVxMs1AWouoAXcxa4Q0XzLioHA9fcNQqWVfonj3M8Ww10x1XZVAmYrUqkp5uLEUBcoxV9Cl1ioxwAsbOA6HPLF7BO6P4mMFGNiDLWeDEWuiG3yxYE+/p28+ojQobFRXJBUFwA5eiEcalhdQFVNiGI2MHg8RhSxWre7amLn5iH3VQMoDQCvogYw1wi1dBi8xKyoULV98x8N6xp7lwMmOj+4YQltVRMtYvCm/iVJm7gmmKIUGNMo1EYzcqolTIzWV/YlRrwQXy5PEC+GQhR6Iio3GqoiFKIs3EWmY0SukqW+71LBmkq2BxR5gxVgjs7BjBXXVMzJpbek+4S1FAP1qIQ10rmM05a1GNIdcW+O5ckhSX/aGICBo4I7SI5Y7V0RdVcZD9P2RrbjHOJXtAjkJDXY3S5CLY7Q2hdTnC/LUS9g7RPnGIMaIrkUncVeFVZCqn4ywZ/WPiIOLmnzKukpzLKIwL7Xz+ozkkJpmKrQyMb4dSihEC1fmmWIPmxyT5WFU/ktPqBA3MF/M1aOwP3BjnnOvmLm1ZNaraRQalvoQozAEVF8AXAQgHUFQlesmYQ7IMg8k4ZFP0EZb3KHEcgIomnxM2MjL/RFo1kZ4tNelQHa8QGlk0WGU7BUOp6hRzqIcasjF5dS1Z5g4PAqTzuIMxcrJ1EQdLBbv5i23lcf1AOo6wjQoB5ViB+iku/2QKnzCcq+z4H+IZI5EKikD4T66jKp7PXyTEE2dMxavJn+JcmMaw/pjXgUK+EaE/bAP1qYyzBMOgdrF9Y9vdfP9JRqzGS4WkK5qPvQSrSwAgsA4qjP6luQSsk5iRhUE9QWvJC87IAlnC2CsvA4gioC91fFwIj1vcJ0cw8lp9cVCIEMZ1UoTkYoxqORYyyILp+YcuuLqnP1DtmtoM+PEa2DGIPLUBChqotpSs1xDiNV7AHHmAIForDKzgdxEtJfRXjmBDgoVlzn7gjBdAHRLXyCw2LDLJGAgXflMazt/sg+A16hDsqb0xITYwjm8i9eYkpXYtf5mQBdMuWTyDEqVFRRA7GssozmUxv9QoQN4UPzD5k5OT4hpRcbxIwmVFiWTDbSomsPC1M6gAAhxBUod/IwAEUzhGpY28EK9pw8ocwUo7/o8xEiVaC/uLFSmRKlbR1vqL41WD+dy4U24L5lIQqy638pm7VsAoPPE9ZEAebOJca2qAR5ZmrWYuq6i/E8m5tCtRvx7zqOgqaM+IKqb4PWI0rXKNuU+IlaEmteXmItv4FV8JKoLGhWMT0d2zWTTM00G2mmu5UlVtHJf+RnLyMcS7UFmbTNfqALOy8FX6j/AJFcpfVynhNosvzgOoxSuR/KoHcB5tZe1HH3BhMBJfeXroYIjIIEbbXjxxLVih/1GkVheIiYVRM8wOavRxAz1NFQWjsLY+pdvMPv0f8AcUKOa8yrFRckYampQQDcS0fwxDMbrdURNdpmjb4P9xMT4KZ9EFLYLArs7lChLIdi88UfcfbAVTz5ajQ9qVIHuMObVW/EfuEVAUHvuO/PG+Ov/bZe+NAcO2YqyWlvLKUwTWvMatPySpuqqIlRC9amYCpja4AoVUCz1GQZor4gaqMx4B1WkXRGBtvX12hFAc5ryr5mAjQvZVfDi0KCZyNFWnCsvZqHAi4VyqTBTAol2ei87qCaI8TZ2KOFvhnCO1dDubFws4fMPEFNVq3fsI8arOUbztDAs1Rn0kTUHRCkeW1SwN71AK2BamASxxcZHaf2hR12ZxfmN74zim8QLaTVg1+oJvbJj8jEUV7zfPqH3Y51XgzFLJ2lV86IBG35X3CVVnPjmLmFlBUQOiDP3L5wF4Sg4FV1HvaLILdrsNXDNuE7OGZQctQFbajSWRIKSt5FqOgbLE/3Edz5ldqUmnC9qPVdTVKIqTMTatMyrKcp2w0zN85xAdrTXQu8+4/85Iq9rLIbasJa5lnuqGLfGYMOetUsFzbvF6MxVoi1qFH9WI0NxzZjhahB1e1Zk9Ulyg76lHcq1G4NsPIYHVDXQJqEoImDtSi+DVnk+YKmAWoLdLiJ7MRlFa/9DmMQpOJhtipy9zidlsFeJSknTUAG7otHMu5dQoGQz3CMZGCs4pgo4hIWajj9Gmu5UioXnG+4jbDp8Wmb5PmYU+Arn7hZ3djPxCVNwZVw/qWIQgNgrn/sD4Fx10Qv7lZr0CcvmuIR2ktoXtEqAJFoO9wWylP+2mEdg0QTSxa+s8TGFZeYZ3AbRcBFwNbDDIcXUsJ3FbuC6llZpq+mOsaEseoo1YkIeTjzFvEocJWSBfssJCqJwAbG36N//ZYQkFNW6h+4rDgdw2AHNwYEHDalCJVBgbtg4iOQvB+pvBmzD/UNjeK0n6gTLrw/uFsLq1B+M1Ck4OgH7IKv48UyIRTPZGXMZeSPbuUV5qVOJsrjEF5dCrmuFbmJX/4hcvDiyi/RHY845v3HVgEM74wSt+4moa7ybTWpRXVZaUcb2balXvK4E6b591GhzHZVqqstEwpY7WmbdX3xFTUYiO7LnBJoh9rbX1FwWPSCwmVDfGoZzEhV563qCjFEWMYO8SyEUU1zIcWMXuDXdOh2vK9sx7VMTc0HDUu1kFgQ3w1ABGLvW6EJLIDaN2x9rB6gOUkEaARd4IYrV4jhqA7z8cEVFy4quWI9UBfmZg5Cu+YgVm+IRMvu9VFfpZvdy1FQma7iC4GnziXGJKmP/dRzyZHe7ip+CFkaUrcRRdmbl2Can1UHCR2AsoQlTFji0IDIrURNcTbyDl9x1SeOINqyxaaNUeIdY6onHzBBxXBs/qAQ0JAV8q/cvC50z1odzkW6sHq3lp1qF4WNVKgJTowZVEAZYCP72w/4qGnaXeoiqZo7fuAOSUHoMwzlNVeG/wCIc1gDFe6F0uEprZJdYXlfiIlB2sUeVb9RThKjRPhqUKh0IyFpqs+ojBK3ef0TJiuL/g1Gfcst57hWdZBMFjVfywREQWJwvPMqHHEL+oViQmd1gpChMhqjqADBRTjAxxVKKhDMutQCI5LxcQsAV0Z61CrDk9q9LmQGlOagGaBVMDRXxBxqNgfCFzHdXKKv6iboq3G8d+I5fBwCglyyFLHpIJpMJzmDSZLeLlRjlQER7Kq7YQIrKEo61wW49efN6h801SOgLTqCYAeTp5mYuir+SJOro4NwzEXGqfuUsBfMqEltVniDqj2Hi0DPS4EgAMPbmocwoZxBwyx43Nh8DqV5nbcy7mPNr5JYyq/RAVTi4ONo2TuEtBt5i1It1n+JiyS8qLr9RECtTTWGDlqTngG6ljK5AoKXMBPaWLw1nRsgJhcRTzMhipUVW5eXr13LILtrMFKmz40RfGuWEyboyHL6lrNvFK+DrzL2075MuyU1BLBinmWOQ4riIvlXG2z/AJAGHM3DCMjrL6iEVt5Wqxpj7iR5Kires1gxuKvVMFqtvncCU4ZSgfMDq1ArdLqZ5hUsz1GoUciB9DNEMPOo2RZQthfE8oBc/EddlmpSDe291CejcwoMYal50cgZuVG0Yj2pRWazqNXYI2d7jpL+9UpRYrbzCMyZy15jRfmNWurMfzCLoH5IbwgV1KzAIoUwC0Xetrq+ZkuDdtwdsfpZWspPDN+DxApqm7aYdYE1ySuwlqWKsoYCtkYxyZiWVGd5fEdWArCvlLIiUTR4yF4iqdeW/E6xIrX+0V3sijfL+YsayXdK/wCyg4MKpAUu60ds2zolLYffUAG8HL3BE8iwjlBo4OZjWqiqGx9cTJOHknhWYdvVXsFfuJXNWZ1AYtwC3TLeyFgOxar0TpUrivmC0W8Um3zmM8XUWD2dXF6llk5NZalhGWwKADBaGwRzZtTDRHIL/UMLnts/iUBa1tx6gkNglR/3CvGTNjLFLdjiuI0ERSD9xwDQoVebM5gFLtWuk2eHxBQtA1rQwdAoaTmNR5HllmIbDV8Q2LCzRmNYt5ryoh0bZdXiAuqB9biJpfZ1AAxvmAnAmHNEMXuVA7inLrzGVoHcp4mADavUIHamiyfO4eEXJYU8bmWDDyvcW3g5KaYmHloAhvmENgKrJT/EL6jlYBKe4yq2SmpXebwJ+4ADVV4B+ppMpp1mZMsrC8PGYndEUJ3ZEt7gjTU08QTgJTf6VKBYYMBEztuM1AU4DNvjiBTWKhB5vzHQxF4M3xziELpKLgK2uR6SWf8AUseYUjp8Q31tqNKXhe/EvIAaFC8MIjIoP1FQssjT/UAN0XE18FtRRE6MKYjyuoOBaDmjGK7isKDIIPluDv7G2BxdtEO2iyX4oUBUhuaGlO36jpddWoZeH1MW86jrCvCsj4l73yBWcXKXJV3Eh6OGLcr4xLVZ46Rmk6bDD2UKMJRxVxUQl53CAv6TOsPFouoLRGX85CLK2cBCJwZUMeojx+yMueNFSgsEsLa+peMN3DbA8S5usxWaMsCU5ZzKF+iAhVcCnEeLdU0vmCmIMm0swZUFMry7ipYZuFd/E0rKDMUIgOY2bpriVhgdepcsSKrsLp1ANrNRrac1LQgh/wDFqLgGY12Vfm2PiDRxM401cTHIwNClM4hrUd4LXcaFKV6Heal8Qd/xiL0LoF+0SGdrHq5cBW7Rt8w1AxER0UQEL3FEDtpxGkC02id418QHtQHxreVhZCBY+Qu7gwPJdkUmnYBqLwhxSb4OQVMdMMBxwPWPiG6C2uGtS1dNU4qo4cCGMWtbgQullXg+kpgYTbNNRK7s/magu+496iiqLCJpUVn9EpqpLKFZezt4R+Vr9R3oxQB8zJDShlZfTubwKNQVQjurbmSPbC1/MBqXZQsZQ1UV5tHiD1qZxDxfURW6vErycxNa0xjicrJa7XHEzpRUd75iphE27oKwQldm0PmMI8phzdag+leUnBfrUENjwvUQUdt1fmK3tYYxBNIZibDDgdagFVu0EgJS3pli0tBfPEFKpvOaIUlmoTiGTkWn+YkI1uJDi8j4ljF7FlfOZasVWazEIq6AivVED6BbSju0z8QIVaKLVNJtPklOZULacu57LZRCgsKF7YV7js+vHcLGbZxcNE2nq8pPbEKl8bhk0iYa3CGj6tKL+YOK5zjvPfR4jzbfZKjBtV3wTp86l0g5k003oO5lgfD6g7aZ8sBFGFdNXb1sl7FiZbzc0pt8MSm1Eu/iGG9R+ZPZD2BtyfDAQUt5XNwK/A17lc/UlzCAc4P2RWkGNQPCwTxLMG/GNQctzVxMNAYC89zJbao1VVPNB7Ky7WJL4qN6UgryARLcKy1LKpCc5idBh6pqFCGS7t6gYFi72BCd2dsRQbPEKMsVqghouUaHGiUXUrJMBs6LqAQrtvjG5gNskgcXDDhSbxV+kj6EpmB7PcOvM4OQ/ivkWV44P3BIolGjCLnI5lmhnK2cZ6jMp87kZ/6IxF9agt+W2BIOlVAhgAYKqM/fWy78TdZaWvSZw12Nte5wwai/f+4X4emi/wCYkLWytSl2Z9wsLfjEtQKv7qKb8Hmdi9pVE89Tdrhqo1xYKxbl3XZQXvFXL/Wl+ksMQeYMEdioMv1HW7ogRq4cN15ljJPSJwYgOpSU7lb5gZlWmYROm4aNGcXAzDPd/gqtkVeSjU2cVa0r1DatVxZZLeqsVRxXMuhc4MHvVSvE7wCWiKuGTk+IwUbRyQJUGaeGYFwFl18xc02niz3La0O9QCi3tgBdAhbS7iFqrQLtiagQbAESw2fxKoGSGragb7gPYyimnjMtMkfAHi5v/UvC4TXUqQ0Ujwgn8w9KlZGczTMhdkMP6jm5soVe2blsjzFLN0PQYmDR0OvUBSK9WplAugherLOVZg9kBKfMqF5Yglyw03m5dbFbriCqm5RvW5dl5b57ir1SXjqCMcqWTVbj4SqwaCUALAczBRrIwZiF1WoAMCOCqP3LbH6lhd0TQG/P4v8ACj+0otUp/cUUh7hecxOVpDLCnj8UeIIUAtVhgkLk4hIjOi1EtG1OL0TApFwtVLRHcqK8kcVF5jNXuH1ouGUwPrllpyCzh8wSIYaCHWRui9vqtyyC6uwEPaIqhGQpo8t6yS7qqzLfbJ3nzEPIKtvasWUruLgr9FdOXQ/pZbM9kBQezL8MyxaviLMj1EZhGWFpQq/mKsOm81WPVRrmtNF/cIemgKjltr6UlwFQYDB1KQV0QSrQlAVfbqU4q2FtqZXgZambYFcXMAfoljqDzbA1F9G3V4ziIWw8PUQhSOHuMQpiL3hl0iGJd6lbqBg6aiVJeW47TJ5IQ6zYPNYlwFjKSng8M2cVGKbB1EKqAW2w7ehPPJhc1cd5wXCDdjGDce4n7hQsHsl//IJWox6yROCItxb6idCRmyJUIVoyR2ymtE3hRfEFINHK3vJ/ctiQMC/hcId3KHcylDyAYcw9wdD/AMvmajpWk8reZSDQKhQxi9RoCms2y6s2yMW14Nj8agsDgdHCNY89xytbQNi+R6SNBuUJTjzusxFQbjn+5X8N0o17jaAmdJWBRFw5d7GB1t4hTdwlIAy29TAC7tRYywWV3XMFPk5gOiONyytEiuidhcvS9EfIVzFT6nbmC4cMyooIEf8AyYlLxVxtOP5S6RcOXEUvwDf1MTl1/TEbFrGEvIgxkTJiriUcCR+eSGW82CZ0yO2JCzHRMuUeVmV4VdW7jcgFqWphVzOqtl1EcM6o/dw7KF5My13eOMBbeNTzH1PJgnqI8ubqiBK+gRmAY1MhTDNMi3quYGsXYmMcmOz6meyV4H4io1K0UMAAqwY3KU8AolgIDCiCkd2OfmWeAxeJmvk6DKDABbwSpcrOA9kMH15L6Ff4gpHtW08VDmQLQlO7hMIDlX7eYOKZQKPrUTisipl1riHKhYs/EBdBD5XBoIVmAgZYBdEO8LisEDFlR0BcLD3CpDLM9zDQ8xGzFvmOhAQYYq5qG+mXtS12uBeip9tVLAICaDfLLjtoXuin9koQOUsIyJUOqhGsw3qtF9M7zjHJ4c8QKRi0MVHsU+hR8wshUUyfJ2yl9Qb/AANY+VSkP3XG96uCzgazn0S8DZZL0buGneuVXw4nAkS6L64jZRwKTEDx6YoZbYYqrBe9zOD8lNSXeiZiV+F19xVEgZ5InUzVI/MYXrlv7Qu0vC2xas27xuFBeIVSxSinyaj5nI0dicyoSBxSvmA3YhjA+Kj0Ve39BeJfGlKd9I8/MUAYXQe7FXMVer8IErPaROU8J8zUP/yWj1FW3i8iENn3VU7siFy8qYpMdwIAqhs3xLrmAJ9R7aCOWMzZLPmbaI20G2DXJjMGmyPlg7NynhURovEPupSzrmHAmLwRTFibrJdsqASvErIS2HVMBQS8WssuEkq7GF2YZAQhweYQrWCCRSpQOGX+bdVXzKSgNvldsEomc6jL3axN/RqFRQsC3hW0zYJMIOuvNlmhlwfZbXlhOj9QSjCwOBvwSlSjBTuXUun/AFA1p+pbCLTbuBToOkOLN9bfuUnE8uf4jMM+p/cw0b3Y/Uth4hoJSpq0u19QDlKfiKE0AcEotqzm+WCIEy9wRI5cFEv9Fiqu5hTL6JiCOWKZn8zNt+iAoQvPiMtHF0VL76UfuKQXqnD4JUQZLqLj/wAPmDJPP9M9bCMW1ThS8ot0DUAHllBsHpEG21Yal4zGLspu4kLxUU4Y4gir1zEUAg7iCXpwRWTxit2g54MVUVQUmdw0G8v6lRb1W5zzGx4VGwupcAC07XnOJQoIP1GoYsyaoSoDhFKRhRmhP/eI1ck0zWI1QKta51cWo6Bb9zF1f3OelwCOD27hFNSrIaZ1Crt5zLGDiDr+qY7dfzhObzdEYyW1eB8Q0spy9xMcLVFVMwoaWf5gOxP0Y9Slzg6MLILPC/7L1DKZW6nz4NTCpd3nr1KOFteD6lVYCiqxAgpANHzbxCgvyVxLUW23qJlaLkrMQA4hChWu2ORMJe0oLxcxMguM2fTK33xYjbFTgtfoi27+Yf1maIwZhbm1j1t5ilUCuZeqM3uDZWATVsBheHAACFPvJLK2HTM8FYvEI0iK6sg79SllKCqafVcQCNCxMI4bjx5Bbqe3j5hi5LVovyNPuUhUGxW/EArVsB54lULS+l4z1BV/uFCx+ot4Y96D3CdRtFDTl7/ruI1cjX/1SwCNlxq3GLuHY/ce9baPFA9xWosdXGNFUzhvUtwl15IlwK+ZmUa1BCcuqiKHC9h+7l7k54LCI0HNSloCBWQdtsoZXLVFMqUastt8RCYN0PyiFETTqKupbqZpyxSJBTltZnm18G46WijtlpaAAzYh9Mq34tpUjQFsaJWgh2u6r9TTKbxVcxoZun2pNb2zDm2cVphAC2UKu3uC+qpwxbR5x/MSgCikK0UwXxcQCBEtZTUdWGCoGsLizsrN4iHqrlsneYCYLTGjt1j0xZaWgQA9qPqWm/iGBwr4j1nsUPTzLd5eAfMFTS5QMn+05PPmC8V9y9W3KJxE25LOCCHpzMswrDv13Db81UqqNrSLlA/if7hUlC2diCjFuY3eddErvV8LAF3b0GI9DfxiU6ii4GCvJCrgNQPMzX1IdnqAztGXeo1Cfz9Q3GM5D0QTdGSrLP4i24MQHEORsMGUAYwtVNsUQtaEQORqqqohgQNGItQJySjEOfbL/RXop77ZTwaieiMAsiP4gibtMV/cDRoEKNSgVxAouHFzS4lHcVNr/hBnuUM3mNt9loP9xm8O9nq2Vlt09oaqg6iSA98OfPVTAb5sduvL3uLlUS4AInvILdW0mscPcHp41oHzBp1ywOC3EsKq+Q0D7hl6gSqAiThayIIjDUwUX7hh14/FIg4Md3BjhffErYXADGqLWnniowVxsH5f6GOonwod253L0wxWf/iCEKhtwZQR0vFCxX5gs8+oNBbbx/UZUluziXJB/wCYgtKUbiSqpgb5BnJQqIRrHtV05h4QlDs1Ckq0YW+4io95FHzvHUTGeGI+s4ldbbm9jvBLFDgSm9U1Gx98qz+4sve9JmotPGJQD2rxXbFFUFrCWy7B5IzfI2Ni4YZ5AomONhOJ0haPyUrAHBYwNrf/AFBLV1LzbqZWSjiNBUHl4mSP0Fvo2zb3CUX3z4iGQtQrYxkolPO0VQajoiDjEjPCurtzMNUdGpdZWBb54z3HCG2PqE6+ICAyJUOK/wDMcOuwPuU1Ligm0cJBY1mr8QXbi27cHqoQ8ymsn/kuLDVWKv4u5WNtJecR2ewYEdnBdxcqjpjlXxjfqGb0FzAssvml4luq5QpvJLLgRcyvlgNnq3KbtHD7zKeyLnF3RM0aAuA7hyxsqg8QFDnzc0DBKFD7gAKJVgNb4gEsnWi0PxBOQTEHxzBySvKFY83arco9BEPZYsVDlOblLQuXpO8pFsLuynlUo3Vumr7F/iYMsusr/wAMMOWtxpkoO+q9QBoawRO5Qa7ls1L2iBGau4oUxsdnGrikkAUrjczmYvGjx2sWKrajJUEbAt9cstRo42T6I3SmRBP5VLlEphtnd1gj6w7t8EYSeFX2zCKHQl8Y5CKwDyq6DJ+pWxsDhLWXj4r5jHL5GXtc3ZysR4UtmxePqEgSy1YWyqC0IR5Qcv8AUWgOqBaLL6rcQPLSYUjsWxkODzGSlLc5F4WnHcRJe/Bj9/8AZmjxIIeDnfxETPGX+5ae9VZMeiCbFa+S5YvQDsA7GNXPawE/mVi56yNrxLCqU02fEFUtG/MHwy4+hiwqvLZBQAOMNVEsiHO0JCSjiKtYAFlbop4fW3Rz+o3xSGiKlVPYxAVHZbXmM3hspCzHUIC0ecT/ABJ0G0OIP4HIY0VceyjVbhEjmhT69Q8InhC+Tn4luBoxuC9VOQ0zLbbo5jQtMhFW4cRDIBxR3/EYr7OKHVlKIoemz38DqMJZoQRO8auPQS7Psf8AIQkmwu7LXKkS6f6zfcGNgacHw5nLEP6RGQINWFvqty9kXIbiWKCru7miLeWAPcB36IWMhXZgXDgL94mDGKAyrxmJip230IjqI2QTxMJVXPv4j+1rLB5R1YcTPAlYG+HcSFGYwt+4DB5owXtO5ZF9Ng1bx6YGEpc0AQCGjrQO4IUw8YIUDkJSaojgTLywN/cG/lC7qNYB6GGS5sOaTihyE1XxPUiKhqwtie1FQcvGI3BDakD1igoqu4Y25VaforcvPYDcyFDd5fxEqi1tLNQ6ObVlDWWDE52sw6QbH/sQY58o5uIWgHgIQ42XZZxEdOCwOj3FTeB8BKg6auzKF1U6cyg2KwlI+ZZ00i8g49znA1hh4eZZUkxqaTDNV6H43C3Vl7q4uw53iBlAHlDcUbPax1OE0SggfIaxLW0t8E1/cgN/cAfKNZ9QizeFceFW5v4gQdapFP8AczBQrc65sw/qXTAgHFrzl3DStM5k+NwwEb6j21P7o/UOkwjjmrYR26I4OUdWPcW8F1YgcN7fRGLBLYp5GPZdwEdHsYqw0PT3+2AAQ9gscK1F5FyzTFDVnqVbDLTZmIuQ3wRLy3jqAG7X3uaG9635gUcyz3spoWIIsSAUhdU7AEJIc4XiXAGWTnyMGILggq4TiOyiUNNHNql0M1ZdxxLU0W8yqZtQbo531MqeQNReJeRx6nFS9nMZN3jSeR4YkIFh8+4eS9cencdrTeDiLueeqh4vmpksHHNciq8rFdoDQCeLPMY5nJUN+/uA6z5qH51Fpn4ILhGXZ4gcqxNHMI3RbllC1HzC2T2oP5ipC3agExDYC0JF0brJmPPzEawHmuY4MFIBR5GKHRs6uU59SxngEeMtPqXCmwn/AAy7EFoZswzBk7Txi6ljS0q/QBuZ/wCxWFdmIuVwwJlLZ4yqHfJR/wCRCDbXdqr7lAixm3b4hQvvUFz8qwqzlXESMU+YV5B4lOh8GJq72iGotecI6KRbb+JSZF04r5lpAqq3zApxLvZFniIWWolI6pqd2ygLS4uv4jKhU7epWhU3gtX3FmUBS5cXqokuf1UUhP7XAmrXc2W+5dyK68SpDau/7o3AVRRxdTBIOl19ytKsYcfMRpBnJnhRGoM2YUHLBihvGdTHOGb/ANxpWzZNLr+ZQpGGICuJY2fL/wAuAIIG+YgAJvWpsfAcw2sDJuJtK62EANLiCAuq6iI2dwF0XCsMdh6YUG1Sw5gdUbwYBEVFa+WAYJoMrPDa9/6lOagE2PFTFqEGv7vRncZnBun0nHuKxThDnAlmkt/lEtVWWlXAARHKwZo2ky6r3KVgaHA+J0lV1FxaVHNRy0JgCrgO/wAjT6xMMfjTf3FBF2MLPcpFDsGZ9S3xZBZ+peUg8R/CcDWWV6TJnGYwsp8y0qr1p+ozjewsS3knKbmF4LGK2NVVYJUcV6ioETyLj1stV0/dzI+sUpR6DmWRPODK+2BsYzaaLgXQg6q1C6H6YPtzWlav1DgFq4lEQ3tNdPiUMBUziJKVVtVR50kXQwLbK+JQ+UBDDnLMtA2GVsvkk5r+/wDksMjx2xKS311DSsOBoPgmH5OcwaNC6DMXC/nGHIQM20gMrh0Vf2wWWB3bLDBAzmF3M/iDW/3LSA5PBBHICwwuIugObbqpYufvOx6hpEUq8xXMwJDArH13NHPzpDGhDCkWKtr+ZRalfEWiV3dBAEwz11KCjoMhn2ytqrW2LoHU8KsSMXg+4iYay/g95hsd20CKihrTAiAziPKSvPCFcUjgMQXmN09QqtRA7dQDf+p4YgbbZII9zJk/UAaK9Rp3FoCdJcooDu/7SiPg0QJOazUHTAiFI2YIopSw0seIKQqTFZqCJ5V3lA5GOCG1UObGbhe3QmMPPEGmwUGjunDMvNQLT48fErzYxXEb1k2GIuFicTdG+6lGEpcBsDwRyCpoNMMQA9FTQK+0uDRhX2wXVJdxKrITEC8cQ7H1F3RmaKI41zRjAw5yDLVPBGVqu3J4tX3ABTKCgB6iXsm5S5Q0NdwAUAOj8fECoB6RpcrAiXwGlQwHvFZ513PifErxK8Qxx/hq36gtbfcGXLuA5O09k8qNXMr5lfMT5lK2yt8y00CVBLt3KjP2mQnECAsbGFzk8UAUeJcYjgCrkGHMN2SlEC+ZSUuA8wEp5lfMBcB5lJXzAQEdQFvv0bhom/agqQ0lYnMqMolJSUjUV65r49sqgKlYCUlSo6hCXicRU//Z"))
     ]);
-
     const [selectedCard, setSelectedCard] = useState<Card | undefined>(undefined);
 
     const onGuessConfirm = () => {
         setIsShowEvent(false);
         setPlayCards([]);
         setSelectedCard(undefined);
-        // guess story api
+        // guesses story api
         setRoundState(undefined);
     }
 
@@ -45,8 +44,7 @@ const Guess = () => {
         return <>
             {
                 selectedCard ?
-                    <PlayCardDescription card={selectedCard}
-                                         description="猜"
+                    <GuessDescription card={selectedCard}
                                          onConfirmButtonClick={onGuessConfirm}
                                          onCancelButtonClick={onGuessCancel}/>
                     : <EventNotice isShowEvent={isShowEvent} dixitState={roundState}/>
