+++ src/app/dixit/actions/Rank.scss
@@ -1,0 +1,63 @@
+.dixit-score-board {
+  position: absolute;
+  width: 500px;
+  height: 500px;
+  left: 350px;
+  top: 100px;
+  border-radius: 30px;
+
+  font-weight: bold;
+  color: #56776C;
+
+  border-spacing: 0;
+  border-collapse: separate;
+
+  th {
+    height: 50px;
+    font-size: 22px;
+  }
+
+  td {
+    font-size: 26px;
+  }
+
+  tr:first-child th:first-child {
+    border-top-left-radius: 30px;
+  }
+
+  tr:first-child th:last-child {
+    border-top-right-radius: 30px;
+  }
+
+  tr:last-child td:first-child {
+    border-bottom-left-radius: 30px;
+  }
+
+  tr:last-child td:last-child {
+    border-bottom-right-radius: 30px;
+  }
+
+  .red {
+    background: #F6A9A9;
+  }
+
+  .orange {
+    background: #FFBF86;
+  }
+
+  .yellow {
+    background: #FFF47D;
+  }
+
+  .green {
+    background: #C2F784;
+  }
+
+  .blue {
+    background: #E4FBFF;
+  }
+
+  .violet {
+    background: #B8B5FF;
+  }
+}
