+++ src/app/dixit/cards/descriptions/CardDescription.scss
