+++ src/app/dixit/cards/guesses/Guesses.scss
@@ -1,0 +1,51 @@
+.dixit-guesses {
+  position: absolute;
+  width: auto;
+  height: 200px;
+  left: 65px;
+  top: 75px;
+  border-radius: 30px;
+
+  .dixit-guess {
+    width: 140px;
+    height: 210px;
+    margin: 0 15px;
+    border-radius: 30px;
+    border-width: 5px;
+    border-style: solid;
+  }
+
+  .dixit-guess-zone {
+    float: left;
+  }
+
+  .red {
+    @extend .dixit-guess;
+    border-color: #F6A9A9;
+  }
+
+  .orange {
+    @extend .dixit-guess;
+    border-color: #FFBF86;
+  }
+
+  .yellow {
+    @extend .dixit-guess;
+    border-color: #FFF47D;
+  }
+
+  .green {
+    @extend .dixit-guess;
+    border-color: #C2F784;
+  }
+
+  .blue {
+    @extend .dixit-guess;
+    border-color: #E4FBFF;
+  }
+
+  .violet {
+    @extend .dixit-guess;
+    border-color: #B8B5FF;
+  }
+}
