+++ src/app/dixit/cards/handcards/HandCards.scss
@@ -1,0 +1,17 @@
+.dixit-cards {
+  position: absolute;
+  width: 750px;
+  height: 200px;
+  left: 230px;
+  top: 505px;
+  background: #F0E3CA;
+  border-radius: 30px;
+
+  .dixit-card {
+    width: 100px;
+    height: 150px;
+    margin: 25px 6.5px 0;
+    border-radius: 30px;
+    border: 4px solid #56776C;
+  }
+}
