+++ src/app/dixit/logo/Logo.tsx
@@ -1,14 +1,12 @@
 import './Logo.scss';
 
 const Logo = () => {
-    return (
-        <div className="dixit-logo">
-            <span className="indigo">D</span>
-            <span className="violet">I</span>
-            <span className="red">X</span>
-            <span className="yellow">I</span>
-            <span className="indigo">T</span>
-        </div>
-    );
+    return <div className="dixit-logo">
+        <span className="indigo">D</span>
+        <span className="violet">I</span>
+        <span className="red">X</span>
+        <span className="yellow">I</span>
+        <span className="indigo">T</span>
+    </div>;
 }
 export default Logo;
