+++ src/app/dixit/model/domain/Card.ts
