+++ src/app/dixit/model/domain/GameState.ts
