+++ src/app/dixit/model/domain/Guess.ts
