+++ src/app/dixit/model/domain/Story.ts
