+++ src/app/dixit/model/events/DixitRoundStoryTellingEvent.ts
@@ -1,13 +1,13 @@
+import Card from "../domain/Card";
 import Event from "./Event";
-import Player from "../Player";
-import {RoundState} from "../RoundState";
+import {RoundState} from "../domain/RoundState";
 
-export default class DixitRoundOverEvent extends Event {
+export default class DixitRoundStoryTellingEvent extends Event {
     constructor(public readonly gameId: string,
                 public readonly rounds: number,
                 public readonly playerId: string,
                 public readonly roundState: RoundState,
-                public readonly players: Array<Player>) {
+                public readonly handCards: Array<Card>) {
         super(gameId, rounds, playerId);
     }
 }
