+++ src/app/dixit/model/events/Event.ts
