+++ src/app/dixit/playerbar/PlayerBar.scss
@@ -12,4 +12,74 @@
     justify-content: center; /* align horizontal */
     align-items: center; /* align vertical */
   }
+
+  .player-item {
+    width: 180px;
+    height: 50px;
+    background-color: #F0E3CA;
+    border-radius: 30px;
+    margin: 20px auto;
+    display: flex;
+    align-items: center; /* align vertical */
+
+    .font {
+      font-family: Roboto, sans-serif;
+      font-style: normal;
+      font-weight: bold;
+      font-size: 20px;
+      text-align: center;
+    }
+
+    .color {
+      @extend .font;
+      width: 30px;
+      height: 30px;
+      border-radius: 45px;
+      margin-left: 12px;
+      line-height: 30px;
+      color: #596C68;
+    }
+
+    .red {
+      @extend .color;
+      background: #F6A9A9;
+    }
+
+    .orange {
+      @extend .color;
+      background: #FFBF86;
+    }
+
+    .yellow {
+      @extend .color;
+      background: #FFF47D;
+    }
+
+    .green {
+      @extend .color;
+      background: #C2F784;
+    }
+
+    .blue {
+      @extend .color;
+      background: #E4FBFF;
+    }
+
+    .violet {
+      @extend .color;
+      background: #B8B5FF;
+    }
+
+    .player-name {
+      @extend .font;
+      width: 120px;
+      height: 30px;
+      margin-left: 5px;
+      background: #403F48;
+      border-radius: 20px 30px 30px 20px;
+      color: #95A792;
+      letter-spacing: 0.07em;
+    }
+  }
+
 }
